% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AccumulateAttribute}
\alias{AccumulateAttribute}
\title{accumulate}
\description{
Governs whether or not the animation is cumulative.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'sum'}}{Specifies to accumulate values, that is each repeat iteration builds upon the last value of the previous iteration.}
\item{\emph{'none'}}{Specifies not to accumulate values. (The default)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
