% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AElementTargetAttribute}
\alias{AElementTargetAttribute}
\title{target}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{_replace}{The current SVG image is replaced by the linked content in the same rectangular area in the same frame as the current SVG image.}
\item{_self}{The current SVG image is replaced by the linked content in the same frame as the current SVG image. If the attribute is not specified, '_self' is assumed.}
\item{_parent}{The immediate frameset parent of the SVG image is replaced by the linked content.}
\item{_top}{The content of the full window or tab, including any frames, is replaced by the linked content}
\item{_blank}{A new un.named window or tab is requested for the display of the linked content. If this fails, the result is the same as _top}
\item{<XML.Name>}{Specifies the name of the frame, pane, or other relevant presentation context for display of the linked content. If this already exists, it is re.used, replacing the existing content. If it does not exist, it is created (the same as '_blank', except that it now has a name).}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}}
}
}
\keyword{internal}

