% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AElementXLinkActuateAttribute}
\alias{AElementXLinkActuateAttribute}
\title{xlink.actuate}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{onRequest}{This attribute provides documentation to XLink.aware processors that an application should traverse from the starting resource to the ending resource only on a post.loading event triggered for the purpose of traversal. Refer to the XML Linking Language (XLink) [XLINK]. Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}}
}
}
\keyword{internal}

