% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AElementXLinkHrefAttribute}
\alias{AElementXLinkHrefAttribute}
\title{xlink.href}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<IRI>}{The location of the referenced object, expressed as an IRI reference. Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}}
}
}
\keyword{internal}

