% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{altGlyph}
\alias{altGlyph}
\title{Element used to provide control over the glyphs used to render particular character data}
\description{
The altGlyph element allows sophisticated selection of the glyphs used to render its child character data.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{animation event attributes}}{\code{\link[=OnLoadEventAttribute]{onload}}}
\item{\emph{combining attributes}}{\code{\link[=AltGlyphElementDXYAttribute]{dxy}}, \code{\link[=AltGlyphElementXYAttribute]{xy}}}
\item{\emph{conditional processing attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{filter primitive attributes}}{\code{\link[=AltGlyphElementXAttribute]{x}}, \code{\link[=AltGlyphElementYAttribute]{y}}}
\item{\emph{graphical event attributes}}{\code{\link[=OnActivateEventAttribute]{onactivate}}, \code{\link[=OnClickEventAttribute]{onclick}}, \code{\link[=OnFocusInEventAttribute]{onfocusin}}, \code{\link[=OnFocusOutEventAttribute]{onfocusout}}, \code{\link[=OnLoadEventAttribute]{onload}}, \code{\link[=OnMouseDownEventAttribute]{onmousedown}}, \code{\link[=OnMouseMoveEventAttribute]{onmousemove}}, \code{\link[=OnMouseOutEventAttribute]{onmouseout}}, \code{\link[=OnMouseOverEventAttribute]{onmouseover}}, \code{\link[=OnMouseUpEventAttribute]{onmouseup}}}
\item{\emph{presentation attributes}}{\code{\link[=alignment-baseline-presentationAttribute]{alignment.baseline}}, \code{\link[=baseline-shift-presentationAttribute]{baseline.shift}}, \code{\link[=direction-presentationAttribute]{direction}}, \code{\link[=display-presentationAttribute]{display}}, \code{\link[=dominant-baseline-presentationAttribute]{dominant.baseline}}, \code{\link[=fill-presentationAttribute]{fill}}, \code{\link[=fill-opacity-presentationAttribute]{fill.opacity}}, \code{\link[=fill-rule-presentationAttribute]{fill.rule}}, \code{\link[=font-presentationAttribute]{font}}, \code{\link[=font-family-presentationAttribute]{font.family}}, \code{\link[=font-size-presentationAttribute]{font.size}}, \code{\link[=font-size-adjust-presentationAttribute]{font.size.adjust}}, \code{\link[=font-stretch-presentationAttribute]{font.stretch}}, \code{\link[=font-style-presentationAttribute]{font.style}}, \code{\link[=font-variant-presentationAttribute]{font.variant}}, \code{\link[=font-weight-presentationAttribute]{font.weight}}, \code{\link[=glyph-orientation-horizontal-presentationAttribute]{glyph.orientation.horizontal}}, \code{\link[=glyph-orientation-vertical-presentationAttribute]{glyph.orientation.vertical}}, \code{\link[=kerning-presentationAttribute]{kerning}}, \code{\link[=letter-spacing-presentationAttribute]{letter.spacing}}, \code{\link[=stroke-presentationAttribute]{stroke}}, \code{\link[=stroke-dasharray-presentationAttribute]{stroke.dasharray}}, \code{\link[=stroke-dashoffset-presentationAttribute]{stroke.dashoffset}}, \code{\link[=stroke-linecap-presentationAttribute]{stroke.linecap}}, \code{\link[=stroke-linejoin-presentationAttribute]{stroke.linejoin}}, \code{\link[=stroke-miterlimit-presentationAttribute]{stroke.miterlimit}}, \code{\link[=stroke-opacity-presentationAttribute]{stroke.opacity}}, \code{\link[=stroke-width-presentationAttribute]{stroke.width}}, \code{\link[=text-anchor-presentationAttribute]{text.anchor}}, \code{\link[=text-decoration-presentationAttribute]{text.decoration}}, \code{\link[=unicode-bidi-presentationAttribute]{unicode.bidi}}, \code{\link[=visibility-presentationAttribute]{visibility}}, \code{\link[=word-spacing-presentationAttribute]{word.spacing}}}
\item{\emph{unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=AltGlyphElementDXAttribute]{dx}}, \code{\link[=AltGlyphElementDYAttribute]{dy}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=AltGlyphElementFormatAttribute]{format}}, \code{\link[=AltGlyphElementGlyphRefAttribute]{glyphRef}}, \code{\link[=AltGlyphElementRotateAttribute]{rotate}}, \code{\link[=StyleAttribute]{style}}}
\item{\emph{xlink attributes}}{\code{\link[=XLinkActuateAttribute]{xlink.actuate}}, \code{\link[=XLinkArcRoleAttribute]{xlink.arcrole}}, \code{\link[=AltGlyphElementHrefAttribute]{xlink.href}}, \code{\link[=XLinkRoleAttribute]{xlink.role}}, \code{\link[=XLinkShowAttribute]{xlink.show}}, \code{\link[=XLinkTitleAttribute]{xlink.title}}, \code{\link[=XLinkTypeAttribute]{xlink.type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{Any elements or character data.}}
}
}
\keyword{element}

