% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{altGlyphDef}
\alias{altGlyphDef}
\title{Element used to define a set of possible glyph substitutions}
\description{
The altGlyphDef element defines a substitution representation for glyphs. This will consist of  either one or more glyphRef elements or one or more altGlyphitem elements.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=altGlyphItem]{altGlyphItem}}, \code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{element}

