% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementDXAttribute}
\alias{AltGlyphElementDXAttribute}
\title{dx}
\description{
Set shifts in the current x-position prior to rendering.
}
\section{Available Attribute Values}{

The value  represents one or more lengths, given as a vector.
\describe{
\item{<length>[N]}{Specifies the x-shifts values for the \strong{altGlyph} element. If rendering only unicode characters inside an \strong{altGlyph} element, then this specifies the absolute x-shift for the first n characters where n is the length of the dx-vector . If rendering alternate \emph{glyphs}, then dx[1] specifies a x-shift of the current text position, while the remainder, dx[-1], is ignored.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
