% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementDYAttribute}
\alias{AltGlyphElementDYAttribute}
\title{dy}
\description{
Set shifts in the current y-position prior to rendering.
}
\section{Available Attribute Values}{

The value  represents one or more lengths, given as a vector.
\describe{
\item{<length>[N]}{Specifies a sequence of  the y-shifts values for the \emph{altGlyph} element. If rendering only unicode characters inside an \strong{altGlyph} element, then this specifies the absolute y-shift for the first n characters where n is the length of the dy-vector . If rendering alternate \emph{glyphs}, then dy[1] specifies a y-shift of the current text position, while the remainder, dy[-1], is ignored.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
