% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementFormatAttribute}
\alias{AltGlyphElementFormatAttribute}
\title{format}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<string>}{The format of the given font. If the font is in one of the formats listed in CSS2 ([CSS2], section 15.3.5), such as TrueDoc Portable Font Resource or Embedded OpenType, then the <string> must contain the corresponding font format string, such as truedoc.pfr or embedded.opentype. (This attribute has the same meaning as the 'format' attribute on the 'glyphRef' element.)Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}
\keyword{internal}

