% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementGlyphRefAttribute}
\alias{AltGlyphElementGlyphRefAttribute}
\title{glyphRef}
\description{
Sets the alt-glyph reference identifier.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<string>}{Specifies a \emph{glyph} identifier. The format of the font will determine the \emph{glyph} identifier format.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
