% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementRotateAttribute}
\alias{AltGlyphElementRotateAttribute}
\title{rotate}
\description{
Sets a sequence of  one or more rotations to be applied to consecutive glyphs upon rendering.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<angle>[N]}{Specifies an additional rotation (in degrees) to be applied to characters within this \strong{altGlyph}. If rendering only unicode characters inside an \strong{altGlyph} element, then this specifies an aditional rotation (in degrees) to be applied sequentially for the first n characters where n is the length of the rotate-vector and applies the final rotate to the remaining characters (if any). If rendering alternate \emph{glyphs}, then dy[1] specifies a rotation to be applied to the entire sequence, while the remainder, dy[-1], is ignored.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
