% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementYAttribute}
\alias{AltGlyphElementYAttribute}
\title{y}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list.of.coordinates>}{The corresponding absolute Y coordinates for rendering the 'altGlyph' element.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}
\keyword{internal}

