% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{altGlyphItem}
\alias{altGlyphItem}
\title{Element used to provide a set of glyph substitution candidates}
\description{
The altGlyphItem element provides a set of candidates for glyph substitution by the <altglyph> element.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{element}

