% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{animateMotion}
\alias{animateMotion}
\title{An element used to move a shape along a path}
\description{
The animateMotion element causes a referenced element to move along a motion path.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Animation Addition Attributes}}{\code{\link[=AccumulateAttribute]{accumulate}}, \code{\link[=AdditiveAttribute]{additive}}}
\item{\emph{Animation Event Attributes}}{\code{\link[=OnBeginEventAttribute]{onbegin}}, \code{\link[=OnEndEventAttribute]{onend}}, \code{\link[=OnLoadEventAttribute]{onload}}, \code{\link[=OnRepeatEventAttribute]{onrepeat}}}
\item{\emph{Animation Timing Attributes}}{\code{\link[=BeginAttribute]{begin}}, \code{\link[=DurAttribute]{dur}}, \code{\link[=EndAttribute]{end}}, \code{\link[=FillAttribute]{fill}}, \code{\link[=MaxAttribute]{max}}, \code{\link[=MinAttribute]{min}}, \code{\link[=RepeatCountAttribute]{repeatCount}}, \code{\link[=RepeatDurAttribute]{repeatDur}}, \code{\link[=RestartAttribute]{restart}}}
\item{\emph{Animation Value Attributes}}{\code{\link[=ByAttribute]{by}}, \code{\link[=CalcModeAttribute]{calcMode}}, \code{\link[=FromAttribute]{from}}, \code{\link[=KeySplinesAttribute]{keySplines}}, \code{\link[=KeyTimesAttribute]{keyTimes}}, \code{\link[=ToAttribute]{to}}, \code{\link[=ValuesAttribute]{values}}}
\item{\emph{Conditional Processing Attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Graphical Event Attributes}}{\code{\link[=OnLoadEventAttribute]{onload}}}
\item{\emph{Shape Elements}}{\code{\link[=PathAttribute]{path}}}
\item{\emph{Unclassified}}{\code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=KeyPointsAttribute]{keyPoints}}, \code{\link[=OriginAttribute]{origin}}, \code{\link[=RotateAttribute]{rotate}}}
\item{\emph{Xlink Attributes}}{\code{\link[=XLinkActuateAttribute]{xlink.actuate}}, \code{\link[=XLinkArcRoleAttribute]{xlink.arcrole}}, \code{\link[=HrefAttribute]{xlink.href}}, \code{\link[=XLinkRoleAttribute]{xlink.role}}, \code{\link[=XLinkShowAttribute]{xlink.show}}, \code{\link[=XLinkTitleAttribute]{xlink.title}}, \code{\link[=XLinkTypeAttribute]{xlink.type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=mpath]{mpath}}}
}
}

\keyword{element}
