% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AnimateTransformElementTypeAttribute}
\alias{AnimateTransformElementTypeAttribute}
\title{type}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{translate | scale | rotate | skewX | skewY}{Indicates the type of transformation which is to have its values change over time. If the attribute is not specified, then the effect is as if a value of 'translate' were specified.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateTransform]{animateTransform}}}
}
}
\keyword{internal}

