% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxColors.R
\name{as.rgb}
\alias{as.rgb}
\title{convert to rgb representation}
\usage{
as.rgb(r, g, b)
}
\arguments{
\item{r}{red component: 0-255}

\item{g}{green component: 0-255}

\item{b}{blue component: 0-255}
}
\value{
returns character vector representing a rgb color
}
\description{
convert to rgb representation
}

