% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AttributeNameAttribute}
\alias{AttributeNameAttribute}
\title{attributeName}
\description{
Sets the name of the target attribute.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<attribute-name>}{Specifies the name of the target attribute.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
