% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgRIdGenerators.R
\name{autoId}
\alias{autoId}
\title{Auto id generator}
\usage{
autoId(inc = TRUE)
}
\arguments{
\item{inc}{TRUE means increment the counter (default)}
}
\description{
Auto id generator
}
\note{
inc=FALSE is provided to retreive the last inc=TRUE value
}

