% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{baseline-shift-presentationAttribute}
\alias{baseline-shift-presentationAttribute}
\title{baseline.shift}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{baseline}{**ToDo!!!** }
\item{sub}{**ToDo!!!** }
\item{super}{**ToDo!!!** }
\item{<percentage>}{**ToDo!!!** }
\item{<length>}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}
\keyword{internal}

