% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{BeginAttribute}
\alias{BeginAttribute}
\title{begin}
\description{
Governs when the animation should begin.
}
\section{Available Attribute Values}{

The value is a vector composed of one or more of the following:
\describe{
\item{<offset-value>}{Specifies when to begin the animation, based upon an offset (+ or - ) from the implicit synbase (default is seconds)}
\item{<syncbase-value>}{Specifies when to begin the animation, based upon an offset (+ or - ) from the explicit synbase given by the begin or end of the element referenced by Id.value}
\item{<event-value>}{Specifies when to begin the animation, based upon an offset (+ or - ) relative to when the event referenced occurs.}
\item{<repeat-value>}{Specifies when to begin the animation, based upon an offset (+ or - ) relative to a repeating event.}
\item{<accessKey-value>}{Specifies when to begin the animation, based upon an offset (+ or - ) relative to an access key event of the specified character.}
\item{<wallclock-sync-value>}{Specifies when to begin the animation, based on wall clock time.}
\item{\emph{'indefinite'}}{Specifies when to begin the animation, using a \emph{beginElement()} method call or a hyperlink targeted to the element to trigger the begin.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
