% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ByAttribute}
\alias{ByAttribute}
\title{by}
\description{
Sets a relative offset value for the animation.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<value>}{Specifies the relative offset value to be used to compute the final ending value, That is  \eqn{endValue = startingValue + by}. This overrides any \emph{from to} specification.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
