% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{circle}
\alias{circle}
\title{An element used to create a circular shape}
\description{
The circle element is an SVG basic shape, used to create circles based on a center point and a radius.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Animation Event Attributes}}{\code{\link[=OnLoadEventAttribute]{onload}}}
\item{\emph{Complementary Attributes}}{\code{\link[=CircleElementCXYAttribute]{cxy}}}
\item{\emph{Conditional Processing Attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Graphical Event Attributes}}{\code{\link[=OnActivateEventAttribute]{onactivate}}, \code{\link[=OnClickEventAttribute]{onclick}}, \code{\link[=OnFocusInEventAttribute]{onfocusin}}, \code{\link[=OnFocusOutEventAttribute]{onfocusout}}, \code{\link[=OnLoadEventAttribute]{onload}}, \code{\link[=OnMouseDownEventAttribute]{onmousedown}}, \code{\link[=OnMouseMoveEventAttribute]{onmousemove}}, \code{\link[=OnMouseOutEventAttribute]{onmouseout}}, \code{\link[=OnMouseOverEventAttribute]{onmouseover}}, \code{\link[=OnMouseUpEventAttribute]{onmouseup}}}
\item{\emph{Presentation Attributes}}{\code{\link[=clip-path-presentationAttribute]{clip.path}}, \code{\link[=clip-rule-presentationAttribute]{clip.rule}}, \code{\link[=color-presentationAttribute]{color}}, \code{\link[=color-interpolation-presentationAttribute]{color.interpolation}}, \code{\link[=color-rendering-presentationAttribute]{color.rendering}}, \code{\link[=cursor-presentationAttribute]{cursor}}, \code{\link[=display-presentationAttribute]{display}}, \code{\link[=fill-presentationAttribute]{fill}}, \code{\link[=fill-opacity-presentationAttribute]{fill.opacity}}, \code{\link[=fill-rule-presentationAttribute]{fill.rule}}, \code{\link[=filter-presentationAttribute]{filter}}, \code{\link[=mask-presentationAttribute]{mask}}, \code{\link[=opacity-presentationAttribute]{opacity}}, \code{\link[=pointer-events-presentationAttribute]{pointer.events}}, \code{\link[=shape-rendering-presentationAttribute]{shape.rendering}}, \code{\link[=stroke-presentationAttribute]{stroke}}, \code{\link[=stroke-dasharray-presentationAttribute]{stroke.dasharray}}, \code{\link[=stroke-dashoffset-presentationAttribute]{stroke.dashoffset}}, \code{\link[=stroke-linecap-presentationAttribute]{stroke.linecap}}, \code{\link[=stroke-linejoin-presentationAttribute]{stroke.linejoin}}, \code{\link[=stroke-miterlimit-presentationAttribute]{stroke.miterlimit}}, \code{\link[=stroke-opacity-presentationAttribute]{stroke.opacity}}, \code{\link[=stroke-width-presentationAttribute]{stroke.width}}, \code{\link[=visibility-presentationAttribute]{visibility}}}
\item{\emph{Unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=CircleElementCXAttribute]{cx}}, \code{\link[=CircleElementCYAttribute]{cy}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=CircleElementRAttribute]{r}}, \code{\link[=StyleAttribute]{style}}, \code{\link[=TransformAttribute]{transform}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
}
}

\keyword{element}
