% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{CircleElementCXYAttribute}
\alias{CircleElementCXYAttribute}
\title{cxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=CircleElementCXAttribute]{cx}}, \code{\link[=CircleElementCYAttribute]{cy}}
}

\section{Equivalence}{

\describe{
\item{}{cxy=c(value.CX,value.CY)}
\item{and}{cx = value.CX; cy = value.CY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}}
}
}
\keyword{internal}

