% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{clip-path-presentationAttribute}
\alias{clip-path-presentationAttribute}
\title{clip.path}
\description{
Specifies path for clipping.
}
\section{Available Attribute Values}{



\describe{
\item{<clipPath-Element>}{Specifies a \strong{clipPath} element to be used for clipping}
\item{<funciri>}{Specifies a reference to the \strong{clipPath} to be used for clipping by using the functional IRI notation.}
\item{\emph{'\emph{'inherit'}'}}{Specifies that the clip-path to be applied to  this element should  be inherited from this elements parents.}
\item{\emph{'\emph{'none'}'}}{Specifies no clipping to occur.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}, \code{\link[=use]{use}}}
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}, \code{\link[=glyph]{glyph}}, \code{\link[=marker]{marker}}, \code{\link[=mask]{mask}}, \code{\link[=missing-glyph]{missing.glyph}}, \code{\link[=pattern]{pattern}}, \code{\link[=switch]{switch}}}
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Container Elements}}{\code{\link[=defs]{defs}}, \code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=clipPath]{clipPath}}}
}
}

\keyword{internal}
