% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ClipPathElementClipPathUnitsAttribute}
\alias{ClipPathElementClipPathUnitsAttribute}
\title{clipPathUnits}
\description{
Governs how to interpret the coordinates for the contents of the clipPath. That should they be interpeted  as objectBoundingBox or userSpaceOnUse?
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that the coordinate system for the contents of the \emph{'clipPath'} is given in terms of the specific user coordinate system that is in effect at the time of clipPath referrencing. (Default)}
\item{\emph{'objectBoundingBox'}}{Specifies that the coordinate system for the contents of the \emph{'clipPath'} is given in terms of the bounding box of target object(s) to be clipped.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=clipPath]{clipPath}}}
}
}

\keyword{internal}
