% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{color-interpolation-filters-presentationAttribute}
\alias{color-interpolation-filters-presentationAttribute}
\title{color.interpolation.filters}
\description{
Governs the color space for imaging operations performed via filter effects.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'auto'}'}'}'}}{Specifies to auto select the color-interpolation filter}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the color-interpolation methodology from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'linearRGB'}'}'}'}}{Specifies to use the linearized RGB color space for the filter effects color operations .}
\item{\emph{'\emph{'\emph{'\emph{'sRGB'}'}'}'}}{Specifies to use the sRGB color space for the filter effects color operations .}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feBlend]{feBlend}}, \code{\link[=feColorMatrix]{feColorMatrix}}, \code{\link[=feComponentTransfer]{feComponentTransfer}}, \code{\link[=feComposite]{feComposite}}, \code{\link[=feConvolveMatrix]{feConvolveMatrix}}, \code{\link[=feDiffuseLighting]{feDiffuseLighting}}, \code{\link[=feDisplacementMap]{feDisplacementMap}}, \code{\link[=feFlood]{feFlood}}, \code{\link[=feGaussianBlur]{feGaussianBlur}}, \code{\link[=feImage]{feImage}}, \code{\link[=feMerge]{feMerge}}, \code{\link[=feMorphology]{feMorphology}}, \code{\link[=feOffset]{feOffset}}, \code{\link[=feSpecularLighting]{feSpecularLighting}}, \code{\link[=feTile]{feTile}}, \code{\link[=feTurbulence]{feTurbulence}}}
}
}

\keyword{internal}
