% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{color-interpolation-presentationAttribute}
\alias{color-interpolation-presentationAttribute}
\title{color.interpolation}
\description{
Governs the optimization strategy for color-interpolation.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'auto'}'}}{Specifies an auto selection of the color interpolation.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the method of color interpolation from the parents.}
\item{\emph{'\emph{'linearRGB'}'}}{Specifies to use a  linearized RGB color space for color interpolation}
\item{\emph{'\emph{'sRGB'}'}}{Specifies to use sRGB color space for color interpolation}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}}
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}, \code{\link[=use]{use}}}
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}, \code{\link[=glyph]{glyph}}, \code{\link[=marker]{marker}}, \code{\link[=mask]{mask}}, \code{\link[=missing-glyph]{missing.glyph}}, \code{\link[=pattern]{pattern}}, \code{\link[=switch]{switch}}}
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Container Elements}}{\code{\link[=defs]{defs}}, \code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
