% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{color-profile-presentationAttribute}
\alias{color-profile-presentationAttribute}
\title{color.profile}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{auto}{**ToDo!!!** }
\item{sRGB}{**ToDo!!!** }
\item{<name>}{**ToDo!!!** }
\item{<iri>}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=image]{image}}}
}
}
\keyword{internal}

