% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{color-profile}
\alias{color-profile}
\alias{color.profile}
\title{Element used to specify color profiles}
\description{
The element allows describing the color profile used for the image.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{unclassified}}{\code{\link[=ColorProfileElementLocalAttribute]{local}}, \code{\link[=ColorProfileElementNameAttribute]{name}}, \code{\link[=ColorProfileElementRenderingIntentAttribute]{rendering.intent}}}
\item{\emph{xlink attributes}}{\code{\link[=XLinkActuateAttribute]{xlink.actuate}}, \code{\link[=XLinkArcRoleAttribute]{xlink.arcrole}}, \code{\link[=ColorProfileElementHrefAttribute]{xlink.href}}, \code{\link[=XLinkRoleAttribute]{xlink.role}}, \code{\link[=XLinkShowAttribute]{xlink.show}}, \code{\link[=XLinkTitleAttribute]{xlink.title}}, \code{\link[=XLinkTypeAttribute]{xlink.type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
}
}
\keyword{element}

