% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{color-rendering-presentationAttribute}
\alias{color-rendering-presentationAttribute}
\title{color.rendering}
\description{
Governs the color space for gradient interpolations, color animations and alpha compositing.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'auto'}'}}{Specifies to employ a trade-off strategy of optimization of color-rending  which balances speed vs quality.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the color-rendering optimzation from the parent.}
\item{\emph{'\emph{'optimizeQuality'}'}}{Specifies to optimize the color-rendiering  for quality.}
\item{\emph{'\emph{'optimizeSpeed'}'}}{Specifies to optimize the color-rendiering  for speed}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}}
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}, \code{\link[=use]{use}}}
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}, \code{\link[=glyph]{glyph}}, \code{\link[=marker]{marker}}, \code{\link[=mask]{mask}}, \code{\link[=missing-glyph]{missing.glyph}}, \code{\link[=pattern]{pattern}}, \code{\link[=switch]{switch}}}
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Container Elements}}{\code{\link[=defs]{defs}}, \code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
