% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ColorProfileElementHrefAttribute}
\alias{ColorProfileElementHrefAttribute}
\title{xlink.href}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<iri>}{The location of an ICC profile resource.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=color-profile]{color.profile}}}
}
}
\keyword{internal}

