% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ColorProfileElementLocalAttribute}
\alias{ColorProfileElementLocalAttribute}
\title{local}
\description{
Assigns a unique ID for a locally stored color profile, for referencing elsewhere.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<string>}{Specifes a color profile by ID.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=color-profile]{color.profile}}}
}
}

\keyword{internal}
