% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ColorProfileElementNameAttribute}
\alias{ColorProfileElementNameAttribute}
\title{name}
\description{
Sets the color profile name.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<name>}{Specifies the name of the first parameter in \emph{icc-color} specification. Note: \emph{""""""""""""""""sRGB""""""""""""""""} is predefined; and will result in the specification this being ignored.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=color-profile]{color.profile}}}
}
}

\keyword{internal}
