% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ContentScriptTypeAttribute}
\alias{ContentScriptTypeAttribute}
\title{contentScriptType}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{content.type}{Identifies the default scripting language for the given SVG document fragment. This attribute sets the default scripting language used to process the value strings in event attributes. This language must be used for all instances of script that do not specify their own scripting language. The value content.type specifies a media type, per MIME Part Two. Media Types [RFC2046]. The default value is 'application/ecmascript' [RFC4329].Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}}
}
}
\keyword{internal}

