% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ContentStyleTypeAttribute}
\alias{ContentStyleTypeAttribute}
\title{contentStyleType}
\description{
Sets the default style sheet language for the given document.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<content-type>}{Specifies a default style sheet language for the given document.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
