% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{CursorElementHrefAttribute}
\alias{CursorElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to a cursor image element or cursor image file.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<FUNCIRI>}{Specifies a reference to the cursor image. The reference is in the form of a A Functional \emph{IRI} pointing to a file or element.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=cursor]{cursor}}}
}
}

\keyword{internal}
