% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{direction-presentationAttribute}
\alias{direction-presentationAttribute}
\title{direction}
\description{
Controls the base writing direction of text
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit tne direction from an ancestor.}
\item{\emph{'\emph{'ltr'}'}}{Specifies a left to right direction.}
\item{\emph{'\emph{'rtl'}'}}{Specifies a right to left direction.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
