% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{display-presentationAttribute}
\alias{display-presentationAttribute}
\title{display}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{inline}{**ToDo!!!** }
\item{block}{**ToDo!!!** }
\item{list-item}{**ToDo!!!** }
\item{run-in}{**ToDo!!!** }
\item{compact}{**ToDo!!!** }
\item{marker}{**ToDo!!!** }
\item{table}{**ToDo!!!** }
\item{inline-table}{**ToDo!!!** }
\item{table-row-group}{**ToDo!!!** }
\item{table-header-group}{**ToDo!!!** }
\item{table-footer-group}{**ToDo!!!** }
\item{table-row}{**ToDo!!!** }
\item{table-column-group}{**ToDo!!!** }
\item{table-column}{**ToDo!!!** }
\item{table-cell}{**ToDo!!!** }
\item{table-caption}{**ToDo!!!** }
\item{none}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Elements}}{\code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=use]{use}}}
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}, \code{\link[=foreignObject]{foreignObject}}, \code{\link[=image]{image}}, \code{\link[=switch]{switch}}, \code{\link[=text]{text}}}
}
}
\keyword{internal}

