% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{dominant-baseline-presentationAttribute}
\alias{dominant-baseline-presentationAttribute}
\title{dominant.baseline}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{auto}{**ToDo!!!** }
\item{use-script}{**ToDo!!!** }
\item{no-change}{**ToDo!!!** }
\item{reset-size}{**ToDo!!!** }
\item{ideographic}{**ToDo!!!** }
\item{alphabetic}{**ToDo!!!** }
\item{hanging}{**ToDo!!!** }
\item{mathematical}{**ToDo!!!** }
\item{central}{**ToDo!!!** }
\item{middle}{**ToDo!!!** }
\item{text-after-edge}{**ToDo!!!** }
\item{text-before-edge}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

