% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{EllipseElementCYAttribute}
\alias{EllipseElementCYAttribute}
\title{cy}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{The y.axis coordinate of the center of the ellipse. If the attribute is not specified, the effect is as if a value of "0" were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=ellipse]{ellipse}}}
}
}
\keyword{internal}

