% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{EllipseElementRXAttribute}
\alias{EllipseElementRXAttribute}
\title{rx}
\description{
Sets the x-radius of the ellipse.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the x-axis radius, \eqn{r_x} of the \strong{ellipse}. Note \eqn{r_x \geq 0} is required.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=ellipse]{ellipse}}}
}
}

\keyword{internal}
