% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{EllipseElementRYAttribute}
\alias{EllipseElementRYAttribute}
\title{ry}
\description{
Sets the y-radius of the ellipse.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the y-axis radius, \eqn{r_y} of the \strong{ellipse}. Note \eqn{r_y \geq 0} is required.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=ellipse]{ellipse}}}
}
}

\keyword{internal}
