% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{EndAttribute}
\alias{EndAttribute}
\title{end}
\description{
Governs the ending constraint for active duration.
}
\section{Available Attribute Values}{

The value is a vector composed of one or more of the following:
\describe{
\item{<offset-value>}{Specifies when to end the animation, based upon an offset (+ or - ) from the implicit \emph{synbase} (without units the default is seconds)}
\item{<syncbase-value>}{Specifies when to end the animation, based upon an offset (+ or - ) from the explicit \emph{synbase} given by the begin or end of the element referenced by id-value.}
\item{<event-value>}{Specifies when to end the animation, based upon an offset (+ or - ) relative to when the event referenced occurs.}
\item{<repeat-value>}{Specifies when to end the animation, based upon an offset (+ or - ) relative to a repeating event.}
\item{<accessKey-value>}{Specifies when to end the animation, based upon an offset (+ or - ) relative to an access key event of the specified character.}
\item{<wallclock-sync-value>}{Specifies when to end the animation, base on wall clock time.}
\item{\emph{'indefinite'}}{Specifies when to end the animation, using a \emph{endElement()} method call or a hyperlink targeted to the element to trigger the end.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
