% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feBlendModeAttribute}
\alias{feBlendModeAttribute}
\title{mode}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{normal | multiply | screen | darken | lighten}{One of the image blending modes (see table below). If attribute 'mode' is not specified, then the effect is as if a value of normal were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feBlend]{feBlend}}}
}
}
\keyword{internal}

