% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feColorMatrixValuesAttribute}
\alias{feColorMatrixValuesAttribute}
\title{values}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{list of <number>s}{The contents of 'values' depends on the value of attribute 'type'. For type="matrix", 'values' is a list of 20 matrix values (a00 a01 a02 a03 a04 a10 a11 ... a34), separated by whitespace and/or a comma. For example, the identity matrix could be expressed as. type="matrix" values="1 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 1 0" For type="saturate", 'values' is a single real number value (0 to 1). A saturate operation is equivalent to the following matrix operation. | R' | |0.213+0.787s 0.715.0.715s 0.072.0.072s 0 0 | | R || G' | |0.213.0.213s 0.715+0.285s 0.072.0.072s 0 0 | | G || B' | = |0.213.0.213s 0.715.0.715s 0.072+0.928s 0 0 | * | B || A' | | 0 0 0 1 0 | | A || 1 | | 0 0 0 0 1 | | 1 | For type="hueRotate", 'values' is a single one real number value (degrees). A hueRotate operation is equivalent to the following matrix operation. | R' | | a00 a01 a02 0 0 | | R || G' | | a10 a11 a12 0 0 | | G || B' | = | a20 a21 a22 0 0 | * | B || A' | | 0 0 0 1 0 | | A || 1 | | 0 0 0 0 1 | | 1 | where the terms a00, a01, etc. are calculated as follows. | a00 a01 a02 | [+0.213 +0.715 +0.072]| a10 a11 a12 | = [+0.213 +0.715 +0.072] +| a20 a21 a22 | [+0.213 +0.715 +0.072] [+0.787 .0.715 .0.072]cos(hueRotate value) * [.0.213 +0.285 .0.072] + [.0.213 .0.715 +0.928] [.0.213 .0.715+0.928]sin(hueRotate value) * [+0.143 +0.140.0.283] [.0.787 +0.715+0.072] Thus, the upper left term of the hue matrix turns out to be. .213 + cos(hueRotate value)*.787 . sin(hueRotate value)*.213 For type="luminanceToAlpha", 'values' is not applicable. A luminanceToAlpha operation is equivalent to the following matrix operation. | R' | | 0 0 0 0 0 | | R | | G' | | 0 0 0 0 0 | | G | | B' | = | 0 0 0 0 0 | * | B | | A' | | 0.2125 0.7154 0.0721 0 0 | | A | | 1 | | 0 0 0 0 1 | | 1 | If the attribute is not specified, then the default behavior depends on the value of attribute 'type'. If type="matrix", then this attribute defaults to the identity matrix. If type="saturate", then this attribute defaults to the value 1, which results in the identity matrix. If type="hueRotate", then this attribute defaults to the value 0, which results in the identity matrix.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feColorMatrix]{feColorMatrix}}}
}
}
\keyword{internal}

