% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feComponentTransferAmplitudeAttribute}
\alias{feComponentTransferAmplitudeAttribute}
\title{amplitude}
\description{
Sets the amplitude of the gamma function to be used when transfer type=gamma.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numericr>}{Specifies the amplitude to be used to construct the transfer function when the type is \emph{gamma.}}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=feFuncA]{feFuncA}}, \code{\link[=feFuncB]{feFuncB}}, \code{\link[=feFuncG]{feFuncG}}, \code{\link[=feFuncR]{feFuncR}}}
}
}

\keyword{internal}
