% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feComponentTransferOffsetAttribute}
\alias{feComponentTransferOffsetAttribute}
\title{offset}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{When type="gamma", the offset of the gamma function. If the attribute is not specified, then the effect is as if a value of 0 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feFuncA]{feFuncA}}, \code{\link[=feFuncB]{feFuncB}}, \code{\link[=feFuncG]{feFuncG}}, \code{\link[=feFuncR]{feFuncR}}}
}
}
\keyword{internal}

