% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feComponentTransferSlopeAttribute}
\alias{feComponentTransferSlopeAttribute}
\title{slope}
\description{
Sets the slope to be used when the transfer type is linear
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the slope to be used to construct the \emph{transfer} function when the type is \emph{linear}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=feFuncA]{feFuncA}}, \code{\link[=feFuncB]{feFuncB}}, \code{\link[=feFuncG]{feFuncG}}, \code{\link[=feFuncR]{feFuncR}}}
}
}

\keyword{internal}
