% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feComponentTransferTableValuesAttribute}
\alias{feComponentTransferTableValuesAttribute}
\title{tableValues}
\description{
Sets the lookup table to be used when the transfer type is table
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>[N]}{Specifies a numerical vector to be used to the construction of the \emph{transfer function} when the type is either \emph{""""""""""""""""table""""""""""""""""} or \emph{""""""""""""""""step""""""""""""""""}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=feFuncA]{feFuncA}}, \code{\link[=feFuncB]{feFuncB}}, \code{\link[=feFuncG]{feFuncG}}, \code{\link[=feFuncR]{feFuncR}}}
}
}

\keyword{internal}
