% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feComponentTransferTypeAttribute}
\alias{feComponentTransferTypeAttribute}
\title{type}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{identity | table | discrete | linear | gamma}{Indicates the type of component transfer function. The type of function determines the applicability of the other attributes. In the following, C is the initial component (e.g., 'feFuncR'), C' is the remapped component; both in the closed interval [0,1]. For identity. C' = C For table, the function is defined by linear interpolation between values given in the attribute 'tableValues'. The table has n+1 values (i.e., v0 to vn) specifying the start and end values for n evenly sized interpolation regions. Interpolations use the following formula. For a value C < 1 find k such that. k/n <= C < (k+1)/n The result C' is given by. C' = vk + (C . k/n)*n * (vk+1 . vk) If C = 1 then. C' = vn. For discrete, the function is defined by the step function given in the attribute 'tableValues', which provides a list of n values (i.e., v0 to vn.1) in order to identify a step function consisting of n steps. The step function is defined by the following formula. For a value C < 1 find k such that. k/n <= C < (k+1)/n The result C' is given by. C' = vk If C = 1 then. C' = vn.1. For linear, the function is defined by the following linear equation. C' = slope * C + intercept For gamma, the function is defined by the following exponential function. C' = amplitude * pow(C, exponent) + offset Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feFuncA]{feFuncA}}, \code{\link[=feFuncB]{feFuncB}}, \code{\link[=feFuncG]{feFuncG}}, \code{\link[=feFuncR]{feFuncR}}}
}
}
\keyword{internal}

