% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feCompositeK1234Attribute}
\alias{feCompositeK1234Attribute}
\title{k1234}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=feCompositeK1Attribute]{k1}}, \code{\link[=feCompositeK2Attribute]{k2}}, \code{\link[=feCompositeK3Attribute]{k3}}, \code{\link[=feCompositeK4Attribute]{k4}}
}

\section{Equivalence}{

\describe{
\item{}{k1234=c(value.K1,value.K2,value.K3,value.K4)}
\item{and}{k1 = value.K1; k2 = value.K2; k3 = value.K3; k4 = value.K4}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feComposite]{feComposite}}}
}
}
\keyword{internal}

