% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feCompositeK1Attribute}
\alias{feCompositeK1Attribute}
\title{k1}
\description{
Sets the value to be used for k1 when the composite operator is arithmetic.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the \emph{k1} coefficient for the \emph{arithmetic composite operator}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feComposite]{feComposite}}}
}
}

\keyword{internal}
