% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feCompositeK3Attribute}
\alias{feCompositeK3Attribute}
\title{k3}
\description{
Sets the value to be used for k3 when the composite operator is arithmetic.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the \emph{k3} coefficient for the \emph{arithmetic composite operator}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feComposite]{feComposite}}}
}
}

\keyword{internal}
