% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feCompositeOperatorAttribute}
\alias{feCompositeOperatorAttribute}
\title{operator}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{over | in | out | atop | xor | arithmetic}{The compositing operation that is to be performed. All of the 'operator' types except arithmetic match the corresponding operation as described in [PORTERDUFF]. The arithmetic operator is described above. If attribute 'operator' is not specified, then the effect is as if a value of over were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feComposite]{feComposite}}}
}
}
\keyword{internal}

