% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementBiasAttribute}
\alias{feConvolveMatrixElementBiasAttribute}
\title{bias}
\description{
Sets a bias shift which is to be applied, component-wise,  to the result of the convolution.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the value to be added to each image component after applying the \emph{kernelMatrix} and \emph{divisor.}}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
