% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementKernelUnitLengthAttribute}
\alias{feConvolveMatrixElementKernelUnitLengthAttribute}
\title{kernelUnitLength}
\description{
Governs the spacing between successive rows and columns in the kernel matrix, That spacing, dx dy ,is interpreted in the current filter units
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2]}{Specifies a dx, (dy) distance(s) in current filter units between successive rows and columns. Default for dx is 1 pixel, default for dy is dx.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
