% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feConvolveMatrixElementTARGETXYAttribute}
\alias{feConvolveMatrixElementTARGETXYAttribute}
\title{targetXY}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=feConvolveMatrixElementTargetXAttribute]{targetX}}, \code{\link[=feConvolveMatrixElementTargetYAttribute]{targetY}}
}

\section{Equivalence}{

\describe{
\item{}{targetXY=c(value.TARGETX,value.TARGETY)}
\item{and}{targetX = value.TARGETX; targetY = value.TARGETY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}
\keyword{internal}

