% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDiffuseLightingDiffuseConstantAttribute}
\alias{feDiffuseLightingDiffuseConstantAttribute}
\title{diffuseConstant}
\description{
Sets the kd term of the phong lighting model.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies a scaling factor for the resulting RGB components, It is the \eqn{kd} in \emph{Phong lighting model}. Default=1.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDiffuseLighting]{feDiffuseLighting}}}
}
}

\keyword{internal}
