% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDiffuseLightingKernelUnitLengthAttribute}
\alias{feDiffuseLightingKernelUnitLengthAttribute}
\title{kernelUnitLength}
\description{
Sets the dx dy values to be used in the surface normal calculations.
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2]}{Specifies the dx, (dy), to be used to build the implicit pixel grid for the filter. By default, dx is 1 pixel, dy is the value of dx.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDiffuseLighting]{feDiffuseLighting}}}
}
}

\keyword{internal}
