% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDisplacementMapXChannelSelectorAttribute}
\alias{feDisplacementMapXChannelSelectorAttribute}
\title{xChannelSelector}
\description{
Governs which channel (R|G|B|A) of the image given by in2 to be used for the computing the amount of displacement of of the image given by in1 in the x direction.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'R'}}{Specifies to use the R channel from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the x-axis.}
\item{\emph{'G'}}{Specifies to use the G channel from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the x-axis.}
\item{\emph{'B'}}{Specifies to use the B channel from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the x-axis.}
\item{\emph{'A'}}{Specifies to use the A channel from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the x-axis. (The default.)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDisplacementMap]{feDisplacementMap}}}
}
}

\keyword{internal}
