% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDisplacementMapYChannelSelectorAttribute}
\alias{feDisplacementMapYChannelSelectorAttribute}
\title{yChannelSelector}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{R | G | B | A}{Indicates which channel from 'in2' to use to displace the pixels in 'in' along the y.axis. If attribute 'yChannelSelector' is not specified, then the effect is as if a value of A were specified. Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDisplacementMap]{feDisplacementMap}}}
}
}
\keyword{internal}

