% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{feDistantLight}
\alias{feDistantLight}
\title{A primitive filter element that defines a distant light source}
\description{
This filter primitive define a distant light source that can be used within a lighting filter primitive : <fediffuselighting> or <fespecularlighting>.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Unclassified}}{\code{\link[=feDistantLightAzimuthAttribute]{azimuth}}, \code{\link[=feDistantLightElevationAttribute]{elevation}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=set]{set}}}
}
}

\keyword{element}
