% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDistantLightElevationAttribute}
\alias{feDistantLightElevationAttribute}
\title{elevation}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{Direction angle for the light source from the XY plane towards the z axis, in degrees. Note the positive Z.axis points towards the viewer of the content. If the attribute is not specified, then the effect is as if a value of 0 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feDistantLight]{feDistantLight}}}
}
}
\keyword{internal}

