% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{feFuncB}
\alias{feFuncB}
\title{Element used to specify blue component transfer function}
\description{
This filter primitive defines the transfer function for the blue component of the input graphic of its parent <fecomponenttransfer> element.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{transfer function element attributes}}{\code{\link[=feComponentTransferAmplitudeAttribute]{amplitude}}, \code{\link[=feComponentTransferExponentAttribute]{exponent}}, \code{\link[=feComponentTransferInterceptAttribute]{intercept}}, \code{\link[=feComponentTransferOffsetAttribute]{offset}}, \code{\link[=feComponentTransferSlopeAttribute]{slope}}, \code{\link[=feComponentTransferTableValuesAttribute]{tableValues}}, \code{\link[=feComponentTransferTypeAttribute]{type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=set]{set}}}
}
}
\keyword{element}

