% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feGaussianBlurStdDeviationAttribute}
\alias{feGaussianBlurStdDeviationAttribute}
\title{stdDeviation}
\description{
Sets the x and y standard deviations to be used by the Gaussian Blur algorithm..
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2]}{Specifies the standard deviations \eqn{\sigma_x}, \eqn{\sigma_y} for the blur operation along the x and y axis respectively. The default for \eqn{\sigma_x} is 1, the default for \eqn{\sigma_y} is \eqn{\sigma_x}.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feGaussianBlur]{feGaussianBlur}}}
}
}

\keyword{internal}
