% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feImageElementHrefAttribute}
\alias{feImageElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to a source image of the feImage element.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies an \emph{IRI} reference to the image source.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feImage]{feImage}}}
}
}

\keyword{internal}
