% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{feMergeNode}
\alias{feMergeNode}
\title{Element used to wrap a filter primitive prior to being used by femerge}
\description{
The feMergeNode takes the result of another filter to be processed by its parent <femerge>.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=set]{set}}}
}
}
\keyword{element}

