% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feMorphologyOperatorAttribute}
\alias{feMorphologyOperatorAttribute}
\title{operator}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{erode | dilate}{A keyword indicating whether to erode (i.e., thin) or dilate (fatten) the source graphic. If attribute 'operator' is not specified, then the effect is as if a value of erode were specified. Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feMorphology]{feMorphology}}}
}
}
\keyword{internal}

