% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feMorphologyRadiusAttribute}
\alias{feMorphologyRadiusAttribute}
\title{radius}
\description{
Governs the amount of  dilation or erodion in the x and y directions..
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2]}{Specifies the radii \eqn{r_x}, \eqn{r_y}, used in the x, y-directions to produce the \emph{erosion/dilation} effects. The default for \eqn{r_x} is 0, the default for \eqn{r_y} is \eqn{r_x}.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feMorphology]{feMorphology}}}
}
}

\keyword{internal}
