% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feOffsetDxAttribute}
\alias{feOffsetDxAttribute}
\title{dx}
\description{
Sets how much  of an  offset, the input graphics should have along the x-axis.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies to offset the input graphic by the amount dx along the x-axis, where dx is is given in the \emph{primitiveUnits} terms of the \strong{filter}. Default for dx is dx=0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feOffset]{feOffset}}}
}
}

\keyword{internal}
