% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feOffsetDXYAttribute}
\alias{feOffsetDXYAttribute}
\title{dxy}
\description{
Shortcut for assigning dx and dy
}
\section{Combines}{

\code{\link[=feOffsetDXYAttribute]{dx}}, \code{\link[=feOffsetDXYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
dxy =v is equivalent to  dx=v[1], dy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feOffset]{feOffset}}}
}
}

\keyword{internal}
