% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feOffsetDyAttribute}
\alias{feOffsetDyAttribute}
\title{dy}
\description{
Sets how much  of an  offset, the input graphics should have along the y-axis.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies to offset the input graphic by the amount dy along the y-axis, where dy is is given in the \emph{primitiveUnits} terms of the \emph{filter}. Default is dy=0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feOffset]{feOffset}}}
}
}

\keyword{internal}
