% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{fePointLight}
\alias{fePointLight}
\title{A Primitive Filter Element that defines a pointlight light source}
\description{
description=
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{combining attributes}}{\code{\link[=fePointLightXYZAttribute]{xyz}}}
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{filter primitive attributes}}{\code{\link[=fePointLightXAttribute]{x}}, \code{\link[=fePointLightYAttribute]{y}}}
\item{\emph{unclassified}}{\code{\link[=fePointLightZAttribute]{z}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=set]{set}}}
}
}
\keyword{element}

