% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{fePointLightZAttribute}
\alias{fePointLightZAttribute}
\title{z}
\description{
Sets the z-coordinate of the point light source.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the z-coordinate of the light source location with respect to the coordinates system on the filter produce by \emph{primitiveUnits}. (Default=0)}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=fePointLight]{fePointLight}}}
}
}

\keyword{internal}
