% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpecularLightingKernelUnitLengthAttribute}
\alias{feSpecularLightingKernelUnitLengthAttribute}
\title{kernelUnitLength}
\description{
Sets the dx dy values to be used in the surface normal calculations.
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2]}{Specifies the dx, (dy) to be used in calculating the unit normal vector. The default for dx is a small distance, the default for dy is dx. It is advised that at minimal, one should specifies filterRes or kernelUnitLength to assure consistency across displays.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}

\keyword{internal}
