% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpecularLightingSurfaceScaleAttribute}
\alias{feSpecularLightingSurfaceScaleAttribute}
\title{surfaceScale}
\description{
Sets the A_in scale factor used to produce the height of the surface.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies scaling factor to be used to scale the \emph{alpha channel} to produce the height of surface. Thus it is the maximum possible height of the surface and realized when the \emph{alpha channel} =1. The default is 1.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}

\keyword{internal}
