% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightLimitingConeAngleAttribute}
\alias{feSpotLightLimitingConeAngleAttribute}
\title{limitingConeAngle}
\description{
Sets the angle in degrees between the spot light axis and the spot light cone.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies an angle (in degrees) that defines a cone of light emulating a spot light. Default is no cone.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}

\keyword{internal}
