% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightPointsAtXAttribute}
\alias{feSpotLightPointsAtXAttribute}
\title{pointsAtX}
\description{
Sets the X-location of where the spot light is pointing.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<coordinate>}{Specifies the x-coordinate of where light source is point to. The coordinate is given in terms of the \emph{primitiveUnits} of the coordinate system  established on the \strong{filter element}. (Default=0)}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}

\keyword{internal}
