% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feSpotLightPOINTSATXYZAttribute}
\alias{feSpotLightPOINTSATXYZAttribute}
\title{pointsAtXYZ}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=feSpotLightPointsAtXAttribute]{pointsAtX}}, \code{\link[=feSpotLightPointsAtYAttribute]{pointsAtY}}, \code{\link[=feSpotLightPointsAtZAttribute]{pointsAtZ}}
}

\section{Equivalence}{

\describe{
\item{}{pointsAtXYZ=c(value.POINTSATX,value.POINTSATY,value.POINTSATZ)}
\item{and}{pointsAtX = value.POINTSATX; pointsAtY = value.POINTSATY; pointsAtZ = value.POINTSATZ}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}
\keyword{internal}

