% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightPointsAtZAttribute}
\alias{feSpotLightPointsAtZAttribute}
\title{pointsAtZ}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{Z location in the coordinate system established by attribute 'primitiveUnits' on the 'filter' element of the point at which the light source is pointing, assuming that, in the initial coordinate system, the positive Z.axis comes out towards the person viewing the content and assuming that one unit along the Z.axis equals one unit in X and Y. If the attribute is not specified, then the effect is as if a value of 0 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}
\keyword{internal}

