% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightSpecularExponentAttribute}
\alias{feSpotLightSpecularExponentAttribute}
\title{specularExponent}
\description{
Sets the exponent controlling the focus of the spot light.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the \emph{specular exponent}, which controls the shininess 0 is essentially ambient light, infinity approaching a mirror like surface. (Default=1).}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}

\keyword{internal}
