% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightXAttribute}
\alias{feSpotLightXAttribute}
\title{x}
\description{
Sets the x-coordinate of the spot light source.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the x-coordinate of the light source location with respect to the coordinates system on the \strong{filter} produced by \emph{primitiveUnits}. (Default=0)}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}

\keyword{internal}
