% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feSpotLightXYZAttribute}
\alias{feSpotLightXYZAttribute}
\title{xyz}
\description{
Shortcut for assigning x, y and z
}
\section{Combines}{

\code{\link[=feSpotLightXYZAttribute]{x}}, \code{\link[=feSpotLightXYZAttribute]{y}}, \code{\link[=feSpotLightXYZAttribute]{z}}
}

\section{Equivalence}{

\describe{
xyz =v is equivalent to  x=v[1], y=v[2], z=v[3]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}

\keyword{internal}
