% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceBaseFrequencyAttribute}
\alias{feTurbulenceBaseFrequencyAttribute}
\title{baseFrequency}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number.optional.number>}{The base frequency (frequencies) parameter(s) for the noise function. If two <number>s are provided, the first number represents a base frequency in the X direction and the second value represents a base frequency in the Y direction. If one number is provided, then that value is used for both X and Y. A negative value for base frequency is an error (see Error processing). If the attribute is not specified, then the effect is as if a value of 0 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}
\keyword{internal}

