% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceNumOctavesAttribute}
\alias{feTurbulenceNumOctavesAttribute}
\title{numOctaves}
\description{
Sets the numOctaves parameter to be used by the noise generation function for feTurbulence
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<integer>}{Specifies \emph{numOctaves} parameter to be used in noise function. The default is 1.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}

\keyword{internal}
