% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceStitchTilesAttribute}
\alias{feTurbulenceStitchTilesAttribute}
\title{stitchTiles}
\description{
Governs whether or not to stitch the tiles together when performing the turbulence operation..
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'stitch'}}{Specifies to the user agent to adjust  \emph{x-baseFrequency} \eqn{f_x} and \emph{y-baseFrequency} \eqn{f_y} values in such a way as to maintain a smooth transition at the border of the tiles containing the turbulence.}
\item{\emph{'noStitch'}}{Specifies to the user agent to make no attempt to achieve smooth transitions at the borders of the tiles which contain a turbulence function. (Default)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}

\keyword{internal}
