% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceTypeAttribute}
\alias{feTurbulenceTypeAttribute}
\title{type}
\description{
Governs which model of turbublence to use.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'fractalNoise'}}{Specifies to the filter primitive to use the \emph{noise} function.}
\item{\emph{'turbulence'}}{Specifies to the filter primitive to use the \emph{turbulence} function.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}

\keyword{internal}
