% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FillAttribute}
\alias{FillAttribute}
\title{fill}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{freeze}{The animation effect F(t) is defined to freeze the effect value at the last value of the active duration. The animation effect is "frozen" for the remainder of the document duration (or until the animation is restarted . see SMIL Animation. Restarting animation).}
\item{remove}{The animation effect is removed (no longer applied) when the active duration of the animation is over. After the active end of the animation, the animation no longer affects the target (unless the animation is restarted . see SMIL Animation. Restarting animation). This is the default value.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}
\keyword{internal}

