% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{filter-presentationAttribute}
\alias{filter-presentationAttribute}
\title{filter}
\description{
Defines which filter is to be applied to this element. (This element being a graphical element or a container such as **svg** or **g**).
}
\section{Available Attribute Values}{



\describe{
\item{<filter>}{Specifies  a \strong{filter} element as the value of the  \emph{filter} attribute.}
\item{<funciri>}{Specifies  a  reference to a \strong{filter} element given by functional IRI.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the filter from parent.}
\item{\emph{'\emph{'none'}'}}{Specifies that there is no  filter to be applied.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}, \code{\link[=use]{use}}}
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}, \code{\link[=glyph]{glyph}}, \code{\link[=marker]{marker}}, \code{\link[=missing-glyph]{missing.glyph}}, \code{\link[=pattern]{pattern}}, \code{\link[=switch]{switch}}}
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Container Elements}}{\code{\link[=defs]{defs}}, \code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
