% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterElementHeightAttribute}
\alias{FilterElementHeightAttribute}
\title{height}
\description{
Sets the height of the filter effects region.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the \emph{bounding box height} of the filter in terms of the filterUnits. (The default is 1.20).}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=filter]{filter}}}
}
}

\keyword{internal}
