% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterElementPrimitiveUnitsAttribute}
\alias{FilterElementPrimitiveUnitsAttribute}
\title{primitiveUnits}
\description{
Governs the units for coordinate system of the filter primitive subregion.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies to interpret \emph{x, y, width, height} ( or \emph{xy, wh}) as current user coordinates.}
\item{\emph{'objectBoundingBox'}}{Specifies to interpret \emph{x, y, width, height} ( or \emph{xy, wh}) as fractions (or percentages)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=filter]{filter}}}
}
}

\keyword{internal}
