% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterElementXAttribute}
\alias{FilterElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{See Filter effects region.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=filter]{filter}}}
}
}
\keyword{internal}

