% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{FilterPrimitiveIN12Attribute}
\alias{FilterPrimitiveIN12Attribute}
\title{in12}
\description{
Shortcut for assigning in and in2
}
\section{Combines}{

\code{\link[=FilterPrimitiveIN12Attribute]{in1}}, \code{\link[=FilterPrimitiveIN12Attribute]{in2}}
}

\section{Equivalence}{

\describe{
in12 =v is equivalent to  in=v[1], in2=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feBlend]{feBlend}}, \code{\link[=feComposite]{feComposite}}, \code{\link[=feDisplacementMap]{feDisplacementMap}}}
}
}

\keyword{internal}
