% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterPrimitiveIn2Attribute}
\alias{FilterPrimitiveIn2Attribute}
\title{in2}
\description{
Sets the secondary graphics input to be used for the given filter primitive.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'SourceGraphic'}}{Specifies taking as the second graphics input argument, \emph{in2}, the original graphic source.}
\item{\emph{'SourceAlpha'}}{Specifies taking as the second graphics input argument, \emph{in2}, taking the alpha channel of the original graphic source.}
\item{\emph{'BackgroundImage'}}{Specifies taking as the second  graphics input argument, \emph{in2}, the image snapshot of the filter region in effect at the time filter was invoked.}
\item{\emph{'BackgroundAlpha'}}{Specifies taking as the second  graphics input argument, \emph{in2}, the \emph{alpha channel} of the filter region image.}
\item{\emph{'FillPaint'}}{Specifies taking as the second  graphics input argument, \emph{in2}, the \emph{fill} property on the target element.}
\item{\emph{'StrokePaint'}}{Specifies taking as the second  graphics input argument, \emph{in2}, the \emph{stroke} property on the target element.}
\item{<filter-primitive-reference>}{Specifies taking as the second  graphics input argument, \emph{in2}, the results of the referenced filter primitive.}
\item{<filterPrimitveElement>}{Specifies taking as the second  graphics input argument, \emph{in2}, the results of the given filter primitive.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feBlend]{feBlend}}, \code{\link[=feComposite]{feComposite}}, \code{\link[=feDisplacementMap]{feDisplacementMap}}}
}
}

\keyword{internal}
