% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{flood-color-presentationAttribute}
\alias{flood-color-presentationAttribute}
\title{flood.color}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{currentColor |<color> [<icccolor>] | inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feFlood]{feFlood}}}
}
}
\keyword{internal}

