% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{flood-opacity-presentationAttribute}
\alias{flood-opacity-presentationAttribute}
\title{flood.opacity}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{<opacity-value>}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feFlood]{feFlood}}}
}
}
\keyword{internal}

