% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{font-face-name}
\alias{font-face-name}
\alias{font.face.name}
\title{Element used to specify a local font-face by name}
\description{
The font-face-name element points to a locally installed copy of this font, identified by its name. The font-face element is used within a font-face-src element.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{unclassified}}{\code{\link[=FontFaceNameElementNameAttribute]{name}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{Empty.}}
}
}
\keyword{element}

