% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{font-face-src}
\alias{font-face-src}
\alias{font.face.src}
\title{An element for defing the font face by specifing the source.}
\description{
The font-face-src element corresponds to the src property in CSS @font-face descriptions. It serves as container for <font-face-name>, pointing to locally installed copies of this font, and <font-face-uri>, utilizing remotely defined fonts.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face-name]{font.face.name}}, \code{\link[=font-face-uri]{font.face.uri}}}
}
}

\keyword{element}
