% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{font-face}
\alias{font-face}
\alias{font.face}
\title{An element for defining  the font face}
\description{
The font-face element corresponds to the CSS @font-face declaration. It defines a font's outer properties.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Transfer Function Element Attributes}}{\code{\link[=FontFaceElementSlopeAttribute]{slope}}}
\item{\emph{Unclassified}}{\code{\link[=FontFaceElementAccentHeightAttribute]{accent.height}}, \code{\link[=FontFaceElementAlphabeticAttribute]{alphabetic}}, \code{\link[=FontFaceElementAscentAttribute]{ascent}}, \code{\link[=FontFaceElementBboxAttribute]{bbox}}, \code{\link[=FontFaceElementCapHeightAttribute]{cap.height}}, \code{\link[=FontFaceElementDescentAttribute]{descent}}, \code{\link[=FontFaceElementFontFamilyAttribute]{font.family}}, \code{\link[=FontFaceElementFontSizeAttribute]{font.size}}, \code{\link[=FontFaceElementFontStretchAttribute]{font.stretch}}, \code{\link[=FontFaceElementFontStyleAttribute]{font.style}}, \code{\link[=FontFaceElementFontVariantAttribute]{font.variant}}, \code{\link[=FontFaceElementFontWeightAttribute]{font.weight}}, \code{\link[=FontFaceElementHangingAttribute]{hanging}}, \code{\link[=FontFaceElementIdeographicAttribute]{ideographic}}, \code{\link[=FontFaceElementMathematicalAttribute]{mathematical}}, \code{\link[=FontFaceElementOverlinePositionAttribute]{overline.position}}, \code{\link[=FontFaceElementOverlineThicknessAttribute]{overline.thickness}}, \code{\link[=FontFaceElementPanose1Attribute]{panose.1}}, \code{\link[=FontFaceElementStemhAttribute]{stemh}}, \code{\link[=FontFaceElementStemvAttribute]{stemv}}, \code{\link[=FontFaceElementStrikeThroughPositionAttribute]{strikethrough.position}}, \code{\link[=FontFaceElementStrikeThroughThicknessAttribute]{strikethrough.thickness}}, \code{\link[=FontFaceElementUnderlinePositionAttribute]{underline.position}}, \code{\link[=FontFaceElementUnderlineThicknessAttribute]{underline.thickness}}, \code{\link[=FontFaceElementUnicodeRangeAttribute]{unicode.range}}, \code{\link[=FontFaceElementUnitsPerEmAttribute]{units.per.em}}, \code{\link[=FontFaceElementVertAlphabeticAttribute]{v.alphabetic}}, \code{\link[=FontFaceElementVertHangingAttribute]{v.hanging}}, \code{\link[=FontFaceElementVertIdeographicAttribute]{v.ideographic}}, \code{\link[=FontFaceElementVertMathematicalAttribute]{v.mathematical}}, \code{\link[=FontFaceElementWidthsAttribute]{widths}}, \code{\link[=FontFaceElementXHeightAttribute]{x.height}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face-src]{font.face.src}}}
}
}

\keyword{element}
