% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-family-presentationAttribute}
\alias{font-family-presentationAttribute}
\title{font.family}
\description{
Governs the candidiates of  font families to be used in text rendering.
}
\section{Available Attribute Values}{



\describe{
\item{<family-name | generica-family>+}{Specifies a vector of consisting of  font-family names or a generic-font-family names.}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the font family from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
