% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-size-adjust-presentationAttribute}
\alias{font-size-adjust-presentationAttribute}
\title{font.size.adjust}
\description{
Govern the aspect values to preserve the when using a substitute font.
}
\section{Available Attribute Values}{



\describe{
\item{<number>}{Specifies the aspect ratio value(font-size divide by x-height) to preserve the x-height for the first font candidate substitution.}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies the to inherit the aspect ratio value from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'none'}'}'}'}}{Specifies to make no aspect ratio adjustment.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
