% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-stretch-presentationAttribute}
\alias{font-stretch-presentationAttribute}
\title{font.stretch}
\description{
Governs the amount of condensing or expansion in the glyphs used to render the text.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'condensed'}'}'}'}}{Specifies the font stretch rank=3}
\item{\emph{'\emph{'\emph{'\emph{'expanded'}'}'}'}}{Specifies the font stretch rank=7}
\item{\emph{'\emph{'\emph{'\emph{'extra-condensed'}'}'}'}}{Specifies the font stretch rank=2}
\item{\emph{'\emph{'\emph{'\emph{'extra-expanded'}'}'}'}}{Specifies the font stretch rank=8}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies the font  to inhert font-stretch from parent}
\item{\emph{'\emph{'\emph{'\emph{'narrower'}'}'}'}}{Specifies the font stretch rank=-1+stretch rank of parent}
\item{\emph{'\emph{'\emph{'\emph{'normal'}'}'}'}}{Specifies the font stretch rank=5 (default)}
\item{\emph{'\emph{'\emph{'\emph{'semi-condensed'}'}'}'}}{Specifies the font stretch rank=4}
\item{\emph{'\emph{'\emph{'\emph{'semi-expanded'}'}'}'}}{Specifies the font stretch rank=6}
\item{\emph{'\emph{'\emph{'\emph{'ultra-condensed'}'}'}'}}{Specifies the font stretch rank=1}
\item{\emph{'\emph{'\emph{'\emph{'ultra-expanded'}'}'}'}}{Specifies the font stretch rank=9}
\item{\emph{'\emph{'\emph{'\emph{'wider'}'}'}'}}{Specifies the font stretch rank=1+stretch rank of parent}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
