% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-variant-presentationAttribute}
\alias{font-variant-presentationAttribute}
\title{font.variant}
\description{
Governs the which font-variant to use for lower case characters:  normal glyphs or small-caps glyphs.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'inherit'}'}}{Inherit font-variant from the parent.}
\item{\emph{'\emph{'normal'}'}}{Specifies to use a normal variant style.}
\item{\emph{'\emph{'small-caps'}'}}{Specifies to use a small-caps  variant style.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
