% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-weight-presentationAttribute}
\alias{font-weight-presentationAttribute}
\title{font.weight}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{normal}{**ToDo!!!** }
\item{bold}{**ToDo!!!** }
\item{bolder}{**ToDo!!!** }
\item{lighter}{**ToDo!!!** }
\item{100}{**ToDo!!!** }
\item{200}{**ToDo!!!** }
\item{300}{**ToDo!!!** }
\item{400}{**ToDo!!!** }
\item{500}{**ToDo!!!** }
\item{600}{**ToDo!!!** }
\item{700}{**ToDo!!!** }
\item{800}{**ToDo!!!** }
\item{900}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

