% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontElementHorizAdvXAttribute}
\alias{FontElementHorizAdvXAttribute}
\title{horiz.adv.x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The default horizontal advance after rendering a glyph in horizontal orientation. Glyph widths are required to be non.negative, even if the glyph is typically rendered right.to.left, as in Hebrew and Arabic scripts.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font]{font}}}
}
}
\keyword{internal}

