% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontElementHorizOriginXAttribute}
\alias{FontElementHorizOriginXAttribute}
\title{horiz.origin.x}
\description{
Sets the x-coordinate of the origin to be used when rendering the horizontally oriented text.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the glyph origin relative to the font coordinate system in horizontal text mode. Default=0.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font]{font}}}
}
}

\keyword{internal}
