% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{FontElementHORIZORIGINXYAttribute}
\alias{FontElementHORIZORIGINXYAttribute}
\title{horiz.origin.xy}
\description{
Shortcut for assigning horiz-origin-x and horiz-origin-y
}
\section{Combines}{

\code{\link[=FontElementHORIZORIGINXYAttribute]{horiz.origin.x}}, \code{\link[=FontElementHORIZORIGINXYAttribute]{horiz.origin.y}}
}

\section{Equivalence}{

\describe{
horiz-origin-xy =v is equivalent to  horiz-origin-x=v[1], horiz-origin-y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font]{font}}}
}
}

\keyword{internal}
