% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontElementVertAdvYAttribute}
\alias{FontElementVertAdvYAttribute}
\title{vert.adv.y}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The default vertical advance after rendering a glyph in vertical orientation. If the attribute is not specified, the effect is as if a value equivalent of one em were specified (see 'units.per.em').Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font]{font}}}
}
}
\keyword{internal}

