% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontElementVertOriginXAttribute}
\alias{FontElementVertOriginXAttribute}
\title{vert.origin.x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The default X.coordinate in the font coordinate system of the origin of a glyph to be used when drawing vertically oriented text. If the attribute is not specified, the effect is as if the attribute were set to half of the effective value of attribute 'horiz.adv.x'.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font]{font}}}
}
}
\keyword{internal}

