% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{FontElementVERTORIGINXYAttribute}
\alias{FontElementVERTORIGINXYAttribute}
\title{vert.origin.xy}
\description{
Shortcut for assigning vert-origin-x and vert-origin-y
}
\section{Combines}{

\code{\link[=FontElementVERTORIGINXYAttribute]{vert.origin.x}}, \code{\link[=FontElementVERTORIGINXYAttribute]{vert.origin.y}}
}

\section{Equivalence}{

\describe{
vert-origin-xy =v is equivalent to  vert-origin-x=v[1], vert-origin-y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font]{font}}}
}
}

\keyword{internal}
