% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementAccentHeightAttribute}
\alias{FontFaceElementAccentHeightAttribute}
\title{accent.height}
\description{
Sets the distance from the font origin to the top of the accent characters.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the distance from the font origin to the top of the accent characters, given in the units of the font coordinate system.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
