% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementAlphabeticAttribute}
\alias{FontFaceElementAlphabeticAttribute}
\title{alphabetic}
\description{
Sets the alignment coordinate for glyphs to achieve an alphabetic baseline alignment
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies an amount of alignment to be used to produce an ideographic baseline alignment, given that the \emph{glyphs} are  horizontally orientated.  The units of this value is the same as the units given by the font coordinate system.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
