% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementAscentAttribute}
\alias{FontFaceElementAscentAttribute}
\title{ascent}
\description{
Sets the maximum unaccented height of the font, as measured within the font coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the maximum height for an unaccented \emph{glyph} where the height is given in the same units of the font coordinate system.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
