% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementCapHeightAttribute}
\alias{FontFaceElementCapHeightAttribute}
\title{cap.height}
\description{
Sets the height of the uppercase glyphs of the fon. The height is measured in terms of the font coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the height of uppercase \emph{glyphs} of the font using \emph{units per em}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
