% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontFamilyAttribute}
\alias{FontFaceElementFontFamilyAttribute}
\title{font.family}
\description{
Sets a prioritized list of font families.  This instructs the user agent to use  first available font in the list.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<character>[N]}{Specifies, as a character vector, one or more font family names.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
