% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontStretchAttribute}
\alias{FontFaceElementFontStretchAttribute}
\title{font.stretch}
\description{
Governs the width or narrowness, of the font face relative to others within the same font family.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'all'}}{Specifies to apply all font factors.}
\item{\emph{'normal'}}{Specifies no stretching and no condensing of the \emph{glyphs}.}
\item{\emph{'ultra-condensed'}}{Specifies maximum condensing (ultra) of the \emph{glyphs}.}
\item{\emph{'extra-condensed'}}{Specifies a slightly more condensed than \emph{condensed}.}
\item{\emph{'condensed'}}{Specifies a slightly more condensed than \emph{semi-condensed}.}
\item{\emph{'semi-condensed'}}{Specifies a slightly more condensed than \emph{normal}.}
\item{\emph{'semi-expanded'}}{Specifies a slightly more expanded than \emph{normal}.}
\item{\emph{'expanded'}}{Specifies a slightly more expanded than the \emph{semi-expanded}  amount of stretching.}
\item{\emph{'extra-expanded'}}{Specifies a slightly more expanded than \emph{expanded}  amount of stretching.}
\item{\emph{'ultra-expanded'}}{Specifies the maximum amount of stretching.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
