% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontStyleAttribute}
\alias{FontFaceElementFontStyleAttribute}
\title{font.style}
\description{
Governs the font style ( for example, italic, bold, )
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'all'}}{Specifies all three font styles.}
\item{\emph{'normal'}}{Specifies a \emph{normal} font style.}
\item{\emph{'italic'}}{Specifies an \emph{italic} font style}
\item{\emph{'oblique'}}{Specifies an \emph{oblique} font style.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
