% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontVariantAttribute}
\alias{FontFaceElementFontVariantAttribute}
\title{font.variant}
\description{
Governs whether or not to use a small-caps font variant for bicameral fonts.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'normal'}}{Specifies the \emph{normal} font variant.}
\item{\emph{'small-caps'}}{Specifies \emph{small-caps} variant of the font.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
