% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontWeightAttribute}
\alias{FontFaceElementFontWeightAttribute}
\title{font.weight}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{all | [normal | bold | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900] [, [normal | bold | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900]]*}{Same syntax and semantics as the 'font.weight' descriptor within an font.face rule. The weight of a face relative to others in the same font family. Takes on the same values as the 'font.weight' property with three exceptions. relative keywords (bolder, lighter) are not permitteda comma.separated list of values is permitted, for fonts that contain multiple weightsan additional keyword, 'all', is permitted, which means that the font will match for all possible weights; either because it contains multiple weights, or because that face only has a single weight. If the attribute is not specified, the effect is as if a value of 'all' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

