% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementHangingAttribute}
\alias{FontFaceElementHangingAttribute}
\title{hanging}
\description{
Sets the alignment coordinate offset for a horizontally oriented glyph layout in order to achieve a hanging baseline alignment.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the vaule of the alignment amount to use to to produce an ideographic baseline alignment. This amount is given in the same units as in the font coordinate system. This alignment is applicable for horizontally orientated \emph{glyphs}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
