% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementIdeographicAttribute}
\alias{FontFaceElementIdeographicAttribute}
\title{ideographic}
\description{
Sets an offset within the font coordinate system to provide an alignment of a horizontally oriented glyph layout, with the goal of achieving an ideographic baseline alignment.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies, in font coordinates, for horizontally orientated \emph{glyphs}, the alignment to produce an ideographic baseline alignment.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
