% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementMathematicalAttribute}
\alias{FontFaceElementMathematicalAttribute}
\title{mathematical}
\description{
Sets an offset within the font coordinate system to provide an alignment of a horizontally oriented glyph layout, with the goal of achieving an mathematical baseline alignment.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies, in font coordinates, for horizontally orientated \emph{glyphs}, the alignment to produce an \emph{ideographic} baseline alignment.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
