% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementOverlinePositionAttribute}
\alias{FontFaceElementOverlinePositionAttribute}
\title{overline.position}
\description{
Sets the desired overline position, as measured within the font-coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies, in font coordinates, the intended \emph{overline position}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
