% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementPanose1Attribute}
\alias{FontFaceElementPanose1Attribute}
\title{panose.1}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{[<integer>]{10}}{Same syntax and semantics as the 'panose.1' descriptor within an font.face rule. The Panose.1 number, consisting of ten decimal integers, separated by whitespace. Except for any additional information provided in this specification, the normative definition of the attribute is in CSS2 ([CSS2], section 15.3.6). If the attribute is not specified, the effect is as if a value of '0 0 0 0 0 0 0 0 0 0' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

