% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementStemvAttribute}
\alias{FontFaceElementStemvAttribute}
\title{stemv}
\description{
Sets the vertical stem width.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies dominant vertical stem width of the \emph{glyphs} via \emph{units per em}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
