% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementStrikeThroughPositionAttribute}
\alias{FontFaceElementStrikeThroughPositionAttribute}
\title{strikethrough.position}
\description{
Sets the preferred positioning within the font-coordinate system for rendering a strikethrough.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies, in font coordinates, the intended \emph{strikethrough} position.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
