% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementStrikeThroughThicknessAttribute}
\alias{FontFaceElementStrikeThroughThicknessAttribute}
\title{strikethrough.thickness}
\description{
Sets the preferred thickness using units from the font-coordinate system tp render a strikethrough.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies, in font coordinates, the intended \emph{strikethrough} thickness.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
