% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnderlinePositionAttribute}
\alias{FontFaceElementUnderlinePositionAttribute}
\title{underline.position}
\description{
Sets the preferred positioning using  the font-coordinates for rendering an underline.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies, in font coordinates, the intended \emph{underline position}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
