% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnderlineThicknessAttribute}
\alias{FontFaceElementUnderlineThicknessAttribute}
\title{underline.thickness}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The ideal thickness of an underline, expressed as a length within the font coordinate system.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

