% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnicodeRangeAttribute}
\alias{FontFaceElementUnicodeRangeAttribute}
\title{unicode.range}
\description{
Sets the range of ISO 10646 characters [UNICODE] to be covered by the glyphs in the font.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<urange>[N] (vector-of-uranges)}{Specifies range of ISO 10646 characters covered by the \emph{font}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
