% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnitsPerEmAttribute}
\alias{FontFaceElementUnitsPerEmAttribute}
\title{units.per.em}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{Same syntax and semantics as the 'units.per.em' descriptor within an font.face rule. The number of coordinate units on the em square, the size of the design grid on which glyphs are laid out. This value is almost always necessary as nearly every other attribute requires the definition of a design grid. If the attribute is not specified, the effect is as if a value of '1000' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

