% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementVertMathematicalAttribute}
\alias{FontFaceElementVertMathematicalAttribute}
\title{v.mathematical}
\description{
Sets an alignment coordinate adjustment for a vertically oriented glyph layout to aid in achieving an  mathematical baseline alignment.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies, in font coordinates, for vertically orientated \emph{glyphs}, the alignment to produce an \emph{mathematical baseline alignment}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
