% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementWidthsAttribute}
\alias{FontFaceElementWidthsAttribute}
\title{widths}
\description{
Sets the glyphs widths for each individual glyph in terms of the design grid.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<named-list-of-widths>}{Specifies \emph{glyphs} widths, given a \emph{urange} an a number representing the \emph{width.} The format is a list with names as \emph{uranges} and values as width.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
