% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementXHeightAttribute}
\alias{FontFaceElementXHeightAttribute}
\title{x.height}
\description{
Sets the height of the lowercase glyphs of the given font, where the value of the height is in terms of the units used by the font coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the height of lowercase \emph{glyphs} of the font using \emph{units per em}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
