% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceFormatElementStringAttribute}
\alias{FontFaceFormatElementStringAttribute}
\title{string}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<anything>}{The 'string' attribute is a hint to the user agent, and specifies a list of formats that the font referenced by the parent 'font.face.uri' element supports. The syntax of the attribute value is a format string as defined in CSS2, such as 'truetype'. Refer to the description of the 'src' descriptor in CSS2 for details on how the format hint is interpreted ([CSS2], section 15.3.5).Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face-format]{font.face.format}}}
}
}
\keyword{internal}

