% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceNameElementNameAttribute}
\alias{FontFaceNameElementNameAttribute}
\title{name}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<anything>}{The 'name' attribute specifies the name of a local font. Unlike the syntax allowed between the parentheses of the local(...) clause in an font.face rule 'src' descriptor, the font name specified in this attribute is not surrounded in single or double quotes. Refer to the description of the 'src' descriptor in CSS2 for details on how the font name is interpreted ([CSS2], section 15.3.5).Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face-name]{font.face.name}}}
}
}
\keyword{internal}

