% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceUriElementHrefAttribute}
\alias{FontFaceUriElementHrefAttribute}
\title{xlink.href}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<IRI>}{The 'xlink.href' attribute specifies the location of the referenced font.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face-uri]{font.face.uri}}}
}
}
\keyword{internal}

