% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ForeignObjectElementHeightAttribute}
\alias{ForeignObjectElementHeightAttribute}
\title{height}
\description{
Sets the height of the rectangular region into which the referenced document is placed.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the \emph{height} of the rectangular region into which into which the \emph{'foreignObject'} will be rendered. Requires \eqn{height \\geq 0}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=foreignObject]{foreignObject}}}
}
}

\keyword{internal}
