% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ForeignObjectElementWidthAttribute}
\alias{ForeignObjectElementWidthAttribute}
\title{width}
\description{
Set the width of the region into which the  foreign object is to be placed.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the \emph{width} of the rectangular region into which into which the \emph{'foreignObject'} will be rendered. Requires \eqn{width \\geq 0}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=foreignObject]{foreignObject}}}
}
}

\keyword{internal}
