% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ForeignObjectElementXAttribute}
\alias{ForeignObjectElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the left side of rectangular region where the foreign object is to be placed.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the left side of the rectangular region into which the \emph{'foreignObject'} will be rendered. Default=0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=foreignObject]{foreignObject}}}
}
}

\keyword{internal}
