% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FromAttribute}
\alias{FromAttribute}
\title{from}
\description{
Sets the starting value of the animation.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<attribute-value>}{Specifies the initial value of the attribute chosen for animation.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
