% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgRGetNode.R
\name{getsafeNodeAttr}
\alias{getsafeNodeAttr}
\title{Get the attribute value having the name attr.name in the given node}
\usage{
getsafeNodeAttr(attr.name, node)
}
\arguments{
\item{attr.name}{the name of the attribute}

\item{node}{the node on which the value is sought}
}
\description{
Get the attribute value having the name attr.name in the given node
}
\note{
If the attribute with name attr.name fails to exits on the
given node, then an attribute id is generated and added to  the
node
}

