% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{glyph-orientation-horizontal-presentationAttribute}
\alias{glyph-orientation-horizontal-presentationAttribute}
\title{glyph.orientation.horizontal}
\description{
Controls glyph orientation when the inline-progression-direction is horizontal
}
\section{Available Attribute Values}{



\describe{
\item{<angle>}{Sets the angle of orientation (restricted to 0, 90, 180, and 270 degrees).}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the glyph orientation  from the  parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
