% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementArabicFormAttribute}
\alias{GlyphElementArabicFormAttribute}
\title{arabic.form}
\description{
Governs which of the four possible forms for an arabic glyph
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'initial'}}{Specifies the initial \emph{Arabic} \emph{glyph} form.}
\item{\emph{'medial'}}{Specifies the medial \emph{Arabic} \emph{glyph} form.}
\item{\emph{'terminal'}}{Specifies the terminal \emph{Arabic} \emph{glyph} form.}
\item{\emph{'isolated'}}{Specifies the isolated \emph{Arabic} \emph{glyph} form.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}}
}
}

\keyword{internal}
