% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementGlyphNameAttribute}
\alias{GlyphElementGlyphNameAttribute}
\title{glyph.name}
\description{
Sets a name for a glyph
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<glyph-names>[N]}{Specifies a vector of one or more glyph names}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}}
}
}

\keyword{internal}
