% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementHorizAdvXAttribute}
\alias{GlyphElementHorizAdvXAttribute}
\title{horiz.adv.x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The horizontal advance after rendering the glyph in horizontal orientation. If the attribute is not specified, the effect is as if the attribute were set to the value of the font's 'horiz.adv.x' attribute. Glyph widths are required to be non.negative, even if the glyph is typically rendered right.to.left, as in Hebrew and Arabic scripts.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyph]{glyph}}, \code{\link[=missing-glyph]{missing.glyph}}}
}
}
\keyword{internal}

