% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementLangAttribute}
\alias{GlyphElementLangAttribute}
\title{lang}
\description{
Provides a set of languages ot the user agent considered to be supported by this glyph.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<language-codes>}{Specifies a language code name as defined in BCP 47.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}}
}
}

\keyword{internal}
