% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementOrientationAttribute}
\alias{GlyphElementOrientationAttribute}
\title{orientation}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{h | v}{Indicates that the given glyph is only to be used for a particular inline.progression.direction (i.e., horizontal or vertical). If the attribute is not specified, then the glyph can be used in all cases (i.e., both horizontal and vertical inline.progression.direction).Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyph]{glyph}}}
}
}
\keyword{internal}

