% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementUnicodeAttribute}
\alias{GlyphElementUnicodeAttribute}
\title{unicode}
\description{
Sets a string of Unicode characters that are to correspond to this glyph.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<string>}{Specifies one or more Unicode characters indicating the sequence of Unicode characters which corresponds to this \emph{glyph.}}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}}
}
}

\keyword{internal}
