% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementVertAdvYAttribute}
\alias{GlyphElementVertAdvYAttribute}
\title{vert.adv.y}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The vertical advance after rendering a glyph in vertical orientation. If the attribute is not specified, the effect is as if the attribute were set to the value of the font's 'vert.adv.y' attribute.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyph]{glyph}}, \code{\link[=missing-glyph]{missing.glyph}}}
}
}
\keyword{internal}

