% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementVertOriginYAttribute}
\alias{GlyphElementVertOriginYAttribute}
\title{vert.origin.y}
\description{
Sets the y-coordinate of the glyphs origin to be applied to a vertically orientented text. This value is to be interpreted within the fonts coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the default y-coordinate of the origin for a \emph{glyph} relative to coordinate system when in text orientation is vertical. Default is value of vert-origin-y of font.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}, \code{\link[=missing-glyph]{missing.glyph}}}
}
}

\keyword{internal}
