% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{GlyphRefElementDXYAttribute}
\alias{GlyphRefElementDXYAttribute}
\title{dxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=GlyphRefElementDXAttribute]{dx}}, \code{\link[=GlyphRefElementDYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
\item{}{dxy=c(value.DX,value.DY)}
\item{and}{dx = value.DX; dy = value.DY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{internal}

