% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementGlyphRefAttribute}
\alias{GlyphRefElementGlyphRefAttribute}
\title{glyphRef}
\description{
Sets the glyph reference identifier.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<string>}{Specifies a \emph{glyph} identifier.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
