% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementHrefAttribute}
\alias{GlyphRefElementHrefAttribute}
\title{xlink.href}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<iri>}{An IRI reference to a 'glyph' element in an SVG document fragment. The referenced 'glyph' is rendered as an alternate glyph.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{internal}

