% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementXAttribute}
\alias{GlyphRefElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{This value represents the new absolute X coordinate within the font's coordinate system for this glyph. The font coordinate system is based on the em square model described in the Fonts chapter of CSS2 ([CSS2], chapter 15). If the attribute is not specified, for the first 'glyphRef' child element, the effect is as if the attribute were set to "0", whereas for subsequent 'glyphRef' child elements, the effect is as if the attribute were set to the end X coordinate from the previous 'glyphRef' element.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{internal}

