% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{GlyphRefElementXYAttribute}
\alias{GlyphRefElementXYAttribute}
\title{xy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=GlyphRefElementXAttribute]{x}}, \code{\link[=GlyphRefElementYAttribute]{y}}
}

\section{Equivalence}{

\describe{
\item{}{xy=c(value.X,value.Y)}
\item{and}{x = value.X; y = value.Y}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{internal}

