% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementYAttribute}
\alias{GlyphRefElementYAttribute}
\title{y}
\description{
Sets the y-coordinates for the alternative glyphs..
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the y-coordinate for the \emph{glyph} in terms of the fonts coordinate system.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
