% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgRCompound.R
\name{\%<c-\%}
\alias{\%<c-\%}
\title{Convenience operator for creating compounds}
\usage{
name \%<c-\% f
}
\arguments{
\item{name}{the name to assign to the component}

\item{f}{an R function}
}
\value{
A function that evaluates the body of f in the calling environment.
}
\description{
Convenience operator for creating compounds
}

