% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeGraphPaper.R
\name{graphPaper}
\alias{graphPaper}
\title{graphPaper Internal Compound}
\usage{
graphPaper(wh, dxy, labels)
}
\arguments{
\item{wh}{numeric vector for the width-heigth}

\item{dxy}{numeric vector for the spacing between lines}

\item{labels}{add coordinate labels(default=FALSE)}
}
\description{
Internal compound used to add a background with graphpaper appearance.
}
\keyword{display}
\keyword{helper}
