% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{hkern}
\alias{hkern}
\title{An  Element used to specify horizontal distances between glyphs}
\description{
The horizontal distance between two glyphs can be fine-tweaked with an hkern Element. This process is known as Kerning.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{combining attributes}}{\code{\link[=HKernElementG12Attribute]{g12}}, \code{\link[=HKernElementU12Attribute]{u12}}}
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{unclassified}}{\code{\link[=HKernElementG1Attribute]{g1}}, \code{\link[=HKernElementG2Attribute]{g2}}, \code{\link[=HKernElementKAttribute]{k}}, \code{\link[=HKernElementU1Attribute]{u1}}, \code{\link[=HKernElementU2Attribute]{u2}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{Empty.}}
}
}
\keyword{element}

