% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{HKernElementG12Attribute}
\alias{HKernElementG12Attribute}
\title{g12}
\description{
Shortcut for assigning g1 and g2
}
\section{Combines}{

\code{\link[=HKernElementG12Attribute]{g1}}, \code{\link[=HKernElementG12Attribute]{g2}}
}

\section{Equivalence}{

\describe{
g12 =v is equivalent to  g1=v[1], g2=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}

\keyword{internal}
