% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{HKernElementU12Attribute}
\alias{HKernElementU12Attribute}
\title{u12}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=HKernElementU1Attribute]{u1}}, \code{\link[=HKernElementU2Attribute]{u2}}
}

\section{Equivalence}{

\describe{
\item{}{u12=c(value.U1,value.U2)}
\item{and}{u1 = value.U1; u2 = value.U2}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}
\keyword{internal}

