% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{HrefAttribute}
\alias{HrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to the animations target element.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies animation target via an \emph{IRI} reference}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
