% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{ImageElementWHAttribute}
\alias{ImageElementWHAttribute}
\title{wh}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=ImageElementHeightAttribute]{height}}, \code{\link[=ImageElementWidthAttribute]{width}}
}

\section{Equivalence}{

\describe{
\item{}{wh=c(value.HEIGHT,value.WIDTH)}
\item{and}{height = value.HEIGHT; width = value.WIDTH}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=image]{image}}}
}
}
\keyword{internal}

