% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ImageElementXAttribute}
\alias{ImageElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the left side of the rectangular region where the image is to be placed.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the left hand side rectangular region into which the image element is to be placed.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}}
}
}

\keyword{internal}
