% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{kerning-presentationAttribute}
\alias{kerning-presentationAttribute}
\title{kerning}
\description{
Governs kerning:  inter-glyph spacing adjustments.
}
\section{Available Attribute Values}{



\describe{
\item{<length>}{Sets a value for the length for the inter-character spacing. Setting to 0, essentially diables auto kerning.}
\item{\emph{'\emph{'auto'}'}}{Specifies to auto adjust adjust inter-glyph spacing by using kerning tables includess with the font.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the inter-character spacing adjustment from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
