% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{KeyPointsAttribute}
\alias{KeyPointsAttribute}
\title{keyPoints}
\description{
Sets the points corresponding to the keyTime values. The keyTime and keyPoints together form a contraint that at each specific keyTime, the object should be located at the corresponding key point. This constaint may cause the object to move faster or slower along the path between the intervals defined by the keytimes/keypoints. Another of saying this is that keyTime-keyPoints are used to control the pacing of the motion along the path.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric-in-the-[0,1]-range>[N]}{Specifies points as a numeric vector with values in the (0-1) range, with 0 being the start, and 1 being the end. The length of this value should match in length of the keytimes attribute value.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateMotion]{animateMotion}}}
}
}

\keyword{internal}
