% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{KeySplinesAttribute}
\alias{KeySplinesAttribute}
\title{keySplines}
\description{
Sets a set of Bzier control points associated with the keyTimes attribute.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<coordinates>[[N]][4]}{A list of Bezier control points associated with the \emph{'keyTimes'} list. Each pt has 4 values, \emph{x1, y1, x2, y2} with values in the range between 0 and 1.}
\item{<coordinates>[4,N]}{A 4xN matrix whose columns are the Bezier control points associated with the \emph{'keyTimes'} list. Each pt has 4 values, x1, y1, x2, y2 with values in the range between 0 and 1.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
