% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{letter-spacing-presentationAttribute}
\alias{letter-spacing-presentationAttribute}
\title{letter.spacing}
\description{
Governs supplemental to kerning spacing between letters.
}
\section{Available Attribute Values}{



\describe{
\item{<length>}{Sets a value for the amount of supplemental (to kerning) spacing between text characters.}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit  the value of letter-spacing adjustment from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'normal'}'}'}'}}{Specifies no supplemental (to kerning) spacing between text characters.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
