% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{lighting-color-presentationAttribute}
\alias{lighting-color-presentationAttribute}
\title{lighting.color}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{currentColor |<color> [<icccolor>] | inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDiffuseLighting]{feDiffuseLighting}}, \code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}
\keyword{internal}

