% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{linearGradient}
\alias{linearGradient}
\title{An element to define a linear gradient}
\description{
LinearGradient is used to define a linear gradient fill or stroke of  a graphical element.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Complementary Attributes}}{\code{\link[=LinearGradientElementX12Attribute]{x12}}, \code{\link[=LinearGradientElementXY1Attribute]{xy1}}, \code{\link[=LinearGradientElementXY2Attribute]{xy2}}, \code{\link[=LinearGradientElementY12Attribute]{y12}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=LinearGradientElementGradientTransformAttribute]{gradientTransform}}, \code{\link[=LinearGradientElementGradientUnitsAttribute]{gradientUnits}}, \code{\link[=LinearGradientElementSpreadMethodAttribute]{spreadMethod}}, \code{\link[=StyleAttribute]{style}}, \code{\link[=LinearGradientElementX1Attribute]{x1}}, \code{\link[=LinearGradientElementX2Attribute]{x2}}, \code{\link[=LinearGradientElementY1Attribute]{y1}}, \code{\link[=LinearGradientElementY2Attribute]{y2}}}
\item{\emph{Xlink Attributes}}{\code{\link[=XLinkArcRoleAttribute]{xlink.arcrole}}, \code{\link[=LinearGradientElementHrefAttribute]{xlink.href}}, \code{\link[=XLinkRoleAttribute]{xlink.role}}, \code{\link[=XLinkTitleAttribute]{xlink.title}}, \code{\link[=XLinkTypeAttribute]{xlink.type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=stop]{stop}}}
}
}

\keyword{element}
