% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementX12Attribute}
\alias{LinearGradientElementX12Attribute}
\title{x12}
\description{
Shortcut for assigning x1 and x2
}
\section{Combines}{

\code{\link[=LinearGradientElementX12Attribute]{x1}}, \code{\link[=LinearGradientElementX12Attribute]{x2}}
}

\section{Equivalence}{

\describe{
x12 =v is equivalent to  x1=v[1], x2=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
