% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementXY2Attribute}
\alias{LinearGradientElementXY2Attribute}
\title{xy2}
\description{
Shortcut for assigning x2 and y2
}
\section{Combines}{

\code{\link[=LinearGradientElementXY2Attribute]{x2}}, \code{\link[=LinearGradientElementXY2Attribute]{y2}}
}

\section{Equivalence}{

\describe{
xy2 =v is equivalent to  x2=v[1], y2=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
