% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementY12Attribute}
\alias{LinearGradientElementY12Attribute}
\title{y12}
\description{
Shortcut for assigning y1 and y2
}
\section{Combines}{

\code{\link[=LinearGradientElementY12Attribute]{y1}}, \code{\link[=LinearGradientElementY12Attribute]{y2}}
}

\section{Equivalence}{

\describe{
y12 =v is equivalent to  y1=v[1], y2=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
