% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{LinearGradientElementY1Attribute}
\alias{LinearGradientElementY1Attribute}
\title{y1}
\description{
Sets the y-coordinate of the starting point for a linear gradient.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-coordinate of the starting point of the \emph{gradient-vector} of a \strong{linearGradient element}. Can be numeric or percentage.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
