% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LineElementXY2Attribute}
\alias{LineElementXY2Attribute}
\title{xy2}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=LineElementX2Attribute]{x2}}, \code{\link[=LineElementY2Attribute]{y2}}
}

\section{Equivalence}{

\describe{
\item{}{xy2=c(value.X2,value.Y2)}
\item{and}{x2 = value.X2; y2 = value.Y2}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}}
}
}
\keyword{internal}

