% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{marker-mid-presentationAttribute}
\alias{marker-mid-presentationAttribute}
\title{marker.mid}
\description{
Defines the marker  that is to be rendered at the final vertex of the given **path** element or basic shape.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'none'}'}}{Specifies for no \emph{marker} symbol to be rendered.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{marker} symbol from the parent.}
\item{<funciri>}{Sets a functional IRI reference to a \strong{marker element} to be used..}
\item{<marker-element>}{Sets the \strong{marker element} to be used.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}}
}
}

\keyword{internal}
