% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{marker properties-presentationAttribute}
\alias{marker properties-presentationAttribute}
\title{marker properties}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{none}{**ToDo!!!** }
\item{<funciri>}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}}
}
}
\keyword{internal}

