% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{marker-start-presentationAttribute}
\alias{marker-start-presentationAttribute}
\title{marker.start}
\description{
Defines the marker that is to be rendered at the first vertex of the given **path** element or basic shape.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'none'}'}'}'}}{Specifies for no marker symbol to be rendered.}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the marker symbol from the parent.}
\item{<funciri>}{Sets a functional IRI reference to a marker element to be used..}
\item{<marker-element>}{Sets the \strong{marker element} to be used.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}}
}
}

\keyword{internal}
