% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{MarkerElementREFXYAttribute}
\alias{MarkerElementREFXYAttribute}
\title{refXY}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=MarkerElementRefXAttribute]{refX}}, \code{\link[=MarkerElementRefYAttribute]{refY}}
}

\section{Equivalence}{

\describe{
\item{}{refXY=c(value.REFX,value.REFY)}
\item{and}{refX = value.REFX; refY = value.REFY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=marker]{marker}}}
}
}
\keyword{internal}

