% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerElementRefYAttribute}
\alias{MarkerElementRefYAttribute}
\title{refY}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{The y.axis coordinate of the reference point which is to be aligned exactly at the marker position. The coordinate is defined in the coordinate system after application of the 'viewBox' and 'preserveAspectRatio' attributes. If the attribute is not specified, the effect is as if a value of "0" were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=marker]{marker}}}
}
}
\keyword{internal}

