% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerHeightAttribute}
\alias{MarkerHeightAttribute}
\title{markerHeight}
\description{
Sets the height of the marker viewport.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the height of the marker viewport. Must be non-negative. The default is 3.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=marker]{marker}}}
}
}

\keyword{internal}
