% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerUnitsAttribute}
\alias{MarkerUnitsAttribute}
\title{markerUnits}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{strokeWidth | userSpaceOnUse}{Defines the coordinate system for attributes 'markerWidth', 'markerHeight' and the contents of the 'marker'. If markerUnits="strokeWidth", 'markerWidth', 'markerHeight' and the contents of the 'marker' represent values in a coordinate system which has a single unit equal the size in user units of the current stroke width (see the 'stroke.width' property) in place for the graphic object referencing the marker. If markerUnits="userSpaceOnUse", 'markerWidth', 'markerHeight' and the contents of the 'marker' represent values in the current user coordinate system in place for the graphic object referencing the marker (i.e., the user coordinate system for the element referencing the 'marker' element via a 'marker', 'marker.start', 'marker.mid' or 'marker.end' property). If attribute 'markerUnits' is not specified, then the effect is as if a value of 'strokeWidth' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=marker]{marker}}}
}
}
\keyword{internal}

