% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerWidthAttribute}
\alias{MarkerWidthAttribute}
\title{markerWidth}
\description{
Sets the width of the viewport into which the marker is to be fitted
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the width of the marker. Default=3.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=marker]{marker}}}
}
}

\keyword{internal}
