% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{mask}
\alias{mask}
\title{An element used to do masking}
\description{
In SVG, you can specify that any other graphics object or <g> element can be used as an alpha mask for compositing the current object into the background. A mask is defined with the mask element. A mask is used/referenced using the mask property.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Complementary Attributes}}{\code{\link[=MaskElementWHAttribute]{wh}}, \code{\link[=MaskElementXYAttribute]{xy}}}
\item{\emph{Conditional Processing Attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Filter Primitive Attributes}}{\code{\link[=MaskElementHeightAttribute]{height}}, \code{\link[=MaskElementWidthAttribute]{width}}, \code{\link[=MaskElementXAttribute]{x}}, \code{\link[=MaskElementYAttribute]{y}}}
\item{\emph{Presentation Attributes}}{\code{\link[=clip-path-presentationAttribute]{clip.path}}, \code{\link[=color-interpolation-presentationAttribute]{color.interpolation}}, \code{\link[=color-rendering-presentationAttribute]{color.rendering}}, \code{\link[=cursor-presentationAttribute]{cursor}}, \code{\link[=enable-background-presentationAttribute]{enable.background}}, \code{\link[=mask-presentationAttribute]{mask}}}
\item{\emph{Unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=MaskElementMaskContentUnitsAttribute]{maskContentUnits}}, \code{\link[=MaskElementMaskUnitsAttribute]{maskUnits}}, \code{\link[=StyleAttribute]{style}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}, \code{\link[=radialGradient]{radialGradient}}}
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}, \code{\link[=use]{use}}}
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}, \code{\link[=marker]{marker}}, \code{\link[=mask]{mask}}, \code{\link[=pattern]{pattern}}, \code{\link[=switch]{switch}}}
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Container Elements}}{\code{\link[=defs]{defs}}, \code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=altGlyphDef]{altGlyphDef}}, \code{\link[=clipPath]{clipPath}}, \code{\link[=color-profile]{color.profile}}, \code{\link[=cursor]{cursor}}, \code{\link[=filter]{filter}}, \code{\link[=font]{font}}, \code{\link[=font-face]{font.face}}, \code{\link[=foreignObject]{foreignObject}}, \code{\link[=script]{script}}, \code{\link[=style]{style}}, \code{\link[=view]{view}}}
}
}

\keyword{element}
