% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MaskElementMaskContentUnitsAttribute}
\alias{MaskElementMaskContentUnitsAttribute}
\title{maskContentUnits}
\description{
Governs whether to interpret the units for coordinates of the contents of mask either as  as fractions of the bounding box (objectBoundingBox) or using the units inherited from current user coordinate system ( userSpaceOnUse).
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that the coordinate system for contents is given terms of the user coordinate system in effect at the time of \strong{mask} referencing.}
\item{\emph{'objectBoundingBox'}}{Specifies that the coordinate system for contents is given terms of the bounding box of the object to which the \strong{mask} is applied.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=mask]{mask}}}
}
}

\keyword{internal}
