% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MaskElementMaskUnitsAttribute}
\alias{MaskElementMaskUnitsAttribute}
\title{maskUnits}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{userSpaceOnUse | objectBoundingBox}{Defines the coordinate system for attributes 'x', 'y', 'width' and 'height'. If maskUnits="userSpaceOnUse", 'x', 'y', 'width' and 'height' represent values in the current user coordinate system in place at the time when the 'mask' element is referenced (i.e., the user coordinate system for the element referencing the 'mask' element via the 'mask' property). If maskUnits="objectBoundingBox", 'x', 'y', 'width' and 'height' represent fractions or percentages of the bounding box of the element to which the mask is applied. (See Object bounding box units.) If attribute 'maskUnits' is not specified, then the effect is as if a value of 'objectBoundingBox' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=mask]{mask}}}
}
}
\keyword{internal}

