% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{MaskElementWHAttribute}
\alias{MaskElementWHAttribute}
\title{wh}
\description{
Shortcut for assigning height and width
}
\section{Combines}{

\code{\link[=MaskElementWHAttribute]{height}}, \code{\link[=MaskElementWHAttribute]{width}}
}

\section{Equivalence}{

\describe{
wh =v is equivalent to  height=v[1], width=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=mask]{mask}}}
}
}

\keyword{internal}
