% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MaxAttribute}
\alias{MaxAttribute}
\title{max}
\description{
Governs the maximum length of time that the animation remains active..
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<clock.value>}{Specifies the length of the maximum value of the active duration}
\item{\emph{'media'}}{Specifies the maximum value of the active duration as the intrinsic media duration.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
