% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{metadata}
\alias{metadata}
\title{An element that contains metadata}
\description{
Metadata is structured data about data. Metadata which is included with SVG content should be specified within metadata elements. The contents of the metadata should be elements from other XML namespaces such as RDF, FOAF, etc.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
{No Content Elements Applicable}{!}
}
}

\keyword{element}
