% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MPathElementHrefAttribute}
\alias{MPathElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to the path element that is used the path motion for animateMotion.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies an \emph{IRI} reference to the path element to be used to define the path motion.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=mpath]{mpath}}}
}
}

\keyword{internal}
