% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnEndEventAttribute}
\alias{OnEndEventAttribute}
\title{onend}
\description{
Sets the script to execute when an end event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when a \emph{endEvent} event is fired. An \emph{endEvent} occurs at the end of an animation.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
