% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnErrorEventAttribute}
\alias{OnErrorEventAttribute}
\title{onerror}
\description{
Sets the script to execute when an error event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when an \emph{SVGError} event is fired. \emph{SvgErrpr} occurs when either a script error occurs or an element fails to load properly.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
