% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnResizeEventAttribute}
\alias{OnResizeEventAttribute}
\title{onresize}
\description{
Sets the script to execute when an resize event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when an \emph{SVGResize} event is fired. \emph{SvgResize} event occurs after a resizing and applies only to the outermost svg element.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
