% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnScrollEventAttribute}
\alias{OnScrollEventAttribute}
\title{onscroll}
\description{
Sets the script to execute when an scroll event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when an \emph{SVGScroll} event is fired. \emph{SVGScroll} event occurs whenever the view is shifted. This occurs in one of two ways: direct user intervention or a change on the \emph{currentTranslate} property. The \emph{SVGScroll} event applies only to the \emph{outermost svg element}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
