% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnUnloadEventAttribute}
\alias{OnUnloadEventAttribute}
\title{onunload}
\description{
Sets the script to execute when an unload event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when an \emph{SVGUnload} event is fired. \emph{SvgUnload} occurs when the svg document is removed from the window or frame.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
