% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnZoomEventAttribute}
\alias{OnZoomEventAttribute}
\title{onzoom}
\description{
Sets the script to execute when an zoom event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when an \emph{SVGZoom} event is fired. The \emph{SVGZoom} event occurs if zoom level of the document changes. This occurs for one of two reasons: the direct interaction by the user orchange to the \emph{currentScale Property. The }SVGZoom* event applies only to the outermost svg element.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
