% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OriginAttribute}
\alias{OriginAttribute}
\title{origin}
\description{
Sets the origin, but has no effect in SVG, but howerver may be applicable when using CSS for positioning.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<default-value>}{Specifies the origin as defined in the \emph{SMIL} specification.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateMotion]{animateMotion}}}
}
}

\keyword{internal}
