% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PathLengthAttribute}
\alias{PathLengthAttribute}
\title{pathLength}
\description{
Sets the desired length of the path in user units.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies a path length, in user units, given to the user agent whose goal is to attempt to match that value by using an appropriate amount of scaling. This can affect the rendering of text along that path, or path stroke operations and possibly even animation along the path.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=path]{path}}}
}
}

\keyword{internal}
