% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementHeightAttribute}
\alias{PatternElementHeightAttribute}
\title{height}
\description{
Sets the height of the pattern element
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies a tile \emph{height.} (Must be non-negative).}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
