% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementPatternContentUnitsAttribute}
\alias{PatternElementPatternContentUnitsAttribute}
\title{patternContentUnits}
\description{
Governs whether to interpret the coordinate of the contents as fractions of the of the pattern bounding box (objectBoundingBox), or to simply inherit (userSpaceOnUse).
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that the coordinate system for the contents of the \strong{pattern element} is the user coordinate system at the time of the \strong{pattern element} referencing. Thus the measurement units \emph{patternContentUnits} coincides with the \emph{userSpaceUnits}.}
\item{\emph{'objectBoundingBox'}}{Specifies that the coordinate system for contents of the \strong{pattern element} is given terms of the bounding box of the object to which the pattern \strong{pattern element} is applied.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
