% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementPatternUnitsAttribute}
\alias{PatternElementPatternUnitsAttribute}
\title{patternUnits}
\description{
Governs whether the pattern coordinate for x, y, width, and height are to be interpreted as fractions of the bounding box or to be inherited from the user space. objectBoundingBox or userSpaceOnUse.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that the units for \strong{pattern element} attribute values for \emph{'x'}, \emph{'y'}, \emph{'width'} and \emph{'height'} are derived from user coordinate system effective at the time of the \strong{pattern element} referencing.}
\item{\emph{'objectBoundingBox'}}{Specifies that the units for \strong{pattern element} attribute values for \emph{'x'}, \emph{'y'}, \emph{'width'} and \emph{'height'} are given in terms of fractions or percentages of the bounding box of the object to which the pattern is being applied. (Default)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
