% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementWidthAttribute}
\alias{PatternElementWidthAttribute}
\title{width}
\description{
Sets the width of the tile of a pattern element
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the title \emph{width.} (Must be non-negative)}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
