% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementXAttribute}
\alias{PatternElementXAttribute}
\title{x}
\description{
Sets the x-coordinate (left) of the tile of the pattern element
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the left-hand side of a tile}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
