% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositePlayBar.R
\name{playBar}
\alias{playBar}
\title{playBar Compound}
\usage{
playBar(wh, UseGraphPaper)
}
\arguments{
\item{wh}{numeric vector for the width-heigth of the
drawing area sans the toggle bar}

\item{UseGraphPaper}{add GraphPaper to background (default=FALSE)}
}
\description{
Internal compound used to control an animation.
}
\details{
recommend placing immediately after the svgR call
}
\keyword{display}
\keyword{helper}

