% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{pointer-events-presentationAttribute}
\alias{pointer-events-presentationAttribute}
\title{pointer.events}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{visiblePainted}{**ToDo!!!** }
\item{visibleFill}{**ToDo!!!** }
\item{visibleStroke}{**ToDo!!!** }
\item{visible}{**ToDo!!!** }
\item{painted}{**ToDo!!!** }
\item{fill}{**ToDo!!!** }
\item{stroke}{**ToDo!!!** }
\item{all}{**ToDo!!!** }
\item{none}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Elements}}{\code{\link[=use]{use}}}
\item{\emph{Unclassfied:}}{\code{\link[=image]{image}}, \code{\link[=text]{text}}}
}
}
\keyword{internal}

