% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PolygonElementPointsAttribute}
\alias{PolygonElementPointsAttribute}
\title{points}
\description{
Sets the points corresponding to the vertices of the polygon.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<coordinates>[2,N]}{Specifies the vertices of the \strong{polygon} (closed \emph{polyline}) by taking the points defined by the columns of this matrix, with adjacent columns corresponding to adjacent vertices. The points are interpreted in the user coordinate system.}
\item{<coordinates>[[N]][2]}{Specifies the vertices of the \strong{polygon} (closed polyine) by taking the points given in by this unnamed list, with adjacent list entries corresponding to adjacent vertices. The points are interpreted in the user coordinate system.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=polygon]{polygon}}}
}
}

\keyword{internal}
