% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RadialGradientElementCXYAttribute}
\alias{RadialGradientElementCXYAttribute}
\title{cxy}
\description{
Shortcut for assigning cx and cy
}
\section{Combines}{

\code{\link[=RadialGradientElementCXYAttribute]{cx}}, \code{\link[=RadialGradientElementCXYAttribute]{cy}}
}

\section{Equivalence}{

\describe{
cxy =v is equivalent to  cx=v[1], cy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}

\keyword{internal}
