% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RadialGradientElementFXAttribute}
\alias{RadialGradientElementFXAttribute}
\title{fx}
\description{
Sets the x-coordinate of the focal point of the radial gradient.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the focal point of the radial gradient. (The focal point being the location of the initial gradient stop.) Default is 0.50  .}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}

\keyword{internal}
