% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RadialGradientElementFXYAttribute}
\alias{RadialGradientElementFXYAttribute}
\title{fxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=RadialGradientElementFXAttribute]{fx}}, \code{\link[=RadialGradientElementFYAttribute]{fy}}
}

\section{Equivalence}{

\describe{
\item{}{fxy=c(value.FX,value.FY)}
\item{and}{fx = value.FX; fy = value.FY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}
\keyword{internal}

