% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RadialGradientElementFYAttribute}
\alias{RadialGradientElementFYAttribute}
\title{fy}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{See 'fx'. If attribute 'fy' is not specified, 'fy' will coincide with the presentational value of 'cy' for the element whether the value for 'cy' was inherited or not. If the element references an element that specifies a value for 'fy', then the value of 'fy' is inherited from the referenced element.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}
\keyword{internal}

