% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RadialGradientElementRAttribute}
\alias{RadialGradientElementRAttribute}
\title{r}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{See 'cx'. A negative value is an error (see Error processing). A value of zero will cause the area to be painted as a single color using the color and opacity of the last gradient stop. If the attribute is not specified, the effect is as if a value of '50' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}
\keyword{internal}

