% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RectElementRXAttribute}
\alias{RectElementRXAttribute}
\title{rx}
\description{
Sets the x-radius of the rounded corner.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies an x-radius \eqn{r_x} radius of an ellipse to be used to round off the corners of the rectangle. Note \eqn{r_x>0}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}

\keyword{internal}
