% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RectElementRXYAttribute}
\alias{RectElementRXYAttribute}
\title{rxy}
\description{
Shortcut for assigning rx and ry
}
\section{Combines}{

\code{\link[=RectElementRXYAttribute]{rx}}, \code{\link[=RectElementRXYAttribute]{ry}}
}

\section{Equivalence}{

\describe{
rxy =v is equivalent to  rx=v[1], ry=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}

\keyword{internal}
