% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RectElementWHAttribute}
\alias{RectElementWHAttribute}
\title{wh}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=RectElementHeightAttribute]{height}}, \code{\link[=RectElementWidthAttribute]{width}}
}

\section{Equivalence}{

\describe{
\item{}{wh=c(value.HEIGHT,value.WIDTH)}
\item{and}{height = value.HEIGHT; width = value.WIDTH}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}
\keyword{internal}

