% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RectElementXAttribute}
\alias{RectElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{The x.axis coordinate of the side of the rectangle which has the smaller x.axis coordinate value in the current user coordinate system. If the attribute is not specified, the effect is as if a value of "0" were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}
\keyword{internal}

