% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RepeatCountAttribute}
\alias{RepeatCountAttribute}
\title{repeatCount}
\description{
Governs the number of iterations of the animation.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<numeric>}{Specifies the total number of cycles through the anaimation.  (can be fraction).}
\item{\emph{'indefinite'}}{Specifies to repeat indefinitely.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
