% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RestartAttribute}
\alias{RestartAttribute}
\title{restart}
\description{
Provides the conditions for a restart of an animation.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'always'}}{Specifies the animation can be restarted at any time. (Default).}
\item{\emph{'whenNotActive'}}{Specifies the animation can only be restarted when it is not active.}
\item{\emph{'never'}}{Specifies the animation cannot be restarted for the duration of the parent container.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
