% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RotateAttribute}
\alias{RotateAttribute}
\title{rotate}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{auto}{Indicates that the object is rotated over time by the angle of the direction (i.e., directional tangent vector) of the motion path.}
\item{auto.reverse}{Indicates that the object is rotated over time by the angle of the direction (i.e., directional tangent vector) of the motion path plus 180 degrees.}
\item{<number>}{Indicates that the target element has a constant rotation transformation applied to it, where the rotation angle is the specified number of degrees.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateMotion]{animateMotion}}}
}
}
\keyword{internal}

