% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{rotateR}
\alias{rotateR}
\title{Helper for tranform}
\arguments{
\item{angle}{rotation angle in radians}

\item{x}{fixed point x coordinate}

\item{y}{fixed point y coordinate}
}
\description{
Used as an value for the transform attribute. Rotates the target by an amount given by angle, where angle is in radians. Center of rotation is x,y
\itemize{
\item{transform=rotate(pi/2,400,100)}{}
}
}
\keyword{helper}
\keyword{transform}

