% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{scale}
\alias{scale}
\title{Helper for tranform}
\arguments{
\item{dx}{change of scale in the x direction}

\item{dy}{change of scale in the y direction}
}
\description{
Used as an value for the transform attribute. Scales the target by an amount given by dx, dy
\itemize{
\item{transform=scale(x,y)}{}
\item{transform=scale(c(x,y))}{}
\item{transform=c(scale(x,y),  translate(sx,sy}{}
}
}
\keyword{helper}
\keyword{transform}

