% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{script}
\alias{script}
\title{Element used to embed javascript}
\description{
Any functions defined within any script element have a global scope across the entire current document.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{transfer function element attributes}}{\code{\link[=ScriptElementTypeAttribute]{type}}}
\item{\emph{unclassified}}{\code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}}
\item{\emph{xlink attributes}}{\code{\link[=XLinkActuateAttribute]{xlink.actuate}}, \code{\link[=XLinkArcRoleAttribute]{xlink.arcrole}}, \code{\link[=ScriptElementHrefAttribute]{xlink.href}}, \code{\link[=XLinkRoleAttribute]{xlink.role}}, \code{\link[=XLinkShowAttribute]{xlink.show}}, \code{\link[=XLinkTitleAttribute]{xlink.title}}, \code{\link[=XLinkTypeAttribute]{xlink.type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{Any elements or character data.}}
}
}
\keyword{element}

