% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ScriptElementTypeAttribute}
\alias{ScriptElementTypeAttribute}
\title{type}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{content.type}{Identifies the scripting language for the given 'script' element. The value content.type specifies a media type, per Multipurpose Internet Mail Extensions (MIME) Part Two [RFC2046]. If a 'type' is not provided, the value of 'contentScriptType' on the 'svg' element shall be used, which in turn defaults to "application/ecmascript" [RFC4329]. If a 'script' element falls outside of the outermost svg element and the 'type' is not provided, the 'type' must default to "application/ecmascript" [RFC4329].Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=script]{script}}}
}
}
\keyword{internal}

