% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SetElementToAttribute}
\alias{SetElementToAttribute}
\title{to}
\description{
Sets the value for the attribute to be assign and  maintained for  the duration of the set element.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<value>}{Specifies a new attribute value to be set to.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=set]{set}}}
}
}

\keyword{internal}
