% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxDisplay.R
\name{shinyShow}
\alias{shinyShow}
\title{Utility to display a svgR file in local browser using shiny}
\usage{
shinyShow(svgMarkup = NULL, fileName = NULL, rCode = NULL)
}
\arguments{
\item{svgMarkup}{A character string containing the svg markup to display (default=NULL)}

\item{fileName}{A file name containing the svg markup to display (default=NULL)}

\item{rCode}{svgR instructions to be evaluated (default=NULL)}
}
\description{
Utility to display a svgR file in local browser using shiny
}
\details{
Specify only one of the three named parameters.
}
\examples{
\dontrun{ display(fileName="dog.r") }


}
