% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stop-color-presentationAttribute}
\alias{stop-color-presentationAttribute}
\title{stop.color}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{currentColor}{**ToDo!!!** }
\item{<color> <icccolor>}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=stop]{stop}}}
}
}
\keyword{internal}

