% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{stop}
\alias{stop}
\title{An element used to define the colors and layouts of a gradient}
\description{
The ramp of colors to use on a gradient is defined by the stop elements that are child elements to either the <lineargradient> element or the <radialgradient> element.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{presentation attributes}}{\code{\link[=color-presentationAttribute]{color}}, \code{\link[=stop-color-presentationAttribute]{stop.color}}, \code{\link[=stop-opacity-presentationAttribute]{stop.opacity}}}
\item{\emph{transfer function element attributes}}{\code{\link[=StopElementOffsetAttribute]{offset}}}
\item{\emph{unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=StyleAttribute]{style}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=set]{set}}}
}
}
\keyword{element}

