% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{StopElementOffsetAttribute}
\alias{StopElementOffsetAttribute}
\title{offset}
\description{
Governs the placement of the gradient stop.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric | percentage>}{Specifies a fraction (or percentage) determining location of the \strong{stop} element. For \strong{linearGradient} elements this fraction is given by distance from start to stop over the distance from start to end. Here start is the point c(x1, y1) or xy1 and end is the point c(x2, y2) or xy2. For \strong{radialGradient} elements the fraction is computed using the outermost circle and the focal point.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=stop]{stop}}}
}
}

\keyword{internal}
