% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stroke-dasharray-presentationAttribute}
\alias{stroke-dasharray-presentationAttribute}
\title{stroke.dasharray}
\description{
Defines the pattern of dashes and gaps used to stroke paths.
}
\section{Available Attribute Values}{



\describe{
\item{<dasharray>}{Sets the stoke-dasharray pattern using a vector of lengths (with optional unit identifier)}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the \emph{stroke-dasharray} value from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'none'}'}'}'}}{Specifies no dash array.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
