% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stroke-dashoffset-presentationAttribute}
\alias{stroke-dashoffset-presentationAttribute}
\title{stroke.dashoffset}
\description{
Defines the distance into the dash pattern to start the dash.
}
\section{Available Attribute Values}{



\describe{
\item{<length>}{Sets distance from dash pattern to dash start given in non-percent units}
\item{<percentage>}{Sets distance from dash pattern to dash start given as a percentage of current viewport}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the \emph{stroke-dashoffset} value from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
