% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stroke-miterlimit-presentationAttribute}
\alias{stroke-miterlimit-presentationAttribute}
\title{stroke.miterlimit}
\description{
Defines a limit on the ratio of the miter length to the stroke-width.
}
\section{Available Attribute Values}{



\describe{
\item{<miterlimit>}{Sets the miterlimit , a number greater than 1 which is used to limit the ratio of \emph{mitreLength} to \emph{stroke-width}.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{stroke-miterlimit} value from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
