% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stroke-width-presentationAttribute}
\alias{stroke-width-presentationAttribute}
\title{stroke.width}
\description{
Defines the width of the stroke on the current object.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{stroke-width} value from the parent.}
\item{<percentage>}{Sets the stroke-width value as a precentage of the current viewport.}
\item{<length>}{Sets the \emph{stroke-width} value using non-percentate units.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
