% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{style}
\alias{style}
\title{Element used to embed a style specification}
\description{
The style element allows style sheets to be embedded directly within SVG content. SVG's style element has the same attributes as the corresponding element in HTML (see HTML's <style> element),.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{descriptive elements}}{\code{\link[=StyleElementTitleAttribute]{title}}}
\item{\emph{transfer function element attributes}}{\code{\link[=StyleElementTypeAttribute]{type}}}
\item{\emph{unclassified}}{\code{\link[=StyleElementMediaAttribute]{media}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{Any elements or character data.}}
}
}
\keyword{element}

