% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{StyleElementMediaAttribute}
\alias{StyleElementMediaAttribute}
\title{media}
\description{
Sets a set of one or more media descriptors used to indicate intended destination medium for the style information.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<media-descriptors>[N]}{Specifies the intended destination medium for style information Default is \emph{all} .}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=style]{style}}}
}
}

\keyword{internal}
