% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{StyleElementTitleAttribute}
\alias{StyleElementTitleAttribute}
\title{title}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{advisory.title}{(For compatibility with HTML 4 [HTML4].) This attribute specifies an advisory title for the 'style' element.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=style]{style}}}
}
}
\keyword{internal}

