% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgRGetNode.R
\name{[[.svgDoc}
\alias{[[.svgDoc}
\title{Edit the document at the node given by the index}
\usage{
\method{[[}{svgDoc}(doc, id = "")
}
\arguments{
\item{doc}{An svg document created by svgDoc.new}

\item{id}{The id of the node to access. The root node is always
named 'root'}
}
\description{
Edit the document at the node given by the index
}

