% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SVGElementBaseProfileAttribute}
\alias{SVGElementBaseProfileAttribute}
\title{baseProfile}
\description{
Sets a minimal SVG language profile which considered to be necessary for proper rendering.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<profile-name>}{Specifies a minimum SVG language profile considered necessary for correct content rendering.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
