% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{SVGElementWHAttribute}
\alias{SVGElementWHAttribute}
\title{wh}
\description{
Shortcut for assigning height and width
}
\section{Combines}{

\code{\link[=SVGElementWHAttribute]{height}}, \code{\link[=SVGElementWHAttribute]{width}}
}

\section{Equivalence}{

\describe{
wh =v is equivalent to  height=v[1], width=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
