% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SVGElementWidthAttribute}
\alias{SVGElementWidthAttribute}
\title{width}
\description{
Sets the width of the svg.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the \emph{width} of the \strong{svg}. This applies to both for an embedded \strong{svg} and an outer-most \strong{svg} (i.e. \strong{svgR}).}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
