% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SVGElementXAttribute}
\alias{SVGElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{(Has no meaning or effect on outermost svg elements.) The x.axis coordinate of one corner of the rectangular region into which an embedded 'svg' element is placed. If the attribute is not specified, the effect is as if a value of '0' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}}
}
}
\keyword{internal}

