% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{SVGElementXYAttribute}
\alias{SVGElementXYAttribute}
\title{xy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=SVGElementXAttribute]{x}}, \code{\link[=SVGElementYAttribute]{y}}
}

\section{Equivalence}{

\describe{
\item{}{xy=c(value.X,value.Y)}
\item{and}{x = value.X; y = value.Y}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}}
}
}
\keyword{internal}

