% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgRDoc.R
\name{svgR}
\alias{svgR}
\title{Creates a new svg Document}
\usage{
svgR(...)
}
\arguments{
\item{...}{Additional paramaeters}
}
\description{
Creates a new svg Document
}
\note{
By default, name space definition are given by
"http://www.w3.org/2000/svg", "http://www.w3.org/1999/xlink",
"http://www.w3.org/2001/xml-events", but can be overriddent by
specifying namespaceDefinitions=c(...) as an additional parameter.
}
