% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{text-decoration-presentationAttribute}
\alias{text-decoration-presentationAttribute}
\title{text.decoration}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{none | [ underline || overline || line-through || blink ] | inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

