% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{text-rendering-presentationAttribute}
\alias{text-rendering-presentationAttribute}
\title{text.rendering}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{auto}{**ToDo!!!** }
\item{optimizeSpeed}{**ToDo!!!** }
\item{optimizeLegibility}{**ToDo!!!** }
\item{geometricPrecision}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

