% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{text}
\alias{text}
\title{An element used to create text}
\description{
The text element defines a graphics element consisting of text. Note that it is possible to apply a gradient, pattern, clipping path, mask or filter to text
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{animation event attributes}}{\code{\link[=OnLoadEventAttribute]{onload}}}
\item{\emph{combining attributes}}{\code{\link[=TextElementDXYAttribute]{dxy}}, \code{\link[=TextElementXYAttribute]{xy}}}
\item{\emph{conditional processing attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{filter primitive attributes}}{\code{\link[=TextElementXAttribute]{x}}, \code{\link[=TextElementYAttribute]{y}}}
\item{\emph{graphical event attributes}}{\code{\link[=OnActivateEventAttribute]{onactivate}}, \code{\link[=OnClickEventAttribute]{onclick}}, \code{\link[=OnFocusInEventAttribute]{onfocusin}}, \code{\link[=OnFocusOutEventAttribute]{onfocusout}}, \code{\link[=OnLoadEventAttribute]{onload}}, \code{\link[=OnMouseDownEventAttribute]{onmousedown}}, \code{\link[=OnMouseMoveEventAttribute]{onmousemove}}, \code{\link[=OnMouseOutEventAttribute]{onmouseout}}, \code{\link[=OnMouseOverEventAttribute]{onmouseover}}, \code{\link[=OnMouseUpEventAttribute]{onmouseup}}}
\item{\emph{presentation attributes}}{\code{\link[=clip-path-presentationAttribute]{clip.path}}, \code{\link[=clip-rule-presentationAttribute]{clip.rule}}, \code{\link[=color-interpolation-presentationAttribute]{color.interpolation}}, \code{\link[=color-rendering-presentationAttribute]{color.rendering}}, \code{\link[=cursor-presentationAttribute]{cursor}}, \code{\link[=direction-presentationAttribute]{direction}}, \code{\link[=display-presentationAttribute]{display}}, \code{\link[=display-presentationAttribute]{display}}, \code{\link[=dominant-baseline-presentationAttribute]{dominant.baseline}}, \code{\link[=fill-presentationAttribute]{fill}}, \code{\link[=fill-opacity-presentationAttribute]{fill.opacity}}, \code{\link[=fill-rule-presentationAttribute]{fill.rule}}, \code{\link[=filter-presentationAttribute]{filter}}, \code{\link[=font-presentationAttribute]{font}}, \code{\link[=font-family-presentationAttribute]{font.family}}, \code{\link[=font-size-presentationAttribute]{font.size}}, \code{\link[=font-size-adjust-presentationAttribute]{font.size.adjust}}, \code{\link[=font-stretch-presentationAttribute]{font.stretch}}, \code{\link[=font-style-presentationAttribute]{font.style}}, \code{\link[=font-variant-presentationAttribute]{font.variant}}, \code{\link[=font-weight-presentationAttribute]{font.weight}}, \code{\link[=glyph-orientation-horizontal-presentationAttribute]{glyph.orientation.horizontal}}, \code{\link[=glyph-orientation-vertical-presentationAttribute]{glyph.orientation.vertical}}, \code{\link[=kerning-presentationAttribute]{kerning}}, \code{\link[=letter-spacing-presentationAttribute]{letter.spacing}}, \code{\link[=mask-presentationAttribute]{mask}}, \code{\link[=opacity-presentationAttribute]{opacity}}, \code{\link[=pointer-events-presentationAttribute]{pointer.events}}, \code{\link[=stroke-presentationAttribute]{stroke}}, \code{\link[=stroke-dasharray-presentationAttribute]{stroke.dasharray}}, \code{\link[=stroke-dashoffset-presentationAttribute]{stroke.dashoffset}}, \code{\link[=stroke-linecap-presentationAttribute]{stroke.linecap}}, \code{\link[=stroke-linejoin-presentationAttribute]{stroke.linejoin}}, \code{\link[=stroke-miterlimit-presentationAttribute]{stroke.miterlimit}}, \code{\link[=stroke-opacity-presentationAttribute]{stroke.opacity}}, \code{\link[=stroke-width-presentationAttribute]{stroke.width}}, \code{\link[=text-anchor-presentationAttribute]{text.anchor}}, \code{\link[=text-decoration-presentationAttribute]{text.decoration}}, \code{\link[=text-rendering-presentationAttribute]{text.rendering}}, \code{\link[=unicode-bidi-presentationAttribute]{unicode.bidi}}, \code{\link[=visibility-presentationAttribute]{visibility}}, \code{\link[=visibility-presentationAttribute]{visibility}}, \code{\link[=word-spacing-presentationAttribute]{word.spacing}}, \code{\link[=writing-mode-presentationAttribute]{writing.mode}}}
\item{\emph{unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=TextElementDXAttribute]{dx}}, \code{\link[=TextElementDYAttribute]{dy}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=TextElementLengthAdjustAttribute]{lengthAdjust}}, \code{\link[=TextElementRotateAttribute]{rotate}}, \code{\link[=StyleAttribute]{style}}, \code{\link[=TextElementTextLengthAttribute]{textLength}}, \code{\link[=TransformAttribute]{transform}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}}
}
}
\keyword{element}

