% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{TextElementDXYAttribute}
\alias{TextElementDXYAttribute}
\title{dxy}
\description{
Shortcut for assigning dx and dy
}
\section{Combines}{

\code{\link[=TextElementDXYAttribute]{dx}}, \code{\link[=TextElementDXYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
dxy =v is equivalent to  dx=v[1], dy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
