% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextElementDYAttribute}
\alias{TextElementDYAttribute}
\title{dy}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list.of.lengths>}{Shifts in the current text position along the y.axis for the characters within this element or any of its descendants. Refer to the description of the 'dy' attribute on the 'tspan' element. If the attribute is not specified on this element or any of its descendants, no supplemental shifts along the y.axis will occur.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

