% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextElementLengthAdjustAttribute}
\alias{TextElementLengthAdjustAttribute}
\title{lengthAdjust}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{spacing|spacingAndGlyphs}{Indicates the type of adjustments which the user agent shall make to make the rendered length of the text match the value specified on the 'textLength' attribute.'spacing' indicates that only the advance values are adjusted. The glyphs themselves are not stretched or compressed.'spacingAndGlyphs' indicates that the advance values are adjusted and the glyphs themselves stretched or compressed in one axis (i.e., a direction parallel to the inline.progression.direction). The user agent is required to achieve correct start and end positions for the text strings, but the locations of intermediate glyphs are not predictable because user agents might employ advanced algorithms to stretch or compress text strings in order to balance correct start and end positioning with optimal typography. Note that, for a text string that contains n characters, the adjustments to the advance values often occur only for n1 characters (see description of attribute 'textLength'), whereas stretching or compressing of the glyphs will be applied to all n characters. If the attribute is not specified, the effect is as a value of 'spacing' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

