% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextElementRotateAttribute}
\alias{TextElementRotateAttribute}
\title{rotate}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list.of.numbers>}{The supplemental rotation about the current text position that will be applied to all of the glyphs corresponding to each character within this element. Refer to the description of the 'rotate' attribute on the 'tspan' element. If the attribute is not specified on this element or any of its descendants, no supplemental rotations will occur.Animatable. yes (non.additive).}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

