% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextElementTextLengthAttribute}
\alias{TextElementTextLengthAttribute}
\title{textLength}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{The author's computation of the total sum of all of the advance values that correspond to character data within this element, including the advance value on the glyph (horizontal or vertical), the effect of properties 'kerning', 'letter.spacing' and 'word.spacing' and adjustments due to attributes 'dx' and 'dy' on 'tspan' elements. This value is used to calibrate the user agent's own calculations with that of the author. The purpose of this attribute is to allow the author to achieve exact alignment, in visual rendering order after any bidirectional reordering, for the first and last rendered glyphs that correspond to this element; thus, for the last rendered character (in visual rendering order after any bidirectional reordering), any supplemental inter.character spacing beyond normal glyph advances are ignored (in most cases) when the user agent determines the appropriate amount to expand/compress the text string to fit within a length of 'textLength'. A negative value is an error (see Error processing). If the attribute is not specified, the effect is as if the author's computation exactly matched the value calculated by the user agent; thus, no advance adjustments are made.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

