% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextPathElementSpacingAttribute}
\alias{TextPathElementSpacingAttribute}
\title{spacing}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{auto | exact}{Indicates how the user agent should determine the spacing between glyphs that are to be rendered along a path. A value of exact indicates that the glyphs should be rendered exactly according to the spacing rules as specified in Text on a path layout rules. A value of auto indicates that the user agent should use text.on.a.path layout algorithms to adjust the spacing between glyphs in order to achieve visually appealing results. If the attribute is not specified, the effect is as if a value of exact were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=textPath]{textPath}}}
}
}
\keyword{internal}

