% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextPathElementStartOffsetAttribute}
\alias{TextPathElementStartOffsetAttribute}
\title{startOffset}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{An offset from the start of the 'path' for the initial current text position, calculated using the user agent's distance along the path algorithm. If a <length> other than a percentage is given, then the 'startOffset' represents a distance along the path measured in the current user coordinate system. If a percentage is given, then the 'startOffset' represents a percentage distance along the entire path. Thus, startOffset="0" indicates the start point of the 'path' and startOffset="100" indicates the end point of the 'path'. If the attribute is not specified, the effect is as if a value of "0" were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=textPath]{textPath}}}
}
}
\keyword{internal}

