% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ToAttribute}
\alias{ToAttribute}
\title{to}
\description{
Sets the value to be attained by the end of the animation.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<value>}{Specifies the ending attribute value of the animation.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
