% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgRCompound.R
\name{toCompound}
\alias{toCompound}
\title{Wraps an R function to be used as a compounds}
\usage{
toCompound(f, envir = parent.frame())
}
\arguments{
\item{f}{an R function}

\item{envir}{environment for this function (default is parent.frame)}
}
\value{
A function that evaluates the f in the svg environment.
}
\description{
Wraps an R function to be used as a compounds
}

