% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeToggleBar.R
\name{toggleBar}
\alias{toggleBar}
\title{toggleBar Compound}
\usage{
toggleBar(wh)
}
\arguments{
\item{wh}{numeric vector for the width-heigth of the
drawing area sans the toggle bar}
}
\description{
Internal compound used to toggle animations .
}
\keyword{display}
\keyword{helper}

