% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{translate}
\alias{translate}
\title{Helper for tranform}
\arguments{
\item{dx}{Change in x coordinate for translation}

\item{dy}{Change in y coordinate for translation}
}
\description{
Used as an value for the transform attribute. Translates the target by an amount given by dx, dy
\itemize{
\item{transform=translate(x,y)}{}
\item{transform=translate(c(x,y))}{}
\item{transform=c(translate(x,y), rotate(a,cx,cy), scale(sx,sy}{}
}
}
\keyword{helper}
\keyword{transform}

