% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{tspan}
\alias{tspan}
\title{Element used to control sections of text data within a text element}
\description{
Within a <text> element, text and font properties and the current text position can be adjusted with absolute or relative coordinate values by including a tspan element.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{animation event attributes}}{\code{\link[=OnLoadEventAttribute]{onload}}}
\item{\emph{combining attributes}}{\code{\link[=TSpanElementDXYAttribute]{dxy}}, \code{\link[=TSpanElementXYAttribute]{xy}}}
\item{\emph{conditional processing attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{filter primitive attributes}}{\code{\link[=TSpanElementXAttribute]{x}}, \code{\link[=TSpanElementYAttribute]{y}}}
\item{\emph{graphical event attributes}}{\code{\link[=OnActivateEventAttribute]{onactivate}}, \code{\link[=OnClickEventAttribute]{onclick}}, \code{\link[=OnFocusInEventAttribute]{onfocusin}}, \code{\link[=OnFocusOutEventAttribute]{onfocusout}}, \code{\link[=OnLoadEventAttribute]{onload}}, \code{\link[=OnMouseDownEventAttribute]{onmousedown}}, \code{\link[=OnMouseMoveEventAttribute]{onmousemove}}, \code{\link[=OnMouseOutEventAttribute]{onmouseout}}, \code{\link[=OnMouseOverEventAttribute]{onmouseover}}, \code{\link[=OnMouseUpEventAttribute]{onmouseup}}}
\item{\emph{presentation attributes}}{\code{\link[=alignment-baseline-presentationAttribute]{alignment.baseline}}, \code{\link[=baseline-shift-presentationAttribute]{baseline.shift}}, \code{\link[=direction-presentationAttribute]{direction}}, \code{\link[=display-presentationAttribute]{display}}, \code{\link[=dominant-baseline-presentationAttribute]{dominant.baseline}}, \code{\link[=fill-presentationAttribute]{fill}}, \code{\link[=fill-opacity-presentationAttribute]{fill.opacity}}, \code{\link[=fill-rule-presentationAttribute]{fill.rule}}, \code{\link[=font-presentationAttribute]{font}}, \code{\link[=font-family-presentationAttribute]{font.family}}, \code{\link[=font-size-presentationAttribute]{font.size}}, \code{\link[=font-size-adjust-presentationAttribute]{font.size.adjust}}, \code{\link[=font-stretch-presentationAttribute]{font.stretch}}, \code{\link[=font-style-presentationAttribute]{font.style}}, \code{\link[=font-variant-presentationAttribute]{font.variant}}, \code{\link[=font-weight-presentationAttribute]{font.weight}}, \code{\link[=glyph-orientation-horizontal-presentationAttribute]{glyph.orientation.horizontal}}, \code{\link[=glyph-orientation-vertical-presentationAttribute]{glyph.orientation.vertical}}, \code{\link[=kerning-presentationAttribute]{kerning}}, \code{\link[=letter-spacing-presentationAttribute]{letter.spacing}}, \code{\link[=stroke-presentationAttribute]{stroke}}, \code{\link[=stroke-dasharray-presentationAttribute]{stroke.dasharray}}, \code{\link[=stroke-dashoffset-presentationAttribute]{stroke.dashoffset}}, \code{\link[=stroke-linecap-presentationAttribute]{stroke.linecap}}, \code{\link[=stroke-linejoin-presentationAttribute]{stroke.linejoin}}, \code{\link[=stroke-miterlimit-presentationAttribute]{stroke.miterlimit}}, \code{\link[=stroke-opacity-presentationAttribute]{stroke.opacity}}, \code{\link[=stroke-width-presentationAttribute]{stroke.width}}, \code{\link[=text-anchor-presentationAttribute]{text.anchor}}, \code{\link[=text-decoration-presentationAttribute]{text.decoration}}, \code{\link[=unicode-bidi-presentationAttribute]{unicode.bidi}}, \code{\link[=visibility-presentationAttribute]{visibility}}, \code{\link[=word-spacing-presentationAttribute]{word.spacing}}}
\item{\emph{unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=TSpanElementDXAttribute]{dx}}, \code{\link[=TSpanElementDYAttribute]{dy}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=TextElementLengthAdjustAttribute]{lengthAdjust}}, \code{\link[=TSpanElementRotateAttribute]{rotate}}, \code{\link[=StyleAttribute]{style}}, \code{\link[=TSpanElementTextLengthAttribute]{textLength}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=set]{set}}}
\item{\emph{Descriptive Elements}}{\code{\link[=desc]{desc}}, \code{\link[=metadata]{metadata}}, \code{\link[=title]{title}}}
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}}
}
}
\keyword{element}

