% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TSpanElementDXAttribute}
\alias{TSpanElementDXAttribute}
\title{dx}
\description{
Set shifts in the current x-position prior to rendering.
}
\section{Available Attribute Values}{

The value  represents one or more lengths, given as a vector.
\describe{
\item{<length>[N]}{Specifies a sequence of x-shifts starting with the current text position for the characters within this \strong{tspan} or any of its descendants.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
