% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{TSpanElementDXYAttribute}
\alias{TSpanElementDXYAttribute}
\title{dxy}
\description{
Shortcut for assigning dx and dy
}
\section{Combines}{

\code{\link[=TSpanElementDXYAttribute]{dx}}, \code{\link[=TSpanElementDXYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
dxy =v is equivalent to  dx=v[1], dy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
