% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{TSpanElementXYAttribute}
\alias{TSpanElementXYAttribute}
\title{xy}
\description{
Shortcut for assigning x and y
}
\section{Combines}{

\code{\link[=TSpanElementXYAttribute]{x}}, \code{\link[=TSpanElementXYAttribute]{y}}
}

\section{Equivalence}{

\describe{
xy =v is equivalent to  x=v[1], y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
