% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.cm}
\alias{u.cm}
\title{Helper function to declare units}
\arguments{
\item{x}{value in centimeters (cm)}
}
\description{
Declares the units to be centimeters (cm)
\itemize{
\item{transform=translate(u.cm(0,2)}{}
}
}
