% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.em}
\alias{u.em}
\title{Helper function to declare units}
\arguments{
\item{x}{value in em}
}
\description{
Declares the units to be em's. (An em isa  measure relative to current font size)
\itemize{
\item{transform=translate(u.em(c(0,2))}{}
}
}
