% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.ex}
\alias{u.ex}
\title{Helper function to declare units}
\arguments{
\item{x}{value in ex}
}
\description{
Declares the units to be ex's. (An ex is a measure relative to the height of the character x)
\itemize{
\item{transform=translate(u.ex(c(0,2))}{}
}
}
