% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.in}
\alias{u.in}
\title{Helper function to declare units}
\arguments{
\item{x}{value in inches (in)}
}
\description{
Declares the units to be inches (in)
\itemize{
\item{transform=translate(u.in(0,.5)}{}
}
}

