% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.mm}
\alias{u.mm}
\title{Helper function to declare units}
\arguments{
\item{x}{value in millimeters (mm)}
}
\description{
Declares the units to be millimeters (mm)
\itemize{
\item{transform=translate(u.mm(0,20)}{}
}
}
