% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.pc}
\alias{u.pc}
\title{Helper function to declare units}
\arguments{
\item{x}{valus in picas (pc)}
}
\description{
Declares the units to be pica's. (1pc=1/6 inch)
\itemize{
\item{transform=translate(u.pc(0,3))}{}
}
}
