% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.prct}
\alias{u.prct}
\title{Helper function to declare units}
\arguments{
\item{x}{value as a percentage (\%)}
}
\description{
Declares the units to be percentage (\%)
\itemize{
\item{viewBox=u.\%(c(0,0,50,50))}{}
}
}
