% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.pt}
\alias{u.pt}
\title{Helper function to declare units}
\arguments{
\item{x}{value in points (pt)}
}
\description{
Declares the units to be points's. (1pt=1/72 inch)
\itemize{
\item{transform=translate(u.pt(0,36))}{}
}
}
