% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.px}
\alias{u.px}
\title{Helper function to declare units}
\arguments{
\item{x}{value in pixels (default)}
}
\description{
Declares the units to be pixels.
\itemize{
\item{transform=translate(u.px(c(0,2))}{}
}
}
