% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_HelperPages.R
\name{u.rd}
\alias{u.rd}
\title{Helper function to declare units}
\arguments{
\item{x}{value in radians}
}
\description{
Declares the units to be radians
\itemize{
\item{transform=rotate(u.rd(pi/2), 100,100)}{}
}
}

