% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{UseElementHeightAttribute}
\alias{UseElementHeightAttribute}
\title{height}
\description{
Sets the height of the rectangular region into which the referenced element is to be placed.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the \emph{height} of the rectangular region into which the referenced element by use is to be placed.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=use]{use}}}
}
}

\keyword{internal}
