% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{UseElementYAttribute}
\alias{UseElementYAttribute}
\title{y}
\description{
Sets the y-coordinate of the top edge of the region where an element referenced by use is to be placed.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-coordinate of the top side rectangular region into which the referenced element by use is to be placed.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=use]{use}}}
}
}

\keyword{internal}
