% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ValuesAttribute}
\alias{ValuesAttribute}
\title{values}
\description{
Sets thef animation values to be used for the given animation type.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<values>[[N]] (list-of-values)}{Specifies an unamed list attributes values over which to iterate.}
\item{<values>[N,M] (matrix-of-values)}{A matrix with columns corresponding to the attribute values. Obviously assumes that the length of each attribute value is the same. Thisos useful for defining animations involving points.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
