% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ViewBoxAttribute}
\alias{ViewBoxAttribute}
\title{viewBox}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{c(x,y,w,h)}{x, y upper left hand corner; w,h width height}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Unclassfied:}}{\code{\link[=marker]{marker}}, \code{\link[=pattern]{pattern}}, \code{\link[=view]{view}}}
}
}
\keyword{internal}

