% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ViewElementViewTargetAttribute}
\alias{ViewElementViewTargetAttribute}
\title{viewTarget}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{XML_Name [XML_NAME]*}{Indicates the target object associated with the view. Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=view]{view}}}
}
}
\keyword{internal}

