% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{word-spacing-presentationAttribute}
\alias{word-spacing-presentationAttribute}
\title{word.spacing}
\description{
Governs the  spacing behavior between words.
}
\section{Available Attribute Values}{



\describe{
\item{<length>}{Sets a value for the amount of  spacing between words.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{word-spacing} value from the parent.}
\item{\emph{'\emph{'normal'}'}}{Specifies to use the default \emph{word-spacing} value.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
