% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{writing-mode-presentationAttribute}
\alias{writing-mode-presentationAttribute}
\title{writing.mode}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{lr-tb}{**ToDo!!!** }
\item{rl-tb}{**ToDo!!!** }
\item{tb-rl}{**ToDo!!!** }
\item{lr}{**ToDo!!!** }
\item{rl}{**ToDo!!!** }
\item{tb}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

