% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{XLinkShowAttribute}
\alias{XLinkShowAttribute}
\title{xlink.show}
\description{
Sets the display context in which target is intended to be placed, This applies only to Xlink aware processors.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'new'}}{Specifies \emph{new} to Xlink-aware processors. This attribute is provided for backwards compatibility with \emph{SVG 1.1}.}
\item{\emph{'replace'}}{Specifies \emph{replace} to Xlink-aware processors. This attribute is provided for backwards compatibility with \emph{SVG 1.1}.}
\item{\emph{'embed'}}{Specifies \emph{embed} to Xlink-aware processors. This attribute is provided for backwards compatibility with \emph{SVG 1.1}.}
\item{\emph{'other'}}{Specifies \emph{other} to Xlink-aware processors. This attribute is provided for backwards compatibility with \emph{SVG 1.1}.}
\item{\emph{'none'}}{Specifies \emph{none} to Xlink-aware processors. This attribute is provided for backwards compatibility with \emph{SVG 1.1}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
\item{\emph{Filter Primitive Elements}}{\code{\link[=feImage]{feImage}}}
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}, \code{\link[=use]{use}}}
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}, \code{\link[=pattern]{pattern}}}
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=color-profile]{color.profile}}, \code{\link[=cursor]{cursor}}, \code{\link[=filter]{filter}}, \code{\link[=font-face-uri]{font.face.uri}}, \code{\link[=glyphRef]{glyphRef}}, \code{\link[=mpath]{mpath}}, \code{\link[=script]{script}}}
}
}

\keyword{internal}
