% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ZoomAndPanAttribute}
\alias{ZoomAndPanAttribute}
\title{zoomAndPan}
\description{
Governs whether or not the outermost svg can zoom and pan.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'disable'}}{disables zooming}
\item{\emph{'magnify'}}{enables zooming (applies only to the for outermost \emph{svg}, typically arrising from an \emph{svgR} call)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=view]{view}}}
}
}

\keyword{internal}
