% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AdditiveAttribute}
\alias{AdditiveAttribute}
\title{additive}
\description{
Governs whether or not the animation is additive.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'sum'}}{Specifies that the animation is to add to the underlying value of the attribute and any other lower priority animations.}
\item{\emph{'replace'}}{Specifies that the the animation is to replace (override) the underlying value of the attributes. This is the default}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
