% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AElementTargetAttribute}
\alias{AElementTargetAttribute}
\title{target}
\description{
Governs the manner in which the link is opened. That is where is the target placed?
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'_replace'}}{Specifies to replace in the current \emph{SVG image} by the linked content. The replacement will occupy the same region and frame as the current SVG.}
\item{\emph{'_self'}}{Specifies to replace in the current \emph{SVG image} by the linked content The replacement will occupy the frame as the current SVG. This is the default.}
\item{\emph{'_parent'}}{Specifies to replace the parent frameset \emph{SVG image} by the linked content.}
\item{\emph{'_top'}}{Specifies to replace the full window/tab (and frames) byt the linked content.}
\item{\emph{'_blank'}}{Specifies to create a new (unnamed) window/tab display and fill with the linked content.}
\item{<name>}{Specifies to find an existing window/tab/pane having the target name, and if one does not exist create one. Then fill it with the linked content.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}}
}
}

\keyword{internal}
