% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AElementXLinkActuateAttribute}
\alias{AElementXLinkActuateAttribute}
\title{xlink.actuate}
\description{
For XLink-aware processors, governs the conditions of loading the linked resource.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{\emph{'onRequest'}}{Specifies for Xlink-aware processors to perform only on a trigger of post-loading event trigger, a traversal from the starting resource to the ending resource.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}}
}
}

\keyword{internal}
