% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AElementXLinkHrefAttribute}
\alias{AElementXLinkHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference for location of a resource.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies a reference to an object as an \emph{IRI} reference.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=a]{a}}}
}
}

\keyword{internal}
