% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{alignment-baseline-presentationAttribute}
\alias{alignment-baseline-presentationAttribute}
\title{alignment.baseline}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{auto}{**ToDo!!!** }
\item{baseline}{**ToDo!!!** }
\item{before-edge}{**ToDo!!!** }
\item{text-before-edge}{**ToDo!!!** }
\item{middle}{**ToDo!!!** }
\item{central}{**ToDo!!!** }
\item{after-edge}{**ToDo!!!** }
\item{text-after-edge}{**ToDo!!!** }
\item{ideographic}{**ToDo!!!** }
\item{alphabetic}{**ToDo!!!** }
\item{hanging}{**ToDo!!!** }
\item{mathematical}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}
\keyword{internal}

