% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{AltGlyphElementDXYAttribute}
\alias{AltGlyphElementDXYAttribute}
\title{dxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=AltGlyphElementDXAttribute]{dx}}, \code{\link[=AltGlyphElementDYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
\item{}{dxy=c(value.DX,value.DY)}
\item{and}{dx = value.DX; dy = value.DY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}
\keyword{internal}

