% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementFormatAttribute}
\alias{AltGlyphElementFormatAttribute}
\title{format}
\description{
Sets the fonts alt-glyph format.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<glyph-format>}{Specifies a \emph{alt-Glyph} format.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
