% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementGlyphRefAttribute}
\alias{AltGlyphElementGlyphRefAttribute}
\title{glyphRef}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<string>}{The glyph identifier, the format of which is dependent on the 'format' of the given font. (Same meaning as the 'glyphRef' attribute on the 'glyphRef' element.)Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}
\keyword{internal}

