% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementHrefAttribute}
\alias{AltGlyphElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to either a glyph or altGlyphDef.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies an \emph{IRI} reference to either a glyph or an altGlyphDef.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
