% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementXAttribute}
\alias{AltGlyphElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list.of.coordinates>}{The <coordinate> values are processed in the same manner as the 'x' attribute on the 'tspan' element, with the following exception. If the referenced alternate glyphs are rendered instead of the Unicode characters inside the 'altGlyph' element, then any absolute X coordinates specified via an 'x' attribute on this element or any ancestor 'text' or 'tspan' elements for Unicode characters 2 through n within the 'altGlyph' element are ignored. Any absolute X coordinate specified via an 'x' attribute on this element or any ancestor 'text' or 'tspan' elements for the first Unicode character within the 'altGlyph' element sets a new absolute X coordinate for the current text position before rendering the first alternate glyph.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}
\keyword{internal}

