% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AltGlyphElementYAttribute}
\alias{AltGlyphElementYAttribute}
\title{y}
\description{
Sets the y-coordinates for the alternative glyphs..
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{Specifies corresponding absolute y-coordinates for \strong{altGlyph} element. If rendering only unicode characters inside an \strong{altGlyph} element, then this specifies the absolute y-coordinates for the first n characters where n is the length of the y-vector . If rendering alternate \emph{glyphs}, then y[1] specifies a new the absolute y-coordinate for the current text position, so that text begins as y[1], with the remainder, y[-1], being ignored.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}}
}
}

\keyword{internal}
