% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{AnimateTransformElementTypeAttribute}
\alias{AnimateTransformElementTypeAttribute}
\title{type}
\description{
Governs the type of transformation which is to have its values change over time.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'translate'}}{Specifies a translational transform}
\item{\emph{'scale'}}{Specifies a scaling transform}
\item{\emph{'rotate'}}{Specifies a rotational transform}
\item{\emph{'skewX'}}{Specifies a skew transform along the X axis}
\item{\emph{'skewY'}}{Specifies a skew transform along the Y axis}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
