% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{CalcModeAttribute}
\alias{CalcModeAttribute}
\title{calcMode}
\description{
Governs the interpolation model for the animation.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'discrete'}}{Specifies a discrete jump to the next value without any intermediate transitioning.}
\item{\emph{'linear'}}{Specifies an intermediate transitioning based on a linear interpolation across time from the starting value to the ending value. (This is the the default for all except for \emph{animateMotion})}
\item{\emph{'paced'}}{Specifies an even paced intermediate transitioning. This is the default for \emph{animateMotion}.}
\item{\emph{'spline'}}{Specifies an intermediate transitioning based on a cubic Bezier spline with points defined using the \emph{keyTimes} and \emph{keySplines} attributes.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}

\keyword{internal}
