% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{CircleElementCXYAttribute}
\alias{CircleElementCXYAttribute}
\title{cxy}
\description{
Shortcut for assigning cx and cy
}
\section{Combines}{

\code{\link[=CircleElementCXYAttribute]{cx}}, \code{\link[=CircleElementCXYAttribute]{cy}}
}

\section{Equivalence}{

\describe{
cxy =v is equivalent to  cx=v[1], cy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}}
}
}

\keyword{internal}
