% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{CircleElementRAttribute}
\alias{CircleElementRAttribute}
\title{r}
\description{
Sets the radius of a circle.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the radius, \eqn{r} of the \strong{circle}. Note: \eqn{r\geq0}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}}
}
}

\keyword{internal}
