% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{color-profile-presentationAttribute}
\alias{color-profile-presentationAttribute}
\title{color.profile}
\description{
Governs what color profile to use.
}
\section{Available Attribute Values}{



\describe{
\item{<iri>}{Specifies an IRI reference to the color profile to be used}
\item{<name>}{Specifies the color profile by name}
\item{\emph{'\emph{'\emph{'\emph{'auto'}'}'}'}}{Specifies to auto select the color profile}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the color profiel from the parents color profile.}
\item{\emph{'\emph{'\emph{'\emph{'sRGB'}'}'}'}}{Specifies to use the sRGB  color profile}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}}
}
}

\keyword{internal}
