% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ColorProfileElementHrefAttribute}
\alias{ColorProfileElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to an ICC color profile resource.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies a reference to an \emph{ICC profile} resource.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=color-profile]{color.profile}}}
}
}

\keyword{internal}
