% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ColorProfileElementNameAttribute}
\alias{ColorProfileElementNameAttribute}
\title{name}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<name>}{The name which is used as the first parameter for icc.color specifications within 'fill', 'stroke', 'stop.color', 'flood.color' and 'lighting.color' property values to identify the color profile to use for the ICC color specification and the name which can be the value of the 'color.profile' property. Note that if <name> is not provided, it will be impossible to reference the given color profile description. The name "sRGB" is predefined; any color profile descriptions with <name> set to "sRGB" will be ignored. For consistency with CSS lexical scanning and parsing rules ([CSS2], section D.2), the keyword "sRGB" is case.insensitive; however, it is recommended that the mixed capitalization "sRGB" be used for consistency with common industry practice.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=color-profile]{color.profile}}}
}
}
\keyword{internal}

