% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ColorProfileElementRenderingIntentAttribute}
\alias{ColorProfileElementRenderingIntentAttribute}
\title{rendering.intent}
\description{
Governs a color profile rendering intent.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'auto'}}{Directs the user agent to determine intent based on the content type. This is the Default.}
\item{\emph{'perceptual'}}{Specifies to perserve relative \emph{color}, used often for images.}
\item{\emph{'saturation'}}{Specifies to preserve \emph{saturation.}}
\item{\emph{'Relative-colorimetric'}}{Specifies to not change colors that fall within the range supported by the rendering device.}
\item{\emph{'Absolute-colorimetric'}}{Specifies the disable any \emph{white} point matching during color convertion. \emph{Not recommended.}}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=color-profile]{color.profile}}}
}
}

\keyword{internal}
