% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ContentScriptTypeAttribute}
\alias{ContentScriptTypeAttribute}
\title{contentScriptType}
\description{
Sets the default scripting language for the given SVG document fragment.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<content-type>}{Specifies the default scripting language of the event attributes value strings. The default is \emph{'application/ecmascript'}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
