% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ContentStyleTypeAttribute}
\alias{ContentStyleTypeAttribute}
\title{contentStyleType}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{content.type}{Identifies the default style sheet language for the given document. That language must then be used for all instances of style that do not specify their own style sheet language, such as the 'style' attributes that are available on many elements. The value content.type specifies a media type, per MIME Part Two. Media Types [RFC2046]. The default value is "text/css" [RFC2318].Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}}
}
}
\keyword{internal}

