% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{CursorElementHrefAttribute}
\alias{CursorElementHrefAttribute}
\title{xlink.href}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<funciri>}{A Functional IRI reference to the file or element which provides the image of the cursor.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=cursor]{cursor}}}
}
}
\keyword{internal}

