% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{CursorElementXAttribute}
\alias{CursorElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the cursor point, this coordinate being interpreted using the cursors coordinate system.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies x-coordinate of the point that the \strong{cursor} is pointing to. This position is expressed in the \emph{cursor's} coordinate system. The default is 0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=cursor]{cursor}}}
}
}

\keyword{internal}
