% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{CursorElementYAttribute}
\alias{CursorElementYAttribute}
\title{y}
\description{
Sets the y-coordinate of the cursor point, this coordinate being interpreted using the cursors coordinate system.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-coordinate of the point that the cursor is pointing to. This position is expressed in the cursor's coordinate system. The default is 0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=cursor]{cursor}}}
}
}

\keyword{internal}
