% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{DurAttribute}
\alias{DurAttribute}
\title{dur}
\description{
Governs the simple duration.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{<clock-value>}{Specifies the length of the simple duration in presentation time. This value must be greater than 0. (The default unis is in seconds)}
\item{\emph{'media'}}{Specifies the simple duration as the intrinsic media duration.}
\item{\emph{'indefinite'}}{Specifies the simple duration as \emph{indefinite}, that is the animation is to continue indefinitely.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
