% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{EllipseElementRXAttribute}
\alias{EllipseElementRXAttribute}
\title{rx}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{The x.axis radius of the ellipse. A negative value is an error (see Error processing). A value of zero disables rendering of the element.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=ellipse]{ellipse}}}
}
}
\keyword{internal}

