% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{EllipseElementRXYAttribute}
\alias{EllipseElementRXYAttribute}
\title{rxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=EllipseElementRXAttribute]{rx}}, \code{\link[=EllipseElementRYAttribute]{ry}}
}

\section{Equivalence}{

\describe{
\item{}{rxy=c(value.RX,value.RY)}
\item{and}{rx = value.RX; ry = value.RY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=ellipse]{ellipse}}}
}
}
\keyword{internal}

