% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{EndAttribute}
\alias{EndAttribute}
\title{end}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{end.value.list ..= end.value (S? ";" S? end.value.list )?}{A semicolon separated list of end values. The interpretation of a list of end times is detailed below.}
\item{end.value ..= ( offset.value | syncbase.value | event.value | repeat.value | accessKey.value | wallclock.sync.value | "indefinite" )}{Describes the active end of the animation.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}
\keyword{internal}

