% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feBlendModeAttribute}
\alias{feBlendModeAttribute}
\title{mode}
\description{
Governs which algorithm to used for blending of the two image sources.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'normal'}}{Specifies the \emph{normal blending mode} See below.}
\item{\emph{'multiply'}}{Specifies the \emph{multiply blending mode} See below.}
\item{\emph{'screen'}}{Specifies the \emph{screen blending mode} See below.}
\item{\emph{'darken'}}{Specifies the \emph{darken blending mode} See below.}
\item{\emph{'lighten'}}{Specifies the \emph{lighten blending mode} See below.}
}
}

\section{feBlend Mode Details}{

feBlend Modes are defned as follows

\strong{normal mode (no blending)}

\deqn{ \alpha_{out} = 1 - (1- \alpha_{in1})  (1- \alpha_{in2}) }
\deqn{ \alpha_{out}  \times C_{out} =(1-\alpha_{in2})  \alpha_{in1} C_{in1} + \alpha_{in1}  C_{in1} }

\strong{multiply blend mode}

\deqn{ \alpha_{out} = 1 - (1- \alpha_{in1}) \times (1- \alpha_{in2)} }
\deqn{\alpha_{out}  \times C_{out} = (1-\alpha_{in2})  \alpha_{in1} C_{in1} + (1-\alpha_{in1})  \alpha_{in2} C_{in2} + \alpha_{in1} \alpha_{in2}  C_{in1}   C_{in2} }

\strong{screen blend mode}

\deqn{ \alpha_{out} = 1 - (1- \alpha_{in1}) \times (1- \alpha_{in2)} }
\deqn{ \alpha_{out}  \times C_{out} = \alpha_{in1} C_{in1} + \alpha_{in2} C_{in2} + \alpha_{in1} \alpha_{in2}  C_{in1}   C_{in2} }

\strong{darken blend mode}

\deqn{ \alpha_{out} = 1 - (1- \alpha_{in1}) \times (1- \alpha_{in2)} }
\deqn{\alpha_{out}  \times C_{out} = min( (1-\alpha_{in1})  \alpha_{in2} C_{in2} + \alpha_{in1}C_{in1}, (1-\alpha_{in2})  \alpha_{in1} C_{in1} + \alpha_{in2}C_{in2} )}

\strong{lighten blend mode}

\deqn{ \alpha_{out} = 1 - (1- \alpha_{in1}) \times (1- \alpha_{in2)}}
\deqn{ \alpha_{out}  \times C_{out} = max((1-\alpha_{in1}) \times \alpha_{in2} C_{in2} + \alpha_{in1}C_{in1},(1-\alpha_{in2}) \times \alpha_{in1} C_{in1} + \alpha_{in2}C_{in2})}#'
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feBlend]{feBlend}}}
}
}

\keyword{internal}
