% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feColorMatrixTypeAttribute}
\alias{feColorMatrixTypeAttribute}
\title{type}
\description{
Governs the type of matrix operation that the feColorMatrix represents.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'matrix'}}{Specifies to interpret the \strong{feColorMatrix} values parameter as a 4x5 matrix . The result color transformation is given by premultiplying this with c(R, G.B, A, 1) Default.}
\item{\emph{'saturate'}}{Specifies that the values parameter is a single numeric between 0 and 1. used to perform color saturation.}
\item{\emph{'hueRotate'}}{Specifies that the values parameter is a single numeric representing degrees between 0 and 360, and is . used to perform hue rotation.}
\item{\emph{'luminanceToAlpha'}}{Specifies to produce a \emph{luminance} to \emph{alpha} effect, and does not use values}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feColorMatrix]{feColorMatrix}}}
}
}

\keyword{internal}
