% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feColorMatrixValuesAttribute}
\alias{feColorMatrixValuesAttribute}
\title{values}
\description{
Sets the values the feColorMatrix that is to be used in the RGBA color transformation.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Depends on the value for type.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feColorMatrix]{feColorMatrix}}}
}
}

\keyword{internal}
