% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feCompositeK1234Attribute}
\alias{feCompositeK1234Attribute}
\title{k1234}
\description{
Shortcut for assigning k1, k2, k3 and k4
}
\section{Combines}{

\code{\link[=feCompositeK1234Attribute]{k1}}, \code{\link[=feCompositeK1234Attribute]{k2}}, \code{\link[=feCompositeK1234Attribute]{k3}}, \code{\link[=feCompositeK1234Attribute]{k4}}
}

\section{Equivalence}{

\describe{
k1234 =v is equivalent to  k1=v[1], k2=v[2], k3=v[3], k4=v[4]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feComposite]{feComposite}}}
}
}

\keyword{internal}
