% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feCompositeK3Attribute}
\alias{feCompositeK3Attribute}
\title{k3}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{Only applicable if operator="arithmetic". If the attribute is not specified, the effect is as if a value of 0 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feComposite]{feComposite}}}
}
}
\keyword{internal}

