% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{feConvolveMatrix}
\alias{feConvolveMatrix}
\title{A filter primitive element that performs image convolution}
\description{
the feConvolveMatrix element applies a matrix convolution filter effect. A convolution combines pixels in the input image with neighboring pixels to produce a resulting image. A wide variety of imaging operations can be achieved through convolutions, including blurring, edge detection, sharpening, embossing and beveling.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Complementary Attributes}}{\code{\link[=feConvolveMatrixElementTARGETXYAttribute]{targetXY}}, \code{\link[=FilterPrimitiveWHAttribute]{wh}}, \code{\link[=FilterPrimitiveXYAttribute]{xy}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Filter Primitive Attributes}}{\code{\link[=FilterPrimitiveHeightAttribute]{height}}, \code{\link[=FilterPrimitiveResultAttribute]{result}}, \code{\link[=FilterPrimitiveWidthAttribute]{width}}, \code{\link[=FilterPrimitiveXAttribute]{x}}, \code{\link[=FilterPrimitiveYAttribute]{y}}}
\item{\emph{Presentation Attributes}}{\code{\link[=color-interpolation-filters-presentationAttribute]{color.interpolation.filters}}}
\item{\emph{Unclassified}}{\code{\link[=feConvolveMatrixElementBiasAttribute]{bias}}, \code{\link[=ClassAttribute]{class}}, \code{\link[=feConvolveMatrixElementDivisorAttribute]{divisor}}, \code{\link[=feConvolveMatrixElementEdgeModeAttribute]{edgeMode}}, \code{\link[=FilterPrimitiveIn1Attribute]{in1}}, \code{\link[=feConvolveMatrixElementKernelMatrixAttribute]{kernelMatrix}}, \code{\link[=feConvolveMatrixElementKernelUnitLengthAttribute]{kernelUnitLength}}, \code{\link[=feConvolveMatrixElementOrderAttribute]{order}}, \code{\link[=feConvolveMatrixElementPreserveAlphaAttribute]{preserveAlpha}}, \code{\link[=StyleAttribute]{style}}, \code{\link[=feConvolveMatrixElementTargetXAttribute]{targetX}}, \code{\link[=feConvolveMatrixElementTargetYAttribute]{targetY}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=set]{set}}}
}
}

\keyword{element}
