% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementDivisorAttribute}
\alias{feConvolveMatrixElementDivisorAttribute}
\title{divisor}
\description{
Sets a divisor to be  applied  to the results of the convolution. This is generally used  for evening out the color values.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies a divisor to scale down the resulting image component values obtained by multiply by the \emph{kernelMatrix}.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
