% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementKernelMatrixAttribute}
\alias{feConvolveMatrixElementKernelMatrixAttribute}
\title{kernelMatrix}
\description{
Sets the kernelMatrix to be used in the convolution filter.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<numeric>[N]}{Specifies the entries of the \emph{kernelMatrix}.}
\item{matrix[N.M]}{Specifies the  \emph{kernelMatrix} using a matrix.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
