% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementOrderAttribute}
\alias{feConvolveMatrixElementOrderAttribute}
\title{order}
\description{
Sets the number of cells in each dimension of the kernel matrix to be used in a convolution.
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2]}{Specifies the number of columns and rows to use for the \emph{kernelMatrix.} Typically usage is c(3, 3).}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
