% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementPreserveAlphaAttribute}
\alias{feConvolveMatrixElementPreserveAlphaAttribute}
\title{preserveAlpha}
\description{
Governs whether or not  to apply the convolution to the alpha channel.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'FALSE'}}{Specifies to apply the \emph{convolution} to all channels, including the \emph{alpha channel}.}
\item{\emph{'TRUE'}}{Specifies to apply the \emph{convolution} to only the colors channels. (Not the \emph{alpha channel}!)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
