% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feConvolveMatrixElementTARGETXYAttribute}
\alias{feConvolveMatrixElementTARGETXYAttribute}
\title{targetXY}
\description{
Shortcut for assigning targetX and targetY
}
\section{Combines}{

\code{\link[=feConvolveMatrixElementTARGETXYAttribute]{targetX}}, \code{\link[=feConvolveMatrixElementTARGETXYAttribute]{targetY}}
}

\section{Equivalence}{

\describe{
targetXY =v is equivalent to  targetX=v[1], targetY=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
