% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feConvolveMatrixElementTargetYAttribute}
\alias{feConvolveMatrixElementTargetYAttribute}
\title{targetY}
\description{
Sets the y-position of the target pixel relative to the convolution matrix.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<integer>}{Specifies the placement of the \emph{convolution matrix} implied by specifying the target y-coordinate relative to the \emph{convolution} origin. By default, the \emph{convolution matrix} is centered about the target.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feConvolveMatrix]{feConvolveMatrix}}}
}
}

\keyword{internal}
