% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDiffuseLightingDiffuseConstantAttribute}
\alias{feDiffuseLightingDiffuseConstantAttribute}
\title{diffuseConstant}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{kd in Phong lighting model. In SVG, this can be any non.negative number. If the attribute is not specified, then the effect is as if a value of 1 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDiffuseLighting]{feDiffuseLighting}}}
}
}
\keyword{internal}

