% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDisplacementMapScaleAttribute}
\alias{feDisplacementMapScaleAttribute}
\title{scale}
\description{
Sets the displacement scaling factor to be applied.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the displacement scale factor, as given in the by the \strong{feDisplacement} Map equations. Note: a value of 0 results in no displacement.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDisplacementMap]{feDisplacementMap}}}
}
}

\keyword{internal}
