% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDisplacementMapYChannelSelectorAttribute}
\alias{feDisplacementMapYChannelSelectorAttribute}
\title{yChannelSelector}
\description{
Governs which channel (R|G|B|A) of the image given by in2 to be used for the computing the amount of displacement of of the image given by in1 in the y direction.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{\emph{'R'}}{Specifies to use the \emph{R-channel} from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the y-axis.}
\item{\emph{'G'}}{Specifies to use the \emph{G-channel} from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the y-axis.}
\item{\emph{'B'}}{Specifies to use the \emph{B-channel} from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the y-axis.}
\item{\emph{'A'}}{Specifies to use the \emph{A-channel} from \emph{'in2'} to calculate a displacement of the pixels in \emph{'in1'} along the y-axis. (The default.)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDisplacementMap]{feDisplacementMap}}}
}
}

\keyword{internal}
