% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDistantLightAzimuthAttribute}
\alias{feDistantLightAzimuthAttribute}
\title{azimuth}
\description{
Sets the angle between the x-axis and the projection into the xy-plane of the vector from the origin to the light source.
}
\section{Available Attribute Values}{

The value is a numeric representing an angle in degrees.
\describe{
\item{<angle>}{Specifies the azimuth angle (in degrees). The azimuth angle is defined as the angle between the x-axis and the projection into the xy-plane of vector from the origin to the light source. This angle is measured clockwise.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feDistantLight]{feDistantLight}}}
}
}

\keyword{internal}
