% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feDistantLightElevationAttribute}
\alias{feDistantLightElevationAttribute}
\title{elevation}
\description{
Sets the angle between the xy-plane and the vector from the origin to the light source.
}
\section{Available Attribute Values}{

The value is a numeric representing an angle in degrees.
\describe{
\item{<angle>}{Specifies the elevation angle (in degrees). The elevation angle is defined as the between the z-axis and the vector from the origin to the light source.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feDistantLight]{feDistantLight}}}
}
}

\keyword{internal}
