% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feMorphologyOperatorAttribute}
\alias{feMorphologyOperatorAttribute}
\title{operator}
\description{
Governs whether to erode (make thin) or dilate (make fatter).
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'erode'}}{Specifies to \emph{erode} (to thin)}
\item{\emph{'dilate'}}{Specifies to \emph{dilate} (to fatten)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feMorphology]{feMorphology}}}
}
}

\keyword{internal}
