% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feMorphologyRadiusAttribute}
\alias{feMorphologyRadiusAttribute}
\title{radius}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number.optional.number>}{The radius (or radii) for the operation. If two <number>s are provided, the first number represents a x.radius and the second value represents a y.radius. If one number is provided, then that value is used for both X and Y. The values are in the coordinate system established by attribute 'primitiveUnits' on the 'filter' element. A negative value is an error (see Error processing). A value of zero disables the effect of the given filter primitive (i.e., the result is a transparent black image). If the attribute is not specified, then the effect is as if a value of 0 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feMorphology]{feMorphology}}}
}
}
\keyword{internal}

