% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feOffsetDXYAttribute}
\alias{feOffsetDXYAttribute}
\title{dxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=feOffsetDxAttribute]{dx}}, \code{\link[=feOffsetDyAttribute]{dy}}
}

\section{Equivalence}{

\describe{
\item{}{dxy=c(value.DX,value.DY)}
\item{and}{dx = value.DX; dy = value.DY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feOffset]{feOffset}}}
}
}
\keyword{internal}

