% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{fePointLightXYZAttribute}
\alias{fePointLightXYZAttribute}
\title{xyz}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=fePointLightXAttribute]{x}}, \code{\link[=fePointLightYAttribute]{y}}, \code{\link[=fePointLightZAttribute]{z}}
}

\section{Equivalence}{

\describe{
\item{}{xyz=c(value.X,value.Y,value.Z)}
\item{and}{x = value.X; y = value.Y; z = value.Z}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=fePointLight]{fePointLight}}}
}
}
\keyword{internal}

