% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpecularLightingSpecularConstantAttribute}
\alias{feSpecularLightingSpecularConstantAttribute}
\title{specularConstant}
\description{
Sets the  ks term in the Phong lighting model.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies a scaling factor for the resulting RGB components, It is the \emph{kd} in \emph{Phong lighting model}. Default=1.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}

\keyword{internal}
