% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpecularLightingSpecularExponentAttribute}
\alias{feSpecularLightingSpecularExponentAttribute}
\title{specularExponent}
\description{
Sets the  specular term (the exponent)  in the specular lighting.  This  controls how shiny the object appears.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the \emph{exponent} for \emph{specular} term, used to control how shiny the object appears . Typical values should range from 1.0 to 128.0. (128 being very shiny.) Default=1.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}

\keyword{internal}
