% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpecularLightingSurfaceScaleAttribute}
\alias{feSpecularLightingSurfaceScaleAttribute}
\title{surfaceScale}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{height of surface when Ain = 1. If the attribute is not specified, then the effect is as if a value of 1 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}
\keyword{internal}

