% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightLimitingConeAngleAttribute}
\alias{feSpotLightLimitingConeAngleAttribute}
\title{limitingConeAngle}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{A limiting cone which restricts the region where the light is projected. No light is projected outside the cone. 'limitingConeAngle' represents the angle in degrees between the spot light axis (i.e. the axis between the light source and the point to which it is pointing at) and the spot light cone. User agents should apply a smoothing technique such as anti.aliasing at the boundary of the cone. If no value is specified, then no limiting cone will be applied.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}
\keyword{internal}

