% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{feSpotLightPOINTSATXYZAttribute}
\alias{feSpotLightPOINTSATXYZAttribute}
\title{pointsAtXYZ}
\description{
Shortcut for assigning pointsAtX, pointsAtY and pointsAtZ
}
\section{Combines}{

\code{\link[=feSpotLightPOINTSATXYZAttribute]{pointsAtX}}, \code{\link[=feSpotLightPOINTSATXYZAttribute]{pointsAtY}}, \code{\link[=feSpotLightPOINTSATXYZAttribute]{pointsAtZ}}
}

\section{Equivalence}{

\describe{
pointsAtXYZ =v is equivalent to  pointsAtX=v[1], pointsAtY=v[2], pointsAtZ=v[3]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Light Source Elements}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}

\keyword{internal}
