% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feSpotLightSpecularExponentAttribute}
\alias{feSpotLightSpecularExponentAttribute}
\title{specularExponent}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{Exponent value controlling the focus for the light source. If the attribute is not specified, then the effect is as if a value of 1 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=feSpotLight]{feSpotLight}}}
}
}
\keyword{internal}

