% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceBaseFrequencyAttribute}
\alias{feTurbulenceBaseFrequencyAttribute}
\title{baseFrequency}
\description{
Sets the base frequency for the noise function.
}
\section{Available Attribute Values}{

The value is a numeric vector having length 1 or 2.
\describe{
\item{<numeric>[1 or 2] (=fx or c(fx,fy))}{Specifies the base frequency parameter, \eqn{f_x}, \eqn{f_y}, used to generate the noise function in the X, Y directions respectively. The default for \eqn{f_x} is 0 (no effect), the default for \eqn{f_y} is \eqn{f_x}.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}

\keyword{internal}
