% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceNumOctavesAttribute}
\alias{feTurbulenceNumOctavesAttribute}
\title{numOctaves}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<integer>}{The numOctaves parameter for the noise function. If the attribute is not specified, then the effect is as if a value of 1 were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}
\keyword{internal}

