% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{feTurbulenceTypeAttribute}
\alias{feTurbulenceTypeAttribute}
\title{type}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{fractalNoise | turbulence}{Indicates whether the filter primitive should perform a noise or turbulence function. If attribute 'type' is not specified, then the effect is as if a value of turbulence were specified. Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feTurbulence]{feTurbulence}}}
}
}
\keyword{internal}

