% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FillAttribute}
\alias{FillAttribute}
\title{fill}
\description{
Governs whether to freeze or to remove the animation upon completion.  Freeze means stop in place, remove means to return to the state prior to the beginning of the animation.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'freeze'}}{Specifies upon completing the animation to freeze in place, there by keeping the last value for that animation frozen.}
\item{\emph{'remove'}}{Specifies upon completing the animation to return to the initial state, that being what value the attribute had prior to the start of the animation.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
