% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterElementPrimitiveUnitsAttribute}
\alias{FilterElementPrimitiveUnitsAttribute}
\title{primitiveUnits}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{userSpaceOnUse | objectBoundingBox}{Specifies the coordinate system for the various length values within the filter primitives and for the attributes that define the filter primitive subregion. If primitiveUnits="userSpaceOnUse", any length values within the filter definitions represent values in the current user coordinate system in place at the time when the 'filter' element is referenced (i.e., the user coordinate system for the element referencing the 'filter' element via a 'filter' property). If primitiveUnits="objectBoundingBox", then any length values within the filter definitions represent fractions or percentages of the bounding box on the referencing element (see Object bounding box units). Note that if only one number was specified in a <number.optional.number> value this number is expanded out before the 'primitiveUnits' computation takes place. If attribute 'primitiveUnits' is not specified, then the effect is as if a value of userSpaceOnUse were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=filter]{filter}}}
}
}
\keyword{internal}

