% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{FilterElementWHAttribute}
\alias{FilterElementWHAttribute}
\title{wh}
\description{
Shortcut for assigning height and width
}
\section{Combines}{

\code{\link[=FilterElementWHAttribute]{height}}, \code{\link[=FilterElementWHAttribute]{width}}
}

\section{Equivalence}{

\describe{
wh =v is equivalent to  height=v[1], width=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=filter]{filter}}}
}
}

\keyword{internal}
