% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterElementWidthAttribute}
\alias{FilterElementWidthAttribute}
\title{width}
\description{
Sets the width of the filter region.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the bounding box width of the \strong{filter} in terms of the \emph{filterUnits}. (default is 1.20 )}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=filter]{filter}}}
}
}

\keyword{internal}
