% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterElementXAttribute}
\alias{FilterElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the left-hand side of the filter effects region.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the left edge of the \strong{filter element} region.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=filter]{filter}}}
}
}

\keyword{internal}
