% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FilterPrimitiveIn1Attribute}
\alias{FilterPrimitiveIn1Attribute}
\title{in1}
\description{
Sets the primary graphics input to be used for the given filter primitive.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'SourceGraphic'}}{Specifies taking as the first graphics input argument, \emph{in1}, the original graphic source.}
\item{\emph{'SourceAlpha'}}{Specifies taking as the first graphics input argument, \emph{in1}, taking the alpha channel of the original graphic source.}
\item{\emph{'BackgroundImage'}}{Specifies taking as the first graphics input argument, \emph{in1}, the image snapshot of the filter region in effect at the time filter was invoked.}
\item{\emph{'BackgroundAlpha'}}{Specifies taking as the first graphics input argument, \emph{in1}, the \emph{alpha channel} of the filter region image.}
\item{\emph{'FillPaint'}}{Specifies taking as the first graphics input argument, \emph{in1}, the \emph{fill} property on the target element.}
\item{\emph{'StrokePaint'}}{Specifies taking as the first graphics input argument, \emph{in1}, the \emph{stroke} property on the target element.}
\item{<filter-primitive-reference>}{Specifies taking as the first graphics input argument, \emph{in1}, the results of the referenced filter primitive.}
\item{<filterPrimitveElement>}{Specifies taking as the first graphics input argument, \emph{in1}, the results of the given filter primitive.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feBlend]{feBlend}}, \code{\link[=feColorMatrix]{feColorMatrix}}, \code{\link[=feComponentTransfer]{feComponentTransfer}}, \code{\link[=feComposite]{feComposite}}, \code{\link[=feConvolveMatrix]{feConvolveMatrix}}, \code{\link[=feDiffuseLighting]{feDiffuseLighting}}, \code{\link[=feDisplacementMap]{feDisplacementMap}}, \code{\link[=feGaussianBlur]{feGaussianBlur}}, \code{\link[=feMorphology]{feMorphology}}, \code{\link[=feOffset]{feOffset}}, \code{\link[=feSpecularLighting]{feSpecularLighting}}, \code{\link[=feTile]{feTile}}}
\item{\emph{Uncategorized Elements}}{\code{\link[=feMergeNode]{feMergeNode}}}
}
}

\keyword{internal}
