% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{flood-opacity-presentationAttribute}
\alias{flood-opacity-presentationAttribute}
\title{flood.opacity}
\description{
Defines the opacity value to be used across the entire filter primitive subregion.
}
\section{Available Attribute Values}{



\describe{
\item{<opacity-value>}{Specifies the opacity value, a  number between 0 and 1 indicating the opacity to be used as fill , with 1 being opaque and 0 being transparent.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the opacity value from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feFlood]{feFlood}}}
}
}

\keyword{internal}
