% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{font-face-format}
\alias{font-face-format}
\alias{font.face.format}
\title{An element for specifing the font face format.}
\description{
The font-face-format element describes the type of font referenced by its parent <font-face-uri>.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Unclassified}}{\code{\link[=FontFaceFormatElementStringAttribute]{string}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
{No Content Elements Applicable}{!}
}
}

\keyword{element}
