% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-stretch-presentationAttribute}
\alias{font-stretch-presentationAttribute}
\title{font.stretch}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{normal}{**ToDo!!!** }
\item{wider}{**ToDo!!!** }
\item{narrower}{**ToDo!!!** }
\item{ultra-condensed}{**ToDo!!!** }
\item{extra-condensed}{**ToDo!!!** }
\item{condensed}{**ToDo!!!** }
\item{semi-condensed}{**ToDo!!!** }
\item{semi-expanded}{**ToDo!!!** }
\item{expanded}{**ToDo!!!** }
\item{extra-expanded}{**ToDo!!!** }
\item{ultra-expanded}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
\item{\emph{Unclassfied:}}{\code{\link[=text]{text}}}
}
}
\keyword{internal}

