% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{font-style-presentationAttribute}
\alias{font-style-presentationAttribute}
\title{font.style}
\description{
Governs which font face to use:  normal, italic or oblique face.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to Inherit the font style from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'italic'}'}'}'}}{Specifies to use an italicl style}
\item{\emph{'\emph{'\emph{'\emph{'normal'}'}'}'}}{Specifies to use a normal style}
\item{\emph{'\emph{'\emph{'\emph{'oblique'}'}'}'}}{Specifies to use an oblique style}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
