% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontElementHorizAdvXAttribute}
\alias{FontElementHorizAdvXAttribute}
\title{horiz.adv.x}
\description{
Sets the amount of horizontal advance to after rendering the given horizontally orientated glyph.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies what the default horizontal advance should be after rendering in a horizontal orientation. Needs to be non-negative.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font]{font}}}
}
}

\keyword{internal}
