% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{FontElementHORIZORIGINXYAttribute}
\alias{FontElementHORIZORIGINXYAttribute}
\title{horiz.origin.xy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=FontElementHorizOriginXAttribute]{horiz.origin.x}}, \code{\link[=FontElementHorizOriginYAttribute]{horiz.origin.y}}
}

\section{Equivalence}{

\describe{
\item{}{horiz.origin.xy=c(value.HORIZ.ORIGIN.X,value.HORIZ.ORIGIN.Y)}
\item{and}{horiz.origin.x = value.HORIZ.ORIGIN.X; horiz.origin.y = value.HORIZ.ORIGIN.Y}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font]{font}}}
}
}
\keyword{internal}

