% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontElementVertOriginXAttribute}
\alias{FontElementVertOriginXAttribute}
\title{vert.origin.x}
\description{
Sets the x-coordinate of the glyphs origin to be applied to a vertically orientented text. This value is to be interpreted within the fonts coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the default x-coordinate of the origin for a \emph{glyph} relative to coordinate system when in text orientation is vertical. Default is half of 'horiz.adv.x.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font]{font}}}
}
}

\keyword{internal}
