% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{FontElementVERTORIGINXYAttribute}
\alias{FontElementVERTORIGINXYAttribute}
\title{vert.origin.xy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=FontElementVertOriginXAttribute]{vert.origin.x}}, \code{\link[=FontElementVertOriginYAttribute]{vert.origin.y}}
}

\section{Equivalence}{

\describe{
\item{}{vert.origin.xy=c(value.VERT.ORIGIN.X,value.VERT.ORIGIN.Y)}
\item{and}{vert.origin.x = value.VERT.ORIGIN.X; vert.origin.y = value.VERT.ORIGIN.Y}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font]{font}}}
}
}
\keyword{internal}

