% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementBboxAttribute}
\alias{FontFaceElementBboxAttribute}
\title{bbox}
\description{
Sets the maximal bounding box of the font.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>[4] (=c(left-x, lower-y, right-x, upper-y))}{Specifies the maximal bounding box of the font. This is represented as a vector c(lower-left-x, lower-left-y, upper-right-x, upper-right-y).}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
