% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementDescentAttribute}
\alias{FontFaceElementDescentAttribute}
\title{descent}
\description{
Sets the maximum unaccented depth of the font. Measurements to be  expressed in terms of  the font coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies an maximum unaccented font depth with respect to the font coordinate system.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
