% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontSizeAttribute}
\alias{FontFaceElementFontSizeAttribute}
\title{font.size}
\description{
Governs the font size.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{\emph{'all'}}{Specifies all size (suitable for scalable fonts)}
\item{<length>}{Specifies the font size (suitable for bitmap fonts)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
