% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontStretchAttribute}
\alias{FontFaceElementFontStretchAttribute}
\title{font.stretch}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{all | [ normal | ultra.condensed | extra.condensed | condensed | semi.condensed | semi.expanded | expanded | extra.expanded | ultra.expanded] [, [ normal | ultra.condensed | extra.condensed | condensed | semi.condensed | semi.expanded | expanded | extra.expanded | ultra.expanded] ]*}{Same syntax and semantics as the 'font.stretch' descriptor within an font.face rule. Indication of the condensed or expanded nature of the face relative to others in the same font family. Takes on the same values as the 'font.stretch' property except that. relative keywords (wider, narrower) are not permitteda comma.separated list is permittedthe keyword 'all' is permitted If the attribute is not specified, the effect is as if a value of 'normal' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

