% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontStyleAttribute}
\alias{FontFaceElementFontStyleAttribute}
\title{font.style}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{all | [ normal | italic | oblique] [, [normal | italic | oblique]]*}{Same syntax and semantics as the 'font.style' descriptor within an font.face rule. The style of a font. Takes on the same values as the 'font.style' property, except that a comma.separated list is permitted. If the attribute is not specified, the effect is as if a value of 'all' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

