% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontVariantAttribute}
\alias{FontFaceElementFontVariantAttribute}
\title{font.variant}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{[normal | small.caps] [,[normal | small.caps]]*}{Same syntax and semantics as the 'font.variant' descriptor within an font.face rule. Indication of whether this face is the small.caps variant of a font. Takes on the same values as the 'font.variant' property, except that a comma.separated list is permitted. If the attribute is not specified, the effect is as if a value of 'normal' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

