% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementFontWeightAttribute}
\alias{FontFaceElementFontWeightAttribute}
\title{font.weight}
\description{
Governs the weight of the font (ie. Bold, etc)
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'all'}}{Specifies to match all possible weights.}
\item{\emph{'normal'}}{Specifies a \emph{normal} weight (i.e. 400)}
\item{\emph{'bold'}}{Specifies a \emph{bold} weight (i.e. 700)}
\item{\emph{'100'}}{Specifies a 100 font weight (lightest)}
\item{\emph{'200'}}{Specifies a 200 font weight.}
\item{\emph{'300'}}{Specifies a 300 font weight.}
\item{\emph{'400'}}{Specifies a 400 font weight. (Normal.)}
\item{\emph{'500'}}{Specifies a 500 font weight.}
\item{\emph{'600'}}{Specifies a 600 font weight}
\item{\emph{'700'}}{Specifies a 700 font weight}
\item{\emph{'800'}}{Specifies a 800 font weight}
\item{\emph{'900'}}{Specifies a 900 font weight. (Boldest)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
