% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementPanose1Attribute}
\alias{FontFaceElementPanose1Attribute}
\title{panose.1}
\description{
Set panrose-1 numbers (for True-type font matching and classification)
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{[<integer>]{10}}{Specifies \emph{panrose} using vector of length 10. (Default=c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0)). Used by Latin fonts}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
