% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementSlopeAttribute}
\alias{FontFaceElementSlopeAttribute}
\title{slope}
\description{
Sets the vertical stroke angle of the font.
}
\section{Available Attribute Values}{

The value is a numeric representing an angle in degrees.
\describe{
\item{<angle>}{Specifies an angle in degrees from the vertical of the dominant vertical strokes of the font. (Default=0)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
