% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnderlineThicknessAttribute}
\alias{FontFaceElementUnderlineThicknessAttribute}
\title{underline.thickness}
\description{
Sets the preferred thickness using the seame units as the font-coordinates to render a underline.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies, in font coordinates, the intended \emph{underline thickness}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
