% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnicodeRangeAttribute}
\alias{FontFaceElementUnicodeRangeAttribute}
\title{unicode.range}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<urange> [, <urange>]*}{Same syntax and semantics as the 'unicode.range' descriptor within an font.face rule. The range of ISO 10646 characters [UNICODE] possibly covered by the glyphs in the font. Except for any additional information provided in this specification, the normative definition of the attribute is in CSS2 ([CSS2], section 15.3.3). If the attribute is not specified, the effect is as if a value of 'U+0.10FFFF' were specified.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

