% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementUnitsPerEmAttribute}
\alias{FontFaceElementUnitsPerEmAttribute}
\title{units.per.em}
\description{
Sets the the number of coordinate units along the sides the em square of the glyph design grid
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies the number of coordinate units on an em square. (Default = 1000)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face]{font.face}}}
}
}

\keyword{internal}
