% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceElementXHeightAttribute}
\alias{FontFaceElementXHeightAttribute}
\title{x.height}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{Same syntax and semantics as the 'x.height' descriptor within an font.face rule. The height of lowercase glyphs in the font within the font coordinate system.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=font-face]{font.face}}}
}
}
\keyword{internal}

