% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceFormatElementStringAttribute}
\alias{FontFaceFormatElementStringAttribute}
\title{string}
\description{
Sets which  formats are supported by the given font.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<character>[N]}{Specifies a list of formats supported by the referenced font. Examples are \emph{truedoc-pfr}, or \emph{opentype}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face-format]{font.face.format}}}
}
}

\keyword{internal}
