% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceNameElementNameAttribute}
\alias{FontFaceNameElementNameAttribute}
\title{name}
\description{
Sets the name of the local font.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<name>}{Specifies a name of a local font.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face-name]{font.face.name}}}
}
}

\keyword{internal}
