% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{FontFaceUriElementHrefAttribute}
\alias{FontFaceUriElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference for the location of a font.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies the location of the referenced font}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=font-face-uri]{font.face.uri}}}
}
}

\keyword{internal}
