% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{foreignObject}
\alias{foreignObject}
\title{An element used to include a non-svg element}
\description{
The foreignObject element allows for inclusion of a foreign XML namespace which has its graphical content drawn by a different user agent. The included foreign graphical content is subject to SVG transformations and compositing.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Animation Event Attributes}}{\code{\link[=OnLoadEventAttribute]{onload}}}
\item{\emph{Complementary Attributes}}{\code{\link[=ForeignObjectElementWHAttribute]{wh}}, \code{\link[=ForeignObjectElementXYAttribute]{xy}}}
\item{\emph{Conditional Processing Attributes}}{\code{\link[=RequiredExtensionsAttribute]{requiredExtensions}}, \code{\link[=RequiredFeaturesAttribute]{requiredFeatures}}, \code{\link[=SystemLanguageAttribute]{systemLanguage}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Filter Primitive Attributes}}{\code{\link[=ForeignObjectElementHeightAttribute]{height}}, \code{\link[=ForeignObjectElementWidthAttribute]{width}}, \code{\link[=ForeignObjectElementXAttribute]{x}}, \code{\link[=ForeignObjectElementYAttribute]{y}}}
\item{\emph{Graphical Event Attributes}}{\code{\link[=OnActivateEventAttribute]{onactivate}}, \code{\link[=OnClickEventAttribute]{onclick}}, \code{\link[=OnFocusInEventAttribute]{onfocusin}}, \code{\link[=OnFocusOutEventAttribute]{onfocusout}}, \code{\link[=OnLoadEventAttribute]{onload}}, \code{\link[=OnMouseDownEventAttribute]{onmousedown}}, \code{\link[=OnMouseMoveEventAttribute]{onmousemove}}, \code{\link[=OnMouseOutEventAttribute]{onmouseout}}, \code{\link[=OnMouseOverEventAttribute]{onmouseover}}, \code{\link[=OnMouseUpEventAttribute]{onmouseup}}}
\item{\emph{Presentation Attributes}}{\code{\link[=clip-presentationAttribute]{clip}}, \code{\link[=display-presentationAttribute]{display}}, \code{\link[=overflow-presentationAttribute]{overflow}}}
\item{\emph{Unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=ExternalResourcesRequiredAttribute]{externalResourcesRequired}}, \code{\link[=StyleAttribute]{style}}, \code{\link[=TransformAttribute]{transform}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
{No Content Elements Applicable}{!}
}
}

\keyword{element}
