% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ForeignObjectElementXAttribute}
\alias{ForeignObjectElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{The x.axis coordinate of one corner of the rectangular region into which the graphics associated with the contents of the 'foreignObject' will be rendered. If the attribute is not specified, the effect is as if a value of '0' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=foreignObject]{foreignObject}}}
}
}
\keyword{internal}

