% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ForeignObjectElementYAttribute}
\alias{ForeignObjectElementYAttribute}
\title{y}
\description{
Sets the y-coord of the top side of rectangular region into which the referenced document is placed.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-coordinate of the top side of the rectangular region into which the \emph{'foreignObject'} will be rendered. Default=0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=foreignObject]{foreignObject}}}
}
}

\keyword{internal}
