% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementArabicFormAttribute}
\alias{GlyphElementArabicFormAttribute}
\title{arabic.form}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{initial | medial | terminal | isolated}{For Arabic glyphs, indicates which of the four possible forms this glyph represents.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyph]{glyph}}}
}
}
\keyword{internal}

