% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementLangAttribute}
\alias{GlyphElementLangAttribute}
\title{lang}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{LanguageCodes;}{The attribute value is a comma.separated list of language names as defined in BCP 47 [BCP47]. The glyph can be used if the 'xml.lang' attribute exactly matches one of the languages given in the value of this parameter, or if the 'xml.lang' attribute exactly equals a prefix of one of the languages given in the value of this parameter such that the first tag character following the prefix is ".".Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyph]{glyph}}}
}
}
\keyword{internal}

