% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementOrientationAttribute}
\alias{GlyphElementOrientationAttribute}
\title{orientation}
\description{
Governs the orientation of the the glyph with respect to the inline-progression-direction.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'h'}}{Specifies that the \strong{glyph element} is to be exclusively rendered in a horizontal inline progression.}
\item{\emph{'v'}}{Specifies that the \strong{glyph element} is to be exclusively rendered in a  vertical inline progression.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}}
}
}

\keyword{internal}
