% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementVertAdvYAttribute}
\alias{GlyphElementVertAdvYAttribute}
\title{vert.adv.y}
\description{
Sets the amount to advance to occur after rendering this glyph, where both the orientation and the direction of advance are vertical.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<numeric>}{Specifies what the default vertical advance should be after rendering in a vertical orientation. Needs to be non-negative. Default is vert-adv-y of font.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=glyph]{glyph}}, \code{\link[=missing-glyph]{missing.glyph}}}
}
}

\keyword{internal}
