% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphElementVertOriginXAttribute}
\alias{GlyphElementVertOriginXAttribute}
\title{vert.origin.x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The X.coordinate in the font coordinate system of the origin of the glyph to be used when drawing vertically oriented text. If the attribute is not specified, the effect is as if the attribute were set to the value of the font's 'vert.origin.x' attribute.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyph]{glyph}}, \code{\link[=missing-glyph]{missing.glyph}}}
}
}
\keyword{internal}

