% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{glyphRef}
\alias{glyphRef}
\title{NA}
\description{
The glyphRef element provides a single possible glyph to the referencing <altglyph> substitution.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{combining attributes}}{\code{\link[=GlyphRefElementDXYAttribute]{dxy}}, \code{\link[=GlyphRefElementXYAttribute]{xy}}}
\item{\emph{core attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{filter primitive attributes}}{\code{\link[=GlyphRefElementXAttribute]{x}}, \code{\link[=GlyphRefElementYAttribute]{y}}}
\item{\emph{unclassified}}{\code{\link[=ClassAttribute]{class}}, \code{\link[=GlyphRefElementDXAttribute]{dx}}, \code{\link[=GlyphRefElementDYAttribute]{dy}}, \code{\link[=GlyphRefElementFormatAttribute]{format}}, \code{\link[=GlyphRefElementGlyphRefAttribute]{glyphRef}}, \code{\link[=StyleAttribute]{style}}}
\item{\emph{xlink attributes}}{\code{\link[=XLinkActuateAttribute]{xlink.actuate}}, \code{\link[=XLinkArcRoleAttribute]{xlink.arcrole}}, \code{\link[=GlyphRefElementHrefAttribute]{xlink.href}}, \code{\link[=XLinkRoleAttribute]{xlink.role}}, \code{\link[=XLinkShowAttribute]{xlink.show}}, \code{\link[=XLinkTitleAttribute]{xlink.title}}, \code{\link[=XLinkTypeAttribute]{xlink.type}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
\item{\emph{Unclassfied:}}{\code{Empty.}}
}
}
\keyword{element}

