% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{GlyphRefElementDXYAttribute}
\alias{GlyphRefElementDXYAttribute}
\title{dxy}
\description{
Shortcut for assigning dx and dy
}
\section{Combines}{

\code{\link[=GlyphRefElementDXYAttribute]{dx}}, \code{\link[=GlyphRefElementDXYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
dxy =v is equivalent to  dx=v[1], dy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
