% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementDYAttribute}
\alias{GlyphRefElementDYAttribute}
\title{dy}
\description{
Set shifts in the current y-position prior to rendering.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies a y-shift for this \emph{glyph} using the font's coordinate system for this glyph.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
