% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementFormatAttribute}
\alias{GlyphRefElementFormatAttribute}
\title{format}
\description{
Sets the fonts glyph font.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<glyph-format>}{Specifies a \emph{Glyph} format.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
