% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementHrefAttribute}
\alias{GlyphRefElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to a glyph element in this svg fragment..
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies a refernce to an alternative glyph.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
