% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementXAttribute}
\alias{GlyphRefElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the glyph, the whose value is interpreted within the fonts coordinate system.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<number>}{Specifies the x-coordinate for the \emph{glyph} in terms of the fonts coordinate system.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
