% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{GlyphRefElementXYAttribute}
\alias{GlyphRefElementXYAttribute}
\title{xy}
\description{
Shortcut for assigning x and y
}
\section{Combines}{

\code{\link[=GlyphRefElementXYAttribute]{x}}, \code{\link[=GlyphRefElementXYAttribute]{y}}
}

\section{Equivalence}{

\describe{
xy =v is equivalent to  x=v[1], y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=glyphRef]{glyphRef}}}
}
}

\keyword{internal}
