% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{GlyphRefElementYAttribute}
\alias{GlyphRefElementYAttribute}
\title{y}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The corresponding new absolute Y coordinate within the font's coordinate system for this glyph.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=glyphRef]{glyphRef}}}
}
}
\keyword{internal}

