% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{HKernElementG12Attribute}
\alias{HKernElementG12Attribute}
\title{g12}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=HKernElementG1Attribute]{g1}}, \code{\link[=HKernElementG2Attribute]{g2}}
}

\section{Equivalence}{

\describe{
\item{}{g12=c(value.G1,value.G2)}
\item{and}{g1 = value.G1; g2 = value.G2}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}
\keyword{internal}

