% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{HKernElementG1Attribute}
\alias{HKernElementG1Attribute}
\title{g1}
\description{
Sets a collection of names to be used as  candiates for the first glyph  in the kerning pair.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<glyph-names>[N]}{Specifies a set of candiate 1st \emph{glyphs} in the kerning pair using a vector of glyph names.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}

\keyword{internal}
