% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{HKernElementU12Attribute}
\alias{HKernElementU12Attribute}
\title{u12}
\description{
Shortcut for assigning u1 and u2
}
\section{Combines}{

\code{\link[=HKernElementU12Attribute]{u1}}, \code{\link[=HKernElementU12Attribute]{u2}}
}

\section{Equivalence}{

\describe{
u12 =v is equivalent to  u1=v[1], u2=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}

\keyword{internal}
