% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{HKernElementU1Attribute}
\alias{HKernElementU1Attribute}
\title{u1}
\description{
Sets the  first glyphs candidates.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<glyph-unicode>[N]}{Specifies possible 1st \emph{glyphs} in the \emph{kerning pair} using a vector of Unicode characters or Unicode ranges.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}

\keyword{internal}
