% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{HKernElementU2Attribute}
\alias{HKernElementU2Attribute}
\title{u2}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{[<character> | <urange>] [, [<character> | <urange>] ]*}{Same as the 'u1' attribute, except that 'u2' specifies possible second glyphs in the kerning pair.Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=hkern]{hkern}}, \code{\link[=vkern]{vkern}}}
}
}
\keyword{internal}

