% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{image-rendering-presentationAttribute}
\alias{image-rendering-presentationAttribute}
\title{image.rendering}
\description{
Governs the optimization strategy of rendering images.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'auto'}'}'}'}}{Specifies to employ a trade-off strategy of image-rendering optimization which balances speed vs quality.}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the image-rendering optimzation from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'optimizeQuality'}'}'}'}}{Specifies to optimize the image-rendiering  for quality.}
\item{\emph{'\emph{'\emph{'\emph{'optimizeSpeed'}'}'}'}}{Specifies to optimize the image-rendiering  for speed}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=image]{image}}}
}
}

\keyword{internal}
