% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ImageElementHrefAttribute}
\alias{ImageElementHrefAttribute}
\title{xlink.href}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<iri>}{A IRI reference.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=image]{image}}}
}
}
\keyword{internal}

