% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{ImageElementXYAttribute}
\alias{ImageElementXYAttribute}
\title{xy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=ImageElementXAttribute]{x}}, \code{\link[=ImageElementYAttribute]{y}}
}

\section{Equivalence}{

\describe{
\item{}{xy=c(value.X,value.Y)}
\item{and}{x = value.X; y = value.Y}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=image]{image}}}
}
}
\keyword{internal}

