% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{KeySplinesAttribute}
\alias{KeySplinesAttribute}
\title{keySplines}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list>}{A set of Bzier control points associated with the 'keyTimes' list, defining a cubic Bzier function that controls interval pacing. The attribute value is a semicolon.separated list of control point descriptions. Each control point description is a set of four values. x1 y1 x2 y2, describing the Bzier control points for one time segment. Note. SMIL allows these values to be separated either by commas with optional whitespace, or by whitespace alone. The 'keyTimes' values that define the associated segment are the Bzier "anchor points", and the 'keySplines' values are the control points. Thus, there must be one fewer sets of control points than there are 'keyTimes'. The values must all be in the range 0 to 1. This attribute is ignored unless the 'calcMode' is set to 'spline'. If there are any errors in the 'keySplines' specification (bad values, too many or too few values), the document fragment is in error (see error processing). Except for any SVG.specific rules explicitly mentioned in this specification, the normative definition for this attribute is the SMIL Animation specification. In particular, see SMIL Animation. 'keySplines' attribute ([SMILANIM], section 3.2.3).}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}}
}
}
\keyword{internal}

