% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{lighting-color-presentationAttribute}
\alias{lighting-color-presentationAttribute}
\title{lighting.color}
\description{
Governs the color of the light source for filter primitives feDiffuseLighting and feSpecularLighting.
}
\section{Available Attribute Values}{



\describe{
\item{<color> [<icccolor>]}{Sets the color of the lighting. The color may be given in any one of 3 ways: name (i.e. red), numeric (#f00 or #ff0000) or as ab rgb (rgb(255,0,0)).}
\item{\emph{'\emph{'currentColor'}'}}{Specifes to use the \emph{currentColor} for the light color. The current color is the value of the color specified by the \emph{color} property.}
\item{\emph{'\emph{'inherit'}'}}{Specifies inheriting the lighting color from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feDiffuseLighting]{feDiffuseLighting}}, \code{\link[=feSpecularLighting]{feSpecularLighting}}}
}
}

\keyword{internal}
