% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{LinearGradientElementGradientUnitsAttribute}
\alias{LinearGradientElementGradientUnitsAttribute}
\title{gradientUnits}
\description{
Governs whether to interpret the coordinates of x1, y1, x2, y2 as either objectBoundingBox or userSpaceOnUse.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that the gradient values for \emph{'x1'}, \emph{'y1'}, \emph{'x2'} and \emph{'y2'} are given in terms of the user coordinate system in effect at the time when the gradient element is referenced.}
\item{\emph{'objectBoundingBox'}}{Specifies that the gradient values for \emph{'x1'}, \emph{'y1'}, \emph{'x2'} and \emph{'y2'} are given in terms fractions (or percentages) of the bounding box of the target element to which the gradient is applied. In this case c(0, 0) is the coordinates of the upper left-hand corner of the bounding, and c(1, 0) is the coordinates of the lower right-hand corner of the bounding box.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
