% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{LinearGradientElementSpreadMethodAttribute}
\alias{LinearGradientElementSpreadMethodAttribute}
\title{spreadMethod}
\description{
Governs the behaviour that should occur if the gradient starts or ends inside the bounds of the target rectangle.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'pad'}}{Specifies to extend the painting beyond the \emph{gradient} vector of the \strong{linearGradient}, by using terminal colors found at the end points. (Default)}
\item{\emph{'reflect'}}{Specifies to extend the painting beyond the \emph{gradient} vector of the \strong{linearGradient}, by reflecting the gradient pattern. The reflections are continued until the entire gradient region is filled. For example, we might have a pattern of red white and blue (red at the start of the gradient vector, blue at the end), and then observe: red white blue blue white red red white blue.}
\item{\emph{'repeat'}}{Specifies to extend the painting beyound the \emph{gradient} vector of the \strong{linearGradient}, by repeating the gradient pattern until the region is filled. For example, For example, we might have a pattern of red white and blue (red at the start of the gradient vector, blue at the end), and then observe: red white blue red white blue red white blue.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
