% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementX12Attribute}
\alias{LinearGradientElementX12Attribute}
\title{x12}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=LinearGradientElementX1Attribute]{x1}}, \code{\link[=LinearGradientElementX2Attribute]{x2}}
}

\section{Equivalence}{

\describe{
\item{}{x12=c(value.X1,value.X2)}
\item{and}{x1 = value.X1; x2 = value.X2}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}
\keyword{internal}

