% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{LinearGradientElementX2Attribute}
\alias{LinearGradientElementX2Attribute}
\title{x2}
\description{
Sets the x-coordinate of the ending point for a linear gradient.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the ending point of the \emph{gradient} vector. Can be numeric or percentage.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
