% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementXY1Attribute}
\alias{LinearGradientElementXY1Attribute}
\title{xy1}
\description{
Shortcut for assigning x1 and y1
}
\section{Combines}{

\code{\link[=LinearGradientElementXY1Attribute]{x1}}, \code{\link[=LinearGradientElementXY1Attribute]{y1}}
}

\section{Equivalence}{

\describe{
xy1 =v is equivalent to  x1=v[1], y1=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}

\keyword{internal}
