% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementXY2Attribute}
\alias{LinearGradientElementXY2Attribute}
\title{xy2}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=LinearGradientElementX2Attribute]{x2}}, \code{\link[=LinearGradientElementY2Attribute]{y2}}
}

\section{Equivalence}{

\describe{
\item{}{xy2=c(value.X2,value.Y2)}
\item{and}{x2 = value.X2; y2 = value.Y2}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}
\keyword{internal}

