% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LinearGradientElementY12Attribute}
\alias{LinearGradientElementY12Attribute}
\title{y12}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=LinearGradientElementY1Attribute]{y1}}, \code{\link[=LinearGradientElementY2Attribute]{y2}}
}

\section{Equivalence}{

\describe{
\item{}{y12=c(value.Y1,value.Y2)}
\item{and}{y1 = value.Y1; y2 = value.Y2}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=linearGradient]{linearGradient}}}
}
}
\keyword{internal}

