% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{LineElementX1Attribute}
\alias{LineElementX1Attribute}
\title{x1}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{The x.axis coordinate of the start of the line. If the attribute is not specified, the effect is as if a value of "0" were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}}
}
}
\keyword{internal}

