% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{LineElementXY1Attribute}
\alias{LineElementXY1Attribute}
\title{xy1}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=LineElementX1Attribute]{x1}}, \code{\link[=LineElementY1Attribute]{y1}}
}

\section{Equivalence}{

\describe{
\item{}{xy1=c(value.X1,value.Y1)}
\item{and}{x1 = value.X1; y1 = value.Y1}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}}
}
}
\keyword{internal}

