% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{LineElementY1Attribute}
\alias{LineElementY1Attribute}
\title{y1}
\description{
Sets the y-coordinate of the first point of the line
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-axis coordinate of the starting point of the \strong{line} element. Default is zero.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=line]{line}}}
}
}

\keyword{internal}
