% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerElementRefXAttribute}
\alias{MarkerElementRefXAttribute}
\title{refX}
\description{
Sets the x-coordinate of the marker reference point within the marker coordinate system.  Upon rendering, the marker reference point coincides with the point on the curve where the marker is to be attached.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the \strong{marker} position. Default=0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=marker]{marker}}}
}
}

\keyword{internal}
