% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{MarkerElementREFXYAttribute}
\alias{MarkerElementREFXYAttribute}
\title{refXY}
\description{
Shortcut for assigning refX and refY
}
\section{Combines}{

\code{\link[=MarkerElementREFXYAttribute]{refX}}, \code{\link[=MarkerElementREFXYAttribute]{refY}}
}

\section{Equivalence}{

\describe{
refXY =v is equivalent to  refX=v[1], refY=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=marker]{marker}}}
}
}

\keyword{internal}
