% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerUnitsAttribute}
\alias{MarkerUnitsAttribute}
\title{markerUnits}
\description{
Determines how to interpret marker coordinates, the choices being either using the stroke width as 1 unit (strokeWidthst), or using the units inherited from current user coordinate system ( userSpaceOnUse).
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that \emph{'markerWidth'}, \emph{'markerHeight'} and any coordinate values of marker contents are in terms of the user coordinate system in effect at the time when this marker is referenced.}
\item{\emph{'strokeWidth'}}{Specifies that \emph{'markerWidth'}, \emph{'markerHeight'} and any-coordinate values of the marker contents are given in terms of a coordinate system which has been scaled in such a way that the current stroke has a width of 1 in that system.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=marker]{marker}}}
}
}

\keyword{internal}
