% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MarkerWidthAttribute}
\alias{MarkerWidthAttribute}
\title{markerWidth}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{Represents the width of the viewport into which the marker is to be fitted when it is rendered. A negative value is an error (see Error processing). A value of zero disables rendering of the element. If the attribute is not specified, the effect is as if a value of "3" were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=marker]{marker}}}
}
}
\keyword{internal}

