% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MaskElementMaskUnitsAttribute}
\alias{MaskElementMaskUnitsAttribute}
\title{maskUnits}
\description{
Governs whether to interpret the coordinates x,y,width, height as objectBoundingBox or userSpaceOnUse
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'userSpaceOnUse'}}{Specifies that the coordinate system for the mask attributes \emph{'x'}, \emph{'y'}, \emph{'width'} and \emph{'height'} represent values are given in terms of the user coordinate system effective at the time of mask referencing.}
\item{\emph{'objectBoundingBox'}}{Specifies that the coordinate system for the mask attributes \emph{'x'}, \emph{'y'}, \emph{'width'} and \emph{'height'} are given in terms of fractions or percentages of the bounding box of the object to which the mask is applied. (Default)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=mask]{mask}}}
}
}

\keyword{internal}
