% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{MaskElementYAttribute}
\alias{MaskElementYAttribute}
\title{y}
\description{
Sets y coordinate of the top-side of the rectangular region defining the largest possible offscreen buffer.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-coordinate of top side of a rectangle defining the offscreen buffer used in computing the masked image. The resulting masked imageis the intersection of the mask and that buffer. The default is y=.1, maskUnits=\emph{'objectBoundingBox'}).}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=mask]{mask}}}
}
}

\keyword{internal}
