% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OnRepeatEventAttribute}
\alias{OnRepeatEventAttribute}
\title{onrepeat}
\description{
Sets the script to execute when an repeat event is received
}
\section{Available Attribute Values}{

The value is a script, likely ecmascript (i.e.  javascript).
\describe{
\item{<script>}{Specifies what to execute when a \emph{repeatEvent} event is fired. A \emph{repeatEvent} occurs at the repeat of an animation.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
