% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{opacity-presentationAttribute}
\alias{opacity-presentationAttribute}
\title{opacity}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\itemize{
\item{<opacity-value>}{**ToDo!!!** }
\item{inherit}{**ToDo!!!** }
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Structural Elements}}{\code{\link[=defs]{defs}}, \code{\link[=g]{g}}, \code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}, \code{\link[=use]{use}}}
\item{\emph{Unclassfied:}}{\code{\link[=a]{a}}, \code{\link[=glyph]{glyph}}, \code{\link[=image]{image}}, \code{\link[=marker]{marker}}, \code{\link[=missing-glyph]{missing.glyph}}, \code{\link[=pattern]{pattern}}, \code{\link[=switch]{switch}}, \code{\link[=text]{text}}}
}
}
\keyword{internal}

