% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OrientAttribute}
\alias{OrientAttribute}
\title{orient}
\description{
Governs the orientation of the marker,
}
\section{Available Attribute Values}{

The value is a numeric representing an angle in degrees.
\describe{
\item{\emph{'auto'}}{Specifies to rotate the \strong{marker} in such a way that the orientation automatically follows the path.}
\item{<angle>}{Specifies the value of much to rotate (in degrees) the \strong{marker}.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=marker]{marker}}}
}
}

\keyword{internal}
