% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{OriginAttribute}
\alias{OriginAttribute}
\title{origin}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{default}{The 'origin' attribute is defined in the SMIL Animation specification ([SMILANIM], section 4.3). It has no effect in SVG.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateMotion]{animateMotion}}}
}
}
\keyword{internal}

