% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PathLengthAttribute}
\alias{PathLengthAttribute}
\title{pathLength}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<number>}{The author's computation of the total length of the path, in user units. This value is used to calibrate the user agent's own distance.along.a.path calculations with that of the author. The user agent will scale all distance.along.a.path computations by the ratio of 'pathLength' to the user agent's own computed value for total path length. 'pathLength' potentially affects calculations for text on a path, motion animation and various stroke operations. A negative value is an error (see Error processing).Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=path]{path}}}
}
}
\keyword{internal}

