% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementHrefAttribute}
\alias{PatternElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to another pattern element residing in the current fragment.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies a reference to a pattern element.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
