% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementWidthAttribute}
\alias{PatternElementWidthAttribute}
\title{width}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{See 'x'. A negative value is an error (see Error processing). A value of zero disables rendering of the element (i.e., no paint is applied). If the attribute is not specified, the effect is as if a value of zero were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=pattern]{pattern}}}
}
}
\keyword{internal}

