% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PatternElementXAttribute}
\alias{PatternElementXAttribute}
\title{x}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{'x', 'y', 'width' and 'height' indicate how the pattern tiles are placed and spaced. These attributes represent coordinates and values in the coordinate space specified by the combination of attributes 'patternUnits' and 'patternTransform'. If the attribute is not specified, the effect is as if a value of zero were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=pattern]{pattern}}}
}
}
\keyword{internal}

