% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{PatternElementXYAttribute}
\alias{PatternElementXYAttribute}
\title{xy}
\description{
Shortcut for assigning x and y
}
\section{Combines}{

\code{\link[=PatternElementXYAttribute]{x}}, \code{\link[=PatternElementXYAttribute]{y}}
}

\section{Equivalence}{

\describe{
xy =v is equivalent to  x=v[1], y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Non-structural Container Elements}}{\code{\link[=pattern]{pattern}}}
}
}

\keyword{internal}
