% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PolygonElementPointsAttribute}
\alias{PolygonElementPointsAttribute}
\title{points}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list.of.points>}{The points that make up the polygon. All coordinate values are in the user coordinate system.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=polygon]{polygon}}}
}
}
\keyword{internal}

