% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{PreserveAspectRatioAttribute}
\alias{PreserveAspectRatioAttribute}
\title{preserveAspectRatio}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{ToDo!!!}{ToDo!!!}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Filter Primitive Elements}}{\code{\link[=feImage]{feImage}}}
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}, \code{\link[=symbol]{symbol}}}
\item{\emph{Unclassfied:}}{\code{\link[=image]{image}}, \code{\link[=marker]{marker}}, \code{\link[=pattern]{pattern}}, \code{\link[=view]{view}}}
}
}
\keyword{internal}

