% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RadialGradientElementCXAttribute}
\alias{RadialGradientElementCXAttribute}
\title{cx}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>}{'cx', 'cy' and 'r' define the largest (i.e., outermost) circle for the radial gradient. The gradient will be drawn such that the 100 gradient stop is mapped to the perimeter of this largest (i.e., outermost) circle. If the attribute is not specified, the effect is as if a value of '50' were specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}
\keyword{internal}

