% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RadialGradientElementCXYAttribute}
\alias{RadialGradientElementCXYAttribute}
\title{cxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=RadialGradientElementCXAttribute]{cx}}, \code{\link[=RadialGradientElementCYAttribute]{cy}}
}

\section{Equivalence}{

\describe{
\item{}{cxy=c(value.CX,value.CY)}
\item{and}{cx = value.CX; cy = value.CY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}
\keyword{internal}

