% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RadialGradientElementFXYAttribute}
\alias{RadialGradientElementFXYAttribute}
\title{fxy}
\description{
Shortcut for assigning fx and fy
}
\section{Combines}{

\code{\link[=RadialGradientElementFXYAttribute]{fx}}, \code{\link[=RadialGradientElementFXYAttribute]{fy}}
}

\section{Equivalence}{

\describe{
fxy =v is equivalent to  fx=v[1], fy=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}

\keyword{internal}
