% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RadialGradientElementRAttribute}
\alias{RadialGradientElementRAttribute}
\title{r}
\description{
Sets the radius of the outer circle of the radial gradient.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies the radius outermost circle of the \strong{radialGradient}. (The outermost circle being the location of the final gradient \emph{stop}.) Default is 0.50 .}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}

\keyword{internal}
