% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RadialGradientElementSpreadMethodAttribute}
\alias{RadialGradientElementSpreadMethodAttribute}
\title{spreadMethod}
\description{
Governs the behaviour that should occur if the gradient starts or ends inside the bounds of the target rectangle.
}
\section{Available Attribute Values}{

The value can be any one of the following:
\describe{
\item{\emph{'pad'}}{Specifies to extend the painting beyond the \emph{outermost-circle} of the \strong{radialGradient}, by using terminal color of the outermost circle. (Default)}
\item{\emph{'reflect'}}{Specifes to extend the painting beyond the \emph{outermost-circle} of the \strong{radialGradient}, by reflecting the gradient pattern. The reflections are continued until the entire gradient region is filled. For example, we might have a pattern of red white blue (red at the \emph{focus}, blue at the \emph{outermost-circle}) and then observe: red white blue blue white red red white blue.}
\item{\emph{'repeat'}}{Specifes to extend the painting beyond the \emph{outermost-circle} of the \strong{radialGradient}, by repeating the gradient pattern. The repetitions are continued until the entire gradient region is filled. For example, we might have a pattern of red white blue (red at the \emph{focus}, blue at the \emph{outermost-circle}) and then observe: red white blue red white blue red white blue.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Gradient Elements}}{\code{\link[=radialGradient]{radialGradient}}}
}
}

\keyword{internal}
