% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RectElementHeightAttribute}
\alias{RectElementHeightAttribute}
\title{height}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{The height of the rectangle. A negative value is an error (see Error processing). A value of zero disables rendering of the element.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}
\keyword{internal}

