% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RectElementRXYAttribute}
\alias{RectElementRXYAttribute}
\title{rxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=RectElementRXAttribute]{rx}}, \code{\link[=RectElementRYAttribute]{ry}}
}

\section{Equivalence}{

\describe{
\item{}{rxy=c(value.RX,value.RY)}
\item{and}{rx = value.RX; ry = value.RY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}
\keyword{internal}

