% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RectElementRYAttribute}
\alias{RectElementRYAttribute}
\title{ry}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<length>}{For rounded rectangles, the y.axis radius of the ellipse used to round off the corners of the rectangle. A negative value is an error (see Error processing). See the notes below about what happens if the attribute is not specified.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}
\keyword{internal}

