% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{RectElementWHAttribute}
\alias{RectElementWHAttribute}
\title{wh}
\description{
Shortcut for assigning height and width
}
\section{Combines}{

\code{\link[=RectElementWHAttribute]{height}}, \code{\link[=RectElementWHAttribute]{width}}
}

\section{Equivalence}{

\describe{
wh =v is equivalent to  height=v[1], width=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}

\keyword{internal}
