% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RectElementXAttribute}
\alias{RectElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the left side of the rectangle.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the left-hand side of the rectangle.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=rect]{rect}}}
}
}

\keyword{internal}
