% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RepeatDurAttribute}
\alias{RepeatDurAttribute}
\title{repeatDur}
\description{
Governs the total amount of time that the animation can continue (including repeats)
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{<clock.value>}{Specifies a duration during which the animation is repeated.}
\item{\emph{'indefinite'}}{Specifies to repeat indefinitely.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}

\keyword{internal}
