% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RestartAttribute}
\alias{RestartAttribute}
\title{restart}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{always}{The animation can be restarted at any time (The default)}
\item{whenNotActive}{The animation can be restarted only when not active (i.e. the active end)}
\item{never}{The element cannot be restarted for the duration of the parent container.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animate]{animate}}, \code{\link[=animateColor]{animateColor}}, \code{\link[=animateMotion]{animateMotion}}, \code{\link[=animateTransform]{animateTransform}}, \code{\link[=set]{set}}}
}
}
\keyword{internal}

