% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{RotateAttribute}
\alias{RotateAttribute}
\title{rotate}
\description{
Governs the rotation of an object along a path.
}
\section{Available Attribute Values}{

The value is a numeric representing an angle in degrees.
\describe{
\item{\emph{'auto'}}{Specifies that as the target element moves along the path, its orientation is to be continously adjusted by an \eqn{\theta(t)}, such that at any time \eqn{t} the vector \eqn{(cos(\theta(t)), sin(\theta(t))} lies in the direction of the objects trajectory. Since, at any given moment, the direction of the trajectory is a vector tangent to the path in the direction of the motion, the orientation of the element with respect to the path is continously maintained.}
\item{\emph{'auto-reverse'}}{Specifies that as the target element moves along the path, its orientation is to be continously adjusted by an \eqn{\theta(t)}, such that at any time \eqn{t} the vector \eqn{(-cos(\theta(t)), -sin(\theta(t))} lies in the direction of the objects trajectory.  Both \emph{auto} and \emph{auto-reverse}  maintain the orientation of the element with respect to the path, but the \emph{auto-reverse} image is a contionus reflection of the \emph{auto} image through a moving reference point on the path corresponding to the origin in the targets coordinate system.}
\item{<angle>}{Specifies to apply a constant rotational transformation on target element, where the rotational angle is the specified number of degrees.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=animateMotion]{animateMotion}}}
}
}

\keyword{internal}
