% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ScriptElementHrefAttribute}
\alias{ScriptElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to  the scripting code.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies an \emph{IRI} reference the script code.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=script]{script}}}
}
}

\keyword{internal}
