% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ScriptElementTypeAttribute}
\alias{ScriptElementTypeAttribute}
\title{type}
\description{
Sets which scripting language to use.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<content-type>[N]}{Specifies the scripting language for the given \emph{'script'} element. Default is \emph{""""""""""""""""application/ecmascript""""""""""""""""}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=script]{script}}}
}
}

\keyword{internal}
