% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SetElementToAttribute}
\alias{SetElementToAttribute}
\title{to}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<value>}{Specifies the value for the attribute during the duration of the 'set' element. The argument value must match the attribute type.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Animation Elements}}{\code{\link[=set]{set}}}
}
}
\keyword{internal}

