% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stop-color-presentationAttribute}
\alias{stop-color-presentationAttribute}
\title{stop.color}
\description{
Defines the color to achieve at the gradient stop position.
}
\section{Available Attribute Values}{



\describe{
\item{<color> <icccolor>}{Sets the stop color. The color may be expressed in one of three ways: name (i.e. red),  or  numeric (#f00 or #ff0000) or  rgb (rgb(255,0,0)). An optional ICC color may also be specified as a 2nd component of the color vector.}
\item{\emph{'\emph{'currentColor'}'}}{Specifies to use the \emph{currentColor} as the stop-colorr. The current color is the value of the color specified by the \emph{color} property.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{stop-color} value from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=stop]{stop}}}
}
}

\keyword{internal}
