% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stop-opacity-presentationAttribute}
\alias{stop-opacity-presentationAttribute}
\title{stop.opacity}
\description{
Defines the opacity  to achieve at the gradient stop position.
}
\section{Available Attribute Values}{



\describe{
\item{<opacity-value>}{Specifies the opacity value, a  number between 0 and 1 indicating the opacity to be used as fill , with 1 being opaque and 0 being transparent.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{stop-opacity} value from the parent.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=stop]{stop}}}
}
}

\keyword{internal}
