% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{stroke-linecap-presentationAttribute}
\alias{stroke-linecap-presentationAttribute}
\title{stroke.linecap}
\description{
Defines the shape to be used at the end of open subpaths when they are stroked.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'butt'}'}'}'}}{Specifies the \emph{stroke-linecap} to be a \emph{butt style cap}.}
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the \emph{stroke-linecap} value from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'round'}'}'}'}}{Specifies the \emph{stroke-linecap} to be a \emph{round style cap}.}
\item{\emph{'\emph{'\emph{'\emph{'square'}'}'}'}}{Specifies the \emph{stroke-linecap} to be a \emph{square style cap}.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Shape Elements}}{\code{\link[=circle]{circle}}, \code{\link[=ellipse]{ellipse}}, \code{\link[=line]{line}}, \code{\link[=path]{path}}, \code{\link[=polygon]{polygon}}, \code{\link[=polyline]{polyline}}, \code{\link[=rect]{rect}}}
\item{\emph{Text Content Elements}}{\code{\link[=altGlyph]{altGlyph}}, \code{\link[=text]{text}}, \code{\link[=textPath]{textPath}}, \code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
