% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{style}
\alias{style}
\title{An element used to embed a style specification}
\description{
The style element allows style sheets to be embedded directly within SVG content. SVG's style element has the same attributes as the corresponding element in HTML (see HTML's <style> element),.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Descriptive Elements}}{\code{\link[=StyleElementTitleAttribute]{title}}}
\item{\emph{Transfer Function Element Attributes}}{\code{\link[=StyleElementTypeAttribute]{type}}}
\item{\emph{Unclassified}}{\code{\link[=StyleElementMediaAttribute]{media}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
{No Content Elements Applicable}{!}
}
}

\keyword{element}
