% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{StyleElementMediaAttribute}
\alias{StyleElementMediaAttribute}
\title{media}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{media.descriptors}{This attribute specifies the intended destination medium for style information. It may be a single media descriptor or a comma.separated list. The default value for this attribute is "all". The set of recognized media.descriptors are the list of media types recognized by CSS2 ([CSS2], section 7.3).Animatable. no.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Unclassfied:}}{\code{\link[=style]{style}}}
}
}
\keyword{internal}

