% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{StyleElementTitleAttribute}
\alias{StyleElementTitleAttribute}
\title{title}
\description{
Sets an advisory title for the style element.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{\emph{'advisory-title'}}{Specifies an advisory title for the style element (HTML 4)}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=style]{style}}}
}
}

\keyword{internal}
