% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{StyleElementTypeAttribute}
\alias{StyleElementTypeAttribute}
\title{type}
\description{
Sets the style sheet language of the element's contents.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{\emph{'content-type'}}{Specifies the style sheet language of the element's contents, such as \emph{""""""""""""""""text/css""""""""""""""""}.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=style]{style}}}
}
}

\keyword{internal}
