% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{SVGElementWHAttribute}
\alias{SVGElementWHAttribute}
\title{wh}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=SVGElementHeightAttribute]{height}}, \code{\link[=SVGElementWidthAttribute]{width}}
}

\section{Equivalence}{

\describe{
\item{}{wh=c(value.HEIGHT,value.WIDTH)}
\item{and}{height = value.HEIGHT; width = value.WIDTH}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}}
}
}
\keyword{internal}

