% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SVGElementXAttribute}
\alias{SVGElementXAttribute}
\title{x}
\description{
Sets the x-coordinate of the left side of an embedded svg. (Revelant only for embedded svg elements)
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the x-coordinate of the left hand side of an embedded \strong{svg}. This has no effect on the outer most \strong{svg}(i.e. \strong{svgR}). The default is 0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
