% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{SVGElementXYAttribute}
\alias{SVGElementXYAttribute}
\title{xy}
\description{
Shortcut for assigning x and y
}
\section{Combines}{

\code{\link[=SVGElementXYAttribute]{x}}, \code{\link[=SVGElementXYAttribute]{y}}
}

\section{Equivalence}{

\describe{
xy =v is equivalent to  x=v[1], y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
