% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{SVGElementYAttribute}
\alias{SVGElementYAttribute}
\title{y}
\description{
Sets y-coordinate of the top side of an embedded svg.
}
\section{Available Attribute Values}{

The value represents a coordinate.
\describe{
\item{<coordinate>}{Specifies the y-coordinate of the top side of an embedded \emph{svg}. This has no effect on the outer most \emph{svg}(i.e. \emph{svgR}). The default is 0.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Container Elements}}{\code{\link[=svg]{svg}}}
}
}

\keyword{internal}
