% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{text-rendering-presentationAttribute}
\alias{text-rendering-presentationAttribute}
\title{text.rendering}
\description{
Governs the optimization strategy of rendering text.
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'auto'}'}}{Specifies to apply a \emph{text-rendering} algorithm that balances the trade offs  that occur between optimizing for speed vs legibility vs geometric precision, with emphasis on legibility.}
\item{\emph{'\emph{'geometricPrecision'}'}}{Specifies to optimize  \emph{text-rendering} for geometric precision.}
\item{\emph{'\emph{'inherit'}'}}{Specifies to inherit the \emph{text-rendering} value from the parent.}
\item{\emph{'\emph{'optimizeLegibility'}'}}{Specifies to optimize  \emph{text-rendering} for legibility.}
\item{\emph{'\emph{'optimizeSpeed'}'}}{Specifies to optimize  \emph{text-rendering} for speed.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
