% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextElementRotateAttribute}
\alias{TextElementRotateAttribute}
\title{rotate}
\description{
Sets a sequence of  one or more rotations to be applied to consecutive glyphs upon rendering.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<angle>[N]}{Specifies an additional rotation (in degrees) to be applied sequentiolly to each character. If the number of characters exceeds the number of rotations, the extra character receive the last rotation.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
