% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextElementTextLengthAttribute}
\alias{TextElementTextLengthAttribute}
\title{textLength}
\description{
Sets the intended text length.
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies a presumed total sum of all of the advance values that correspond to character data within this element to assist by the user agent to in achieving exact alignment. The presumed total sum should include the advance value on the \emph{glyph} (horizontal or vertical), the effect of properties \emph{'kerning'}, \emph{'letter.spacing'} and \emph{'word.spacing'} and any adjustments due to attributes \emph{'dx'} and \emph{'dy'} on \emph{'tspan'} elements.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
