% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{TextElementXYAttribute}
\alias{TextElementXYAttribute}
\title{xy}
\description{
Shortcut for assigning x and y
}
\section{Combines}{

\code{\link[=TextElementXYAttribute]{x}}, \code{\link[=TextElementXYAttribute]{y}}
}

\section{Equivalence}{

\describe{
xy =v is equivalent to  x=v[1], y=v[2]
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
