% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextPathElementHrefAttribute}
\alias{TextPathElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to path element which defines the path along which the text will be rendered.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies a reference to a path element along which the \emph{glyphs} should be placed.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=textPath]{textPath}}}
}
}

\keyword{internal}
