% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextPathElementMethodAttribute}
\alias{TextPathElementMethodAttribute}
\title{method}
\description{
Governs whether to allow stretching and warping of theglyphs along the given path..
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'align'}}{Specifies for rendering that uses simple 2x3 transformations with no stretching/warping of the \emph{glyphs}.}
\item{\emph{'stretch'}}{Specifies that rendering will convert glyph outlines to paths with endpoints and control points adjusted to be normal to the path producing possible stetching and warping. This approach is taken to insure connected \emph{glyphs} remain connect along the path.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=textPath]{textPath}}}
}
}

\keyword{internal}
