% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextPathElementSpacingAttribute}
\alias{TextPathElementSpacingAttribute}
\title{spacing}
\description{
Governs the spacing of individual glyphs when rendering along a given path.
}
\section{Available Attribute Values}{

The value can be either of the one of the following:
\describe{
\item{\emph{'exact'}}{Specifies to the user agent the spacing between _glyphs_ that are to be rendered along a path to be rendered exactly without any adjustment along the path.}
\item{\emph{'auto'}}{Specifies for the user agent to adjust the spacing between \emph{glyphs} using a \emph{text-on-a-path layout algorithm}. The intension is to create a better visual apeal.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=textPath]{textPath}}}
}
}

\keyword{internal}
