% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TextPathElementStartOffsetAttribute}
\alias{TextPathElementStartOffsetAttribute}
\title{startOffset}
\description{
Sets the offset from the start of the path to the initial text position,
}
\section{Available Attribute Values}{

The value represents a length.
\describe{
\item{<length>}{Specifies an offset measured from the starting point of the \emph{'path'} to the initial position for the current text. The offset distance is measured either in percent or in the current user coordinate system and is calculated using the user agent's distance along the path algorithm. If measured in percent, then the total distance of the path is defined to be 100 percent.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: Yes .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=textPath]{textPath}}}
}
}

\keyword{internal}
