% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TRefElementHrefAttribute}
\alias{TRefElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to chararcter data for this tref..
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies an \emph{IRI} reference to an element containing the content character data to be used.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=tref]{tref}}}
}
}

\keyword{internal}
