% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CombAttrPages.R
\name{TSpanElementDXYAttribute}
\alias{TSpanElementDXYAttribute}
\title{dxy}
\description{
ToDo: Needs to be written!!!
}
\section{Combines}{

\code{\link[=TSpanElementDXAttribute]{dx}}, \code{\link[=TSpanElementDYAttribute]{dy}}
}

\section{Equivalence}{

\describe{
\item{}{dxy=c(value.DX,value.DY)}
\item{and}{dx = value.DX; dy = value.DY}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}
\keyword{internal}

