% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TSpanElementDYAttribute}
\alias{TSpanElementDYAttribute}
\title{dy}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{<list.of.lengths>}{The corresponding list of relative Y coordinates for the characters within the 'tspan' element. The processing rules for the 'dy' attribute parallel the processing rules for the 'dx' attribute.Animatable. yes.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Child Elements}}{\code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}
\keyword{internal}

