% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{TSpanElementXAttribute}
\alias{TSpanElementXAttribute}
\title{x}
\description{
Sets the tspan x-coordinate(s)
}
\section{Available Attribute Values}{

\describe{
\item{<coordinate>[N]}{Specifies the absolute x-coordinate for the first n characters where n is the length of the vector x. (default is 0)}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=tref]{tref}}, \code{\link[=tspan]{tspan}}}
}
}

\keyword{internal}
