% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{UseElementHrefAttribute}
\alias{UseElementHrefAttribute}
\title{xlink.href}
\description{
Sets an IRI reference to an element within this fragment to be used.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<IRI>}{Specifies an element or fragment to use by way of an \emph{IRI} reference. \emph{IRI} should be contained within the same document.}
}
}

\section{Animatable}{

Using:  \code{\link[=animate]{animate}}, \code{\link[=set]{set}} .
Supports Additive: No .
}

\section{Used by the Elements}{

\describe{
\item{\emph{Graphics Referencing Elements}}{\code{\link[=use]{use}}}
}
}

\keyword{internal}
