% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ViewElementViewTargetAttribute}
\alias{ViewElementViewTargetAttribute}
\title{viewTarget}
\description{
Sets the target object to be for this view.
}
\section{Available Attribute Values}{

The value is defined as follows:
\describe{
\item{<names>}{Specifies one or more targets to be associated with this view.}
}
}

\section{Animatable}{

Not Animatable
}

\section{Used by the Elements}{

\describe{
\item{\emph{Uncategorized Elements}}{\code{\link[=view]{view}}}
}
}

\keyword{internal}
