% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ElePages.R
\name{vkern}
\alias{vkern}
\title{An element used to specify vertical distances between glyphs}
\description{
The vertical distance between two glyphs in top-to-bottom fonts can be fine-tweaked with an vkern Element. This process is known as Kerning.
}
\section{Available Attributes (Named Parameters)}{

\describe{
\item{\emph{Complementary Attributes}}{\code{\link[=HKernElementG12Attribute]{g12}}, \code{\link[=HKernElementU12Attribute]{u12}}}
\item{\emph{Core Attributes}}{\code{\link[=IDAttribute]{id}}, \code{\link[=XMLBaseAttribute]{xml.base}}, \code{\link[=XMLLangAttribute]{xml.lang}}, \code{\link[=XMLSpaceAttribute]{xml.space}}}
\item{\emph{Unclassified}}{\code{\link[=HKernElementG1Attribute]{g1}}, \code{\link[=HKernElementG2Attribute]{g2}}, \code{\link[=HKernElementKAttribute]{k}}, \code{\link[=HKernElementU1Attribute]{u1}}, \code{\link[=HKernElementU2Attribute]{u2}}}
}
}

\section{Available Content Elements (Unnamed Parameters)}{

\describe{
{No Content Elements Applicable}{!}
}
}

\keyword{element}
