% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_PresAttrPages.R
\name{writing-mode-presentationAttribute}
\alias{writing-mode-presentationAttribute}
\title{writing.mode}
\description{
Governs the initial inline-progression-direction
}
\section{Available Attribute Values}{



\describe{
\item{\emph{'\emph{'\emph{'\emph{'inherit'}'}'}'}}{Specifies to inherit the \emph{writing-mode} value from the parent.}
\item{\emph{'\emph{'\emph{'\emph{'lr'}'}'}'}}{Specifies \emph{writing-mode} is to be left to right.}
\item{\emph{'\emph{'\emph{'\emph{'lr-tb'}'}'}'}}{Specifies \emph{writing-mode} is to be left to right and  top to bottom.}
\item{\emph{'\emph{'\emph{'\emph{'rl'}'}'}'}}{Specifies \emph{writing-mode} is to be  right to left.}
\item{\emph{'\emph{'\emph{'\emph{'rl-tb'}'}'}'}}{Specifies \emph{writing-mode} is to be right to left, top to bottom.}
\item{\emph{'\emph{'\emph{'\emph{'tb'}'}'}'}}{Specifies \emph{writing-mode} is to be top to bottom.}
\item{\emph{'\emph{'\emph{'\emph{'tb-rl'}'}'}'}}{Specifies \emph{writing-mode} is to be top to bottom, right to left.}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Text Content Elements}}{\code{\link[=text]{text}}}
}
}

\keyword{internal}
