% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RegAttrPages.R
\name{ZoomAndPanAttribute}
\alias{ZoomAndPanAttribute}
\title{zoomAndPan}
\description{
ToDo: Needs to be written!!!
}
\section{Available Attribute Values}{

\describe{
\item{disable}{disable zooming}
\item{magnify}{enable zooming (only for outermost svg)}
}
}

\section{Used by the Elements}{

\describe{
\item{\emph{Structural Elements}}{\code{\link[=svg]{svg}}}
\item{\emph{Unclassfied:}}{\code{\link[=view]{view}}}
}
}
\keyword{internal}

